% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{agentClasses-class}
\alias{agentClasses}
\alias{agentClasses-class}
\title{A meta class for agentMatrix and SpatialPointsDataFrame}
\description{
Both these types can be used by NetLogoR to describe turtle agents.
}
\author{
Eliot McIntire
}

