% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{==,agentMatrix,character-method}
\alias{==,agentMatrix,character-method}
\title{Relational Operators}
\usage{
\S4method{==}{agentMatrix,character}(e1, e2)
}
\arguments{
\item{e1}{An \code{agentMatrix} object.}

\item{e2}{atomic vector, symbol, call, or other object for which methods have been written.}
}
\description{
Binary operators which allow the comparison of values in an agentMatrix.
}

