% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,agentMatrix-method}
\alias{head}
\alias{head.agentMatrix}
\alias{length,agentMatrix-method}
\alias{nrow,agentMatrix-method}
\alias{show,agentMatrix-method}
\alias{tail}
\alias{tail.agentMatrix}
\title{Show an object or get information about the object}
\usage{
\S4method{show}{agentMatrix}(object)

\S4method{length}{agentMatrix}(x)

\S4method{nrow}{agentMatrix}(x)

\method{head}{agentMatrix}(x, n = 6L, ...)

\method{tail}{agentMatrix}(x, n = 6L, ...)
}
\arguments{
\item{object}{An \code{agentMatrix} object.}

\item{x}{An \code{agentMatrix} object.}

\item{n}{documentation needed}

\item{...}{documentation needed}
}
\description{
description needed
}

