% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{.bboxCoords}
\alias{.bboxCoords}
\title{Faster .bboxCoords}
\usage{
.bboxCoords(coords)
}
\arguments{
\item{coords}{documentation needed}
}
\description{
This is a drop in replacement for .bboxCoords in raster package.
}

