% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{canMove}
\alias{canMove}
\alias{canMove,NLworlds,SpatialPointsDataFrame,numeric-method}
\title{Can move}
\usage{
canMove(world, turtles, step)

\S4method{canMove}{NLworlds,SpatialPointsDataFrame,numeric}(world, turtles,
  step)
}
\arguments{
\item{world}{A \code{NLworlds} object, representing the world in which the
turtles would move onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtle(s) to be
evaluated.}

\item{step}{Numeric. The distance(s) the turtles would move. Must be of
length 1 if all the turtles would move the same distance or
of the same length as the turtles object.}
}
\value{
A vector of logical of length equal to \code{length(turtles)}.
}
\description{
Reports \code{TRUE} or \code{FALSE} if the turtle(s) can move the given distance(s)
without leaving the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(world = w1, n = 10)
canMove(world = w1, turtles = t1, step = 1:10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

