% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{clearTurtles}
\alias{clearTurtles}
\alias{clearTurtles,SpatialPointsDataFrame-method}
\title{Clear turtles}
\usage{
clearTurtles(turtles, envir = parent.frame(n = 1))

\S4method{clearTurtles}{SpatialPointsDataFrame}(turtles,
  envir = parent.frame(n = 1))
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{envir}{The R environment where \code{turtles} are located.
Default value is \code{envir = parent.frame(n = 1)}.}
}
\description{
Kill all the turtles.
}
\details{
Removes the \code{turtles} object from the R environment. If some of
         the agents in \code{turtles} are present in other agentsets, it does
         not remove them from these other agentsets. This function just remove
         the given \code{turtles} agentset.
}
\examples{
w1 <- createNLworld()
t1 <- createTurtles(n = 10, world = w1)
t1 # show the object containing the turtles
clearTurtles(t1)
t1 # 't1' not found


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#clear-turtles}
}

