% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{NLWorldIndex}
\alias{NLWorldIndex}
\title{Convert NLWorldMatrix indices to vector indices}
\usage{
NLWorldIndex(world, cellNum)
}
\arguments{
\item{world}{NLworldMatrix object.}

\item{cellNum}{Integer. Vector of cells number.}
}
\value{
Numeric vector with indices that will work with an NLWorldMatrix.
}
\description{
This can be used to convert indices that would work with an R \code{RasterLayer},
\code{matrix}, or \code{vector} to NLWorldMatrix indices.
}
\examples{
w1 <- createNLworldMatrix(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- 1:100
w1Ras <- world2raster(w1)
index <- 24
pxpy <- PxcorPycorFromCell(world = w1, cellNum = index)

rasValue <- as.integer(unname(w1Ras[index]))
# Not correct index:
identical(w1[index], rasValue)

# Correct index
identical(w1[NLWorldIndex(w1, index)], rasValue)


}
\author{
Eliot McIntire
}

