% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{convertNLworld}
\alias{convertNLworld}
\alias{convertNLworld,RasterLayer-method}
\alias{convertNLworld,RasterStack-method}
\title{Convert a Raster* object into a NLworlds object}
\usage{
convertNLworld(raster)

\S4method{convertNLworld}{RasterLayer}(raster)

\S4method{convertNLworld}{RasterStack}(raster)
}
\arguments{
\item{raster}{RasterLayer or RasterStack object.}
}
\value{
NLworld or NLworldStack object depending on the input.
        Patches value are retained from the Raster* object.
}
\description{
Convert a RasterLayer object into a NLworld object or a RasterStack object
into a NLworldStack object.
}
\details{
See \code{help("NLworlds-class")} for more details on the NLworlds
         classes.

         The \code{raster} is resampled to match the coordinates system and
         resolution of a NLworlds using the nearest neighbor method. The
         extent will be bigger by 1 on the width and on the height.
}
\examples{
r <- raster(nrows = 21, ncols = 21, xmn = 0, ymn = 0, res = 5)
r[] <- runif(length(r))
plot(r)
extent(r)

world <- convertNLworld(raster = r)
plot(world)
extent(world)
minPxcor(world)
maxPxcor(world)
minPycor(world)
maxPycor(world)


}
\author{
Sarah Bauduin
}

