% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{createNLworld}
\alias{createNLworld}
\alias{createNLworld,missing,missing,missing,missing-method}
\alias{createNLworld,numeric,numeric,numeric,numeric-method}
\title{Create a \code{NLworld} world}
\usage{
createNLworld(minPxcor = -16, maxPxcor = 16, minPycor = -16,
  maxPycor = 16)

\S4method{createNLworld}{numeric,numeric,numeric,numeric}(minPxcor = -16,
  maxPxcor = 16, minPycor = -16, maxPycor = 16)

\S4method{createNLworld}{missing,missing,missing,missing}()
}
\arguments{
\item{minPxcor}{Integer. Minimum \code{pxcor} for the patches (left border).
Default value is \code{minPxcor = -16}.}

\item{maxPxcor}{Integer. Maximum \code{pxcor} for the patches (right border).
Default value is \code{maxPxcor = 16}.}

\item{minPycor}{Integer. Minimum \code{pycor} for the patches (bottom).
Default value is \code{minPycor = -16}.}

\item{maxPycor}{Integer. Maximum \code{pycor} for the patches (top).
Default value is \code{maxPycor = 16}.}
}
\value{
\code{NLworld} object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patches value are \code{NA}.
}
\description{
Create an empty grid of patches of class \code{NLworld}.
}
\details{
See \code{help("NLworld-class")} for more details on the \code{NLworld} class.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
plot(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

