% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{createNLworldMatrix}
\alias{createNLworldMatrix}
\alias{createNLworldMatrix,missing,missing,missing,missing,missing-method}
\alias{createNLworldMatrix,numeric,numeric,numeric,numeric,ANY-method}
\title{Create a world}
\usage{
createNLworldMatrix(minPxcor, maxPxcor, minPycor, maxPycor, data = NA)

\S4method{createNLworldMatrix}{numeric,numeric,numeric,numeric,ANY}(minPxcor,
  maxPxcor, minPycor, maxPycor, data = NA)

\S4method{createNLworldMatrix}{missing,missing,missing,missing,missing}()
}
\arguments{
\item{minPxcor}{Integer. Minimum pxcor for the patches (world's left border).}

\item{maxPxcor}{Integer. Maximum pxcor for the patches (world's right border).}

\item{minPycor}{Integer. Minimum pycor for the patches (world's bottom border).}

\item{maxPycor}{Integer. Maximum pycor for the patches (world's top border).}

\item{data}{Vector of length \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}.
Default is \code{NA}.}
}
\value{
NLworldMatrix object.
}
\description{
Create a world of patches of class NLworldMatrix.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}

