% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{createOTurtlesAM}
\alias{createOTurtlesAM}
\alias{createOTurtlesAM,numeric-method}
\title{Create ordered turtles (AM)}
\usage{
createOTurtlesAM(n, world, heading, breed, color)

\S4method{createOTurtlesAM}{numeric}(n, world, breed, color)
}
\arguments{
\item{n}{Integer.}

\item{world}{NLworlds object.}

\item{heading}{Numeric. Vector of values between 0 and 360. Must be of length 1 or
of length \code{n}. If missing, a random heading is assigned to
each turtle.}

\item{breed}{Character. Vector of "breed" names. Must be of length 1 or of length
\code{n}. If missing, \code{breed = "turtle"} for all turtles.}

\item{color}{Character. Vector of color names. Must be of length \code{n}.
If missing, colors are assigned using the function \code{rainbow(n)}.}
}
\value{
AgentMatrix of length \code{n} with the columns for the
        dataframe being: "who", "heading", "prevX", "prevY", "breed", and "color".
}
\description{
Create \code{n} turtles at the center of the \code{world} with their headings evenly
distributed.
}
\details{
The identity of the turtles is defined by their "who" number. This
         numbering starts at 0 and increments by 1.

         The coordinates from the previous time step are stored in "prevX" and
         "prevY". The initial values are \code{NA}.
}
\author{
Sarah Bauduin and Eliot McIntire
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#create-ordered-turtles}
}

