% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{createTurtles}
\alias{createTurtles}
\alias{createTurtles,numeric,matrix,missing-method}
\alias{createTurtles,numeric,missing,NLworlds-method}
\title{Create turtles}
\usage{
createTurtles(n, coords, world, heading, breed, color)

\S4method{createTurtles}{numeric,matrix,missing}(n, coords, world, heading,
  breed, color)

\S4method{createTurtles}{numeric,missing,NLworlds}(n, coords, world, heading,
  breed, color)
}
\arguments{
\item{n}{Integer. The number of new turtles to create.}

\item{coords}{A matrix (ncol = 2, nrow = n) with the first column \code{xcor}
and the second column \code{ycor} representing the coordinates
for the turtles to be created. \code{nrow(coords)} must be either
equal to 1 if all turtles have the same initial position or equal
to \code{n} if different turtles have different initial position.
Given coordinates must be inside the world's extent. If missing,
turtles are put in the center of the \code{world}.}

\item{world}{A \code{NLworlds} object, representing the world in which the turtles
will evolve. If \code{coords} are provided, \code{world} should
not be provided.}

\item{heading}{Numeric value(s) between 0 and 360. Either of length 1 representing
the heading for all turtles or of length \code{n} if different
turtles have different headings. If missing, a random heading is
assigned to each turtle.}

\item{breed}{String of characters. Either of length 1 representing the breed
for all turtles or of length \code{n} if the different turtles have
different breeds. If missing, \code{breed = "turtle"} for all turtles.}

\item{color}{String of characters of length \code{n} representing the color
of each turtle when plotted. If missing, colors are assigned using
the function \code{rainbow(n)}.}
}
\value{
A SpatialPointsDataFrame object of length \code{n} with the columns for
        the dataframe being: "who", "heading", "prevX", "prevY", "breed", and "color".
}
\description{
Create \code{n} new turtles with a set of defined variables.
}
\details{
The identity of the turtles is defined by their "who" number. This
         numbering starts at 0 and increments by 1.
         The coordinates from the previous time step are stored in "prevX" and
         "prevY". The initial values are \code{NA}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(25)
t1 <- createTurtles(n = 10, coords = cbind(xcor = runif(10, 0, 4), ycor = runif(10, 0, 4)))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1") # automatically uses color column in SpatialPointsDataFrame


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

