% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorld-function.R
\docType{methods}
\name{createWorld}
\alias{createWorld}
\alias{createWorld,ANY-method}
\title{Create a world}
\usage{
createWorld(minPxcor = -16, maxPxcor = 16, minPycor = -16,
  maxPycor = 16, patchSize = 1, ...)

\S4method{createWorld}{ANY}(minPxcor = -16, maxPxcor = 16, minPycor = -16,
  maxPycor = 16, patchSize = 1, ...)
}
\arguments{
\item{minPxcor}{Minimum x coordinate (raster left border).
Default value \code{-16}, as in NetLogo.}

\item{maxPxcor}{Maximum x coordinate (raster right border).
Default value \code{16}, as in NetLogo.}

\item{minPycor}{Minimum y coordinate (raster bottom border).
Default value \code{-16}, as in NetLogo.}

\item{maxPycor}{Maximum y coordinate (raster top border).
Default value \code{16}, as in NetLogo.}

\item{patchSize}{Numeric vector of length 1 or 2 to set the raster resolution.
Default value \1.}

\item{...}{Additional arguments, see Details}
}
\value{
A RasterLayer object with extent and resolution as given by the parameters.
        Cells value are NA.
}
\description{
Create empty patches at the extent and resolution given for the world.
}
\details{
Arguments of the \code{raster} function from the raster package may be added.
}
\examples{
# Create a square world of 25 patches.
world <- createWorld(minPxcor = 0, maxPxcor = 5, minPycor = 0, maxPycor = 5)
# Give the patches random values between 0 and 10.
world[] <- runif(n = 25, min = 0, max = 10)
plot(world)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

