% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{die}
\alias{die}
\alias{die,SpatialPointsDataFrame,numeric-method}
\title{Kill turtles}
\usage{
die(turtles, who)

\S4method{die}{SpatialPointsDataFrame,numeric}(turtles, who)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()}.}

\item{who}{Integers. The who numbers of the turtles to kill.}
}
\value{
SpatialPointsDataFrame of length equal to \code{length(turtles) - length(who)}.
}
\description{
Kill specific turtles.
}
\details{
The who numbers of the other turtles are maintained.

         To remove all the turtles, you can use \code{clearTurtles()}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
length(t1)
t1 <- die(turtles = t1, who = c(2, 3, 4))
length(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#die}
}

