% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{.bboxCoords}
\alias{.bboxCoords}
\title{\code{.bboxCoords} is a drop in replacement for \code{raster::.bboxCoords}.}
\usage{
.bboxCoords(coords)
}
\arguments{
\item{coords}{xy coordinates for all cells, e.g., produced by \code{raster::coordinates}.}
}
\description{
\code{.bboxCoords} is a drop in replacement for \code{raster::.bboxCoords}.
}
