% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{downhill}
\alias{downhill}
\alias{downhill,NLworld,missing,SpatialPointsDataFrame,numeric-method}
\alias{downhill,NLworldStack,character,SpatialPointsDataFrame,numeric-method}
\title{Move downhill}
\usage{
downhill(world, pVar, turtles, nNeighbors, torus = FALSE)

\S4method{downhill}{NLworld,missing,SpatialPointsDataFrame,numeric}(world,
  turtles, nNeighbors, torus)


  \S4method{downhill}{NLworldStack,character,SpatialPointsDataFrame,numeric}(world,
  pVar, turtles, nNeighbors, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the
turtles move onto.}

\item{pVar}{Characters. If the world is a \code{NLworldStack}, \code{pVar}
is the name of the layer used to define the patches's variable
used to move downihll. Should not be provided if the world is
\code{NLworld}.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{nNeighbors}{Integer: 4 or 8. The number of neighbor patches considered to move
downhill.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated
        coordinates and updated data for their headings and previous coordinates
        "prevX" and "prevY".
}
\description{
Move the turtles to their neighboring patch with the lowest value for the patches'
variable.
}
\details{
If no neighboring patch has a smaller value than the patch where the
         turtle is currently located on, the turtle stays on this patch. It still
         moves to the patch center if it was not already on it.

         If there are multiple neighboring patches with the same lowest value,
         the turtle chooses one patch at random.

         If \code{torus = FALSE}, turtles cannot move on the other side of the world.
         If a turtle is located on a patch on the edge of the world, it has fewer
         neighborhing patches for option to move than \code{nNeighbors}. If
         \code{torus = TRUE}, turtles can move on the other side of the world to
         go downhill and their choice of neighborhing patches is always among
         \code{nNeighbors} patches.
}
\examples{
w1 <- createNLworld(minPxcor = 1, maxPxcor = 10, minPycor = 1, maxPycor = 10)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")

t1 <- downhill(world = w1, turtles = t1, nNeighbors = 8)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#downhill}
}

