% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{dx}
\alias{dx}
\alias{dx,SpatialPointsDataFrame,missing-method}
\alias{dx,SpatialPointsDataFrame,numeric-method}
\title{x-increment}
\usage{
dx(turtles, dist = 1)

\S4method{dx}{SpatialPointsDataFrame,numeric}(turtles, dist = 1)

\S4method{dx}{SpatialPointsDataFrame,missing}(turtles)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{dist}{Numeric. Vector of distances the \code{turtles} would have to
move forward to
compute the increment values. Must be of length 1 or of length
\code{turtles}. The default value is \code{dist = 1}.}
}
\value{
Numeric. Vector of length \code{turtles}.
}
\description{
Report the amount by which the \code{turtles}' coordinates xcor would change
if the \code{turtles} were
to move forward the given distances with their current headings.
}
\details{
Report the sine of the \code{turtles}' heading multiplied by the \code{dist}
         values. Heading 0 is north and angles are calculated in degrees in a
         clockwise manner.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createOTurtles(world = w1, n = 10)
dx(turtles = t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#dxy}
}

