% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{dy}
\alias{dy}
\alias{dy,SpatialPointsDataFrame,missing-method}
\alias{dy,SpatialPointsDataFrame,numeric-method}
\title{dy}
\usage{
dy(turtles, step = 1)

\S4method{dy}{SpatialPointsDataFrame,numeric}(turtles, step = 1)

\S4method{dy}{SpatialPointsDataFrame,missing}(turtles)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the agents that were to
move forward.}

\item{step}{Numeric. The distance(s) the turtles would have to move forward to
compute the y-increment value. Must be of length 1 if all the
turtles would have to move the same distance or of the same length
as the turtles object. The default value is \code{step = 1}.}
}
\value{
A vector of the length of the turtles object.
}
\description{
Reports the y-increment (i.e., the amount by which the turtles' ycor would change)
if the turtles were to move forward of \code{step} distance(s) with their current
heading.
}
\details{
Reports the cosine of the turtles' heading multiplied by the \code{step}
         value(s). Heading 0 is north and angles are calculated in degrees in a
         clockwise manner.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createOTurtles(world = w1, n = 10)
dy(turtles = t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

