% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{face}
\alias{face}
\alias{face,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{face,NLworlds,SpatialPointsDataFrame,matrix-method}
\title{Face something}
\usage{
face(world, turtles, agents2, torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,matrix}(world, turtles, agents2,
  torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  turtles, agents2, torus = FALSE)
}
\arguments{
\item{world}{NLworlds object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{agents2}{Matrix (ncol = 2) with the first column "pxcor" and the second
               column "pycor" representing the patches coordinates, or

               SpatialPointsDataFrame created by \code{createTurtles()} or
               by \code{createOTurtles()} representing the moving agents, or

               Matrix (ncol = 2) with the first column "x and the second column
               "y" representing locations coordinates.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated headings.
}
\description{
Set the \code{turtles}' heading towards \code{agents2}.
}
\details{
The number of agents/locations in \code{agents2} must be equal to 1 or
         to the length of \code{turtles}.

         If \code{torus = TRUE} and the distance from one \code{turtles} to
         its corresponding agent/location \code{agents2} is smaller around the
         sides of the \code{world} than across it, then the direction to the agent/location
         \code{agents2} going around the sides of the \code{world} is given to the turtle.

         If a turtle is facing its own location, its heading does not change.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearplot()
Plot(w1)
Plot(t1, addTo = "w1")

t1 <- face(world = w1, turtles = t1, agents2 = cbind(x = 0, y = 0))
t1 <- fd(world = w1, turtles = t1, dist = 0.5)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#face}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#facexy}
}

