% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-arguments.R
\name{fargs}
\alias{fargs}
\title{Function arguments}
\arguments{
\item{n}{Integer.}

\item{world}{NLworlds object.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}

\item{minPxcor}{Integer. Minimum pxcor for the patches (world's left border).}

\item{maxPxcor}{Integer. Maximum pxcor for the patches (world's right border).}

\item{minPycor}{Integer. Minimum pycor for the patches (world's bottom border).}

\item{maxPycor}{Integer. Maximum pycor for the patches (world's top border).}

\item{pxcor}{Integer. Vector of patches pxcor coordinates. Must be of length 1
or of the same length as \code{pycor}.}

\item{pycor}{Integer. Vector of patches pycor coordinates. Must be of length 1
or of the same length as \code{pxcor}.}

\item{pVar}{Character. If the \code{world} is a NLworldStack object, \code{pVar}
is the name of the layer to use to define the patches values.
\code{pVar} must not be provided if the \code{world} is a NLworld object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{agents2}{Matrix (ncol = 2) with the first column "pxcor" and the second
               column "pycor" representing the patches coordinates, or

               SpatialPointsDataFrame created by \code{createTurtles()} or
               by \code{createOTurtles()} representing the moving agents, or

               Matrix (ncol = 2) with the first column "x and the second column
               "y" representing locations coordinates.}

\item{nNeighbors}{Integer: 4 or 8. Represents the number of neihgbor patches
considered.}

\item{dx}{Numeric. Vector of distances to the east (right) from the \code{agents}.
If \code{dx} is negative, the distance to the west (left) is computed.
\code{dx} must be of length 1 or of the same length as number of patches
or turtles in \code{agents}.}

\item{dy}{Numeric. Vector of distances to the north (up) from the \code{agents}.
If \code{dy} is negative, the distance to the south is computed (down).
\code{dy} must be of length 1 or of the same length as number of patches
or turtles in \code{agents}.}
}
\description{
Function arguments
}
\keyword{integral}

