% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{getMaxPycor}
\alias{getMaxPycor}
\alias{getMaxPycor,NLworld-method}
\title{Get maximum pycor}
\usage{
getMaxPycor(world)

\S4method{getMaxPycor}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the patches maximum \code{pycor} in a NLworld.
}
\details{
Does the same as accessing the slot with \code{NLworld@maxPycor}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
getMaxPycor(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

