% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{getMinPxcor}
\alias{getMinPxcor}
\alias{getMinPxcor,NLworld-method}
\title{Get minimum pxcor}
\usage{
getMinPxcor(world)

\S4method{getMinPxcor}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the patches minimum \code{pxcor} in a NLworld.
}
\details{
Does the same as accessing the slot with \code{NLworld@minPxcor}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
getMinPxcor(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

