% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{getMinPycor}
\alias{getMinPycor}
\alias{getMinPycor,NLworld-method}
\title{Get minimum pycor}
\usage{
getMinPycor(world)

\S4method{getMinPycor}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the patches minimum \code{pycor} in a NLworld.
}
\details{
Does the same as accessing the slot with \code{NLworld@minPycor}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
getMinPycor(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

