% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{inCone}
\alias{inCone}
\alias{inCone,SpatialPointsDataFrame,numeric,numeric,SpatialPointsDataFrame,NLworlds-method}
\alias{inCone,SpatialPointsDataFrame,numeric,numeric,matrix,NLworlds-method}
\title{In cone}
\usage{
inCone(turtles, radius, angle, agents2, world, torus = FALSE)


  \S4method{inCone}{SpatialPointsDataFrame,numeric,numeric,matrix,NLworlds}(turtles,
  radius, angle, agents2, world, torus = FALSE)


  \S4method{inCone}{SpatialPointsDataFrame,numeric,numeric,SpatialPointsDataFrame,NLworlds}(turtles,
  radius, angle, agents2, world, torus = FALSE)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) from which the
cone of vision is defined to select agent(s) from \code{agents2}.}

\item{radius}{Numeric. The distance(s) from \code{turtles} to locate agents from
\code{agents2}. Must be of length 1 or of length \code{turtles}.}

\item{angle}{Numeric. The angle(s) to define the size of the cone of vision
for the \code{turtles} as their heading minus \code{angle / 2}
to their heading plus \code{angle / 2}. Must be of length 1 or
of length \code{turtles}.}

\item{agents2}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patch(es) to evaluate their location(s) regarding the turtle(s)
cone of vision.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) to evaluate their
location(s) regarding the turtle(s) cone of vision.}

\item{world}{A \code{NLworlds} object representing the world where the agents
are located.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A list of length equal to \code{length(turtles)}.
        List items are either matrices (ncol = 2) with the first column \code{pxcor}
        and the second column \code{pycor} representing the coordinates of the
        patch(es) within the cone of vision of each turtle if \code{agents2} are patches,
        or SpatialPointsDataFrame objects representing the turtle(s) within
        the cone of vision of each turtle if \code{agents2} are turtles.
}
\description{
Reports the patch(es) or turtle(s) from \code{agents2} within the "cone of
vision" in front of each turtle.
}
\details{
Agents from \code{agents2} are reported if there are within \code{radius}
         distance of the turtle and their direction from the turtle is within
         \code{[-angle, + angle]} of the turtle's heading.
         Distances to patches are calculated to their center.
         If \code{torus = TRUE}, the \code{radius} distance(s) are calculared
         around the sides of the world to select \code{agents2}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10))

p1 <- inCone(turtles = t1, radius = 2, agents2 = patches(w1), angle = 90, world = w1)
t2 <- inCone(turtles = turtle(turtles = t1, who = 0), radius = 2, angle = 90, agents2 = t1, world = w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

