% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{inRadius2}
\alias{inRadius2}
\alias{inRadius2,matrix,numeric,matrix-method}
\title{Agents in radius 2}
\usage{
inRadius2(agents, radius, agents2, world, torus = FALSE)

\S4method{inRadius2}{matrix,numeric,matrix}(agents, radius, agents2, world,
  torus = FALSE)
}
\description{
Report the patches or turtles among \code{agents2} within given distances of
each of the \code{agents}.
For NLwordMs and agentMatrix and return a matrix/df instead of a list
}
\author{
Sarah Bauduin
}

