% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{inspect}
\alias{inspect}
\alias{inspect,SpatialPointsDataFrame,numeric-method}
\title{Inspect}
\usage{
inspect(turtles, who)

\S4method{inspect}{SpatialPointsDataFrame,numeric}(turtles, who)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the existing turtles
in the world.}

\item{who}{Integer. Id (who number(s)) for the turte(s) to inspect.}
}
\value{
A dataframe with the variables for the inspected turtle(s).
}
\description{
Display the variables value for the requested turtle(s).
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createOTurtles(world = w1, n = 10)
inspect(turtles = t1, who = c(2, 3))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

