% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{isNLclass}
\alias{isNLclass}
\alias{isNLclass,SpatialPointsDataFrame,character-method}
\alias{isNLclass,matrix,character-method}
\title{Test class}
\usage{
isNLclass(agents, class)

\S4method{isNLclass}{matrix,character}(agents, class)

\S4method{isNLclass}{SpatialPointsDataFrame,character}(agents, class)
}
\arguments{
\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches to evaluate.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles to evaluate.}

\item{class}{Characters. Must be equal to \code{"agent"}, \code{"agentset"},
\code{"patch"}, \code{"patchset"}. \code{"turtle"} or \code{"turtleset"}.}
}
\value{
Logical. \code{TRUE} if \code{agents} is of the \code{class} tested.
}
\description{
Reports \code{TRUE} or \code{FALSE} depending if the \code{agents} is of the
\code{class} tested.
}
\details{
Careful! The class tested do not correspond to actual R classes.
         \code{agents} is \code{"patch"} if it is a matrix (ncol = 2) with the
         first column \code{pxcor} and the second column \code{pycor} and only
         one row. \code{agents} is \code{"patches"} if the matrix has more than
         one row.
         \code{agents} is \code{"turtle"} if it is a SpatialPointsDataFrame
         of length 1 with the variables created when using \code{createTurtles()}
         or \code{createOTurtles()}. \code{agents} is \code{"turtles"} if the
         SpatialPointsDataFrame is of length larger than 1.
         \code{agents} is \code{"agent"} if it is either \code{"patch"} or
         \code{"turtle"}. \code{agents} is \code{"agentset"} if it is either
         \code{"patches"} or \code{"turtles"}.
         This function does not test if the patch(es) exist(s) within a world
         according to the coordinates. For this, use \code{isPatches()}. This
         function does not test if the turtle(s) exist(s) given their who
         number(s) and breed(s). For this, use \code{isTurtles()}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10), heading = sample(1:3, size = 10, replace= TRUE))
isNLclass(agents = patches(world = w1), class = "pacth")
isNLclass(agents = patches(world = w1), class = "pacthes")
isNLclass(agents = t1, class = "agentset")
isNLclass(agents = t1, class = "turtles")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

