% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{isPatch}
\alias{isPatch}
\title{Is patch?}
\usage{
isPatch(world, pxcor, pycor)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{pxcor}{A numeric value or vector of \code{pxcor} coordinate(s).}

\item{pycor}{A numeric value or vector of \code{pycor} coordinate(s).}
}
\value{
\code{TRUE} or \code{FALSE} if the patch(es) exist(s) or no, in the
        order of the coordinates given.
}
\description{
Reports \code{TRUE} if the patch(es) exist, \code{FALSE} otherwise
}
\examples{
# Create a NLworld
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
isPatch(world = w1, pxcor = -1, pycor = 2)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

