% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{isTurtle}
\alias{isTurtle}
\alias{isTurtle,SpatialPointsDataFrame,numeric,character-method}
\alias{isTurtle,SpatialPointsDataFrame,numeric,missing-method}
\title{Is turtle?}
\usage{
isTurtle(turtles, who, breed)

\S4method{isTurtle}{SpatialPointsDataFrame,numeric,missing}(turtles, who)

\S4method{isTurtle}{SpatialPointsDataFrame,numeric,character}(turtles, who,
  breed)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing turtles.}

\item{who}{Numeric. The who number(s) of the turtle(s) to check.}

\item{breed}{String of characters. To specify the breed of turtles to check
for existence. Must be of length 1 or of length \code{n}. If
\code{breed} is of length \code{n} then the order of breeds
should follow the order of \code{who}. If missing, there is
no distinction based upon breed to check existence.}
}
\value{
\code{TRUE} or \code{FALSE} if the turtle(s) with the given who number(s)
        and potentially given \code{breed} exist(s) or no, in the given \code{turtles}.
}
\description{
Reports \code{TRUE} if the turtle(s) exist(s), \code{FALSE} otherwise.
}
\details{
If \code{breed} is provided, the turtle with the given \code{who} number
         AND given \code{breed} must exists inside \code{turtles} for \code{TRUE}
         to be returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10), breed = c(rep("sheep", 5), rep("wolf", 5)))
isTurtle(turtles = t1, who = 3, breed = "sheep")
isTurtle(turtles = t1, who = 9, breed = "sheep")
isTurtle(turtles = t1, who = c(3, 9))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

