% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{numLayers}
\alias{numLayers}
\alias{numLayers.worldArray}
\alias{numLayers.worldMatrix}
\alias{layerNames,worldArray-method}
\alias{layerNames}
\title{Number of layers and layer names}
\usage{
numLayers(x)

\method{numLayers}{worldArray}(x)

\method{numLayers}{worldMatrix}(x)

\S4method{layerNames}{worldArray}(object)
}
\arguments{
\item{x}{A \code{worldMatrix} or \code{worldArray} object.}

\item{object}{A \code{worldArray} object.}
}
\value{
\code{numLayers} returns an integer representing the number of
layers in a \code{worldArray} or \code{worldMatrix} (which is always \code{1L})

\code{layerNames} returns an character vector representing the names
of the layers in a \code{worldArray}
}
\description{
Accessors for the layers of a \code{worldMatrix} or \code{worldArray}.
}
\examples{
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 1:25)
w2 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 25:1)
numLayers(w1)

a1 <- stackWorlds(w1, w2)
numLayers(a1)
layerNames(a1)
}
\seealso{
\code{\link[=plot.agentMatrix]{plot.agentMatrix()}}
}
