% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{layoutCircle}
\alias{layoutCircle}
\alias{layoutCircle,NLworld,SpatialPointsDataFrame,numeric-method}
\title{Layout turtles on a circle}
\usage{
layoutCircle(world, turtles, radius, torus = FALSE)

\S4method{layoutCircle}{NLworld,SpatialPointsDataFrame,numeric}(world, turtles,
  radius, torus = FALSE)
}
\arguments{
\item{world}{NLworlds object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{radius}{Numeric. Radius of the circle.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated
        coordinates and updated data for their "heading" values and
        previous coordinates "prevX"
        and "prevY".
}
\description{
Relocate the \code{turtles} on a circle centered on the \code{world}.
}
\details{
The \code{turtles} point outwards.

         If the
         \code{radius} value leads turtles outside of the \code{world}'s extent
         and \code{torus = TRUE}, they are
         relocated on the other sides of the \code{world}, inside its extent; if
         \code{torus = FALSE}, the turtles are located past
         the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")

t2 <- layoutCircle(world = w1, turtles = t1, radius = 3)
Plot(t2, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#layout-circle}
}

