% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{left}
\alias{left}
\alias{left,SpatialPointsDataFrame,numeric-method}
\title{Left}
\usage{
left(turtles, nDegrees)

\S4method{left}{SpatialPointsDataFrame,numeric}(turtles, nDegrees)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtle(s) to rotate.}

\item{nDegrees}{Numeric. The number of degrees by which to rotate the turtles'
heading to the left. Must be of length 1 or of the same length
as the \code{turtles}.}
}
\value{
A SpatialPointsDataFrame representing the turtles with updated headings.
}
\description{
Rotate the turtles's heading to the left.
}
\details{
If \code{nDegrees} is negative, the turtle rotate to the right.
}
\examples{
t1 <- createTurtles(world = w1, n = 10)
t1@data
t1 <- left(turtles = t1, nDegrees = 180)
t1@data


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

