% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPxcor}
\alias{maxPxcor}
\alias{maxPxcor,NLworld-method}
\title{Maximum pxcor}
\usage{
maxPxcor(world)

\S4method{maxPxcor}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the patches maximum \code{pxcor} in a NLworld.
}
\details{
Does the same as accessing the slot with \code{NLworld@maxPxcor}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
maxPxcor(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

