% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPycor}
\alias{maxPycor}
\alias{maxPycor,RasterLayer-method}
\alias{maxPycor,RasterStack-method}
\title{Maximum pycor}
\usage{
maxPycor(worldRaster)

\S4method{maxPycor}{RasterLayer}(worldRaster)

\S4method{maxPycor}{RasterStack}(worldRaster)
}
\arguments{
\item{worldRaster}{A \code{Raster} or \code{RasterStack} object.}
}
\value{
A numeric value
}
\description{
Report the maximum y-coordinate for patches.
}
\examples{
# Create a world with the default settings.
world <- createWorld()
maxPycor(worldRaster = world)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

