% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{minOneOf}
\alias{minOneOf}
\alias{minOneOf,NLworld,matrix,missing-method}
\alias{minOneOf,NLworldStack,matrix,character-method}
\title{One with minimum}
\usage{
minOneOf(world, agents, pVar)

\S4method{minOneOf}{NLworld,matrix,missing}(world, agents)

\S4method{minOneOf}{NLworldStack,matrix,character}(world, agents, pVar)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches to be evaluated.}

\item{pVar}{If the world is a \code{NLworldStack}, pVar is the name (characters)
of the layer used for evaluating the minimum patch value.}
}
\value{
A matrix (ncol = 2, nrow = 1) with the first column \code{pxcor} and
        the second column \code{pycor} representing the coordinates for the patch
        (or one of the patches) with the minimum value among the agents.
}
\description{
Reports one patch coordinates \code{pxcor} and \code{pycor} which has its
variable equals to the minimum value.
}
\details{
!!! Only implemented for patches so far !!!

If there are several agents with the minimum value, one is chosen randomly.
         To access to all agents with their variable equal to the minimum value,
         use \code{withMin()}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
pMin <- minOneOf(world = w1, agents = patches(world = w1))
plot(w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

