% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{minPxcor}
\alias{minPxcor}
\alias{minPxcor,NLworld-method}
\alias{minPxcor,NLworldStack-method}
\title{Minimum \code{pxcor}}
\usage{
minPxcor(world)

\S4method{minPxcor}{NLworld}(world)

\S4method{minPxcor}{NLworldStack}(world)
}
\arguments{
\item{world}{\code{NLworlds} object.}
}
\value{
Integer.
}
\description{
Report the patches minimum \code{pxcor} in a \code{NLworlds}.
}
\examples{
w1 <- createNLworld()
minPxcor(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#min-pcor}
}

