% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{NLall}
\alias{NLall}
\alias{NLall,NLworld,matrix,missing,numeric-method}
\alias{NLall,NLworldStack,matrix,character,numeric-method}
\title{All}
\usage{
NLall(world, agents, pVar, val)

\S4method{NLall}{NLworld,matrix,missing,numeric}(world, agents, val)

\S4method{NLall}{NLworldStack,matrix,character,numeric}(world, agents, pVar,
  val)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patches to be evaluated.}

\item{pVar}{If the world is a \code{NLworldStack}, pVar is the name (characters)
of the layer used for evaluating the conditional value.}

\item{val}{Numeric or character depending on the variable class.}
}
\value{
Logical. \code{TRUE} if all the agents have their variable equal to \code{val},
        return \code{FALSE} otherwise.
}
\description{
Reports \code{TRUE} if all the agents have their variable equal to a given value,
or \code{FALSE} otherwise.
}
\details{
!!! Only implemented for patches so far !!!
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
NLall(world = w1, agents = patches(world = w1), val = 5)
w2 <- w1
w2[] <- 5
NLall(world = w2, agents = patches(world = w2), val = 5)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

