% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{NLcount}
\alias{NLcount}
\alias{NLcount,SpatialPointsDataFrame-method}
\alias{NLcount,matrix-method}
\alias{count}
\title{Count agents}
\usage{
NLcount(agents)

\S4method{NLcount}{matrix}(agents)

\S4method{NLcount}{SpatialPointsDataFrame}(agents)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}
}
\value{
Integer.
}
\description{
Report the number of patches or turtles in the \code{agents}.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4) # 25 patches
p1 <- patches(w1)
NLcount(p1) # 25

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
NLcount(t1) # 10


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#count}
}

