% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{NLdist}
\alias{NLdist}
\alias{NLdist,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{NLdist,NLworlds,SpatialPointsDataFrame,matrix-method}
\alias{NLdist,NLworlds,matrix,SpatialPointsDataFrame-method}
\alias{NLdist,NLworlds,matrix,matrix-method}
\title{Distances between agents}
\usage{
NLdist(world, from, to, torus = FALSE, allPairs = FALSE)

\S4method{NLdist}{NLworlds,matrix,matrix}(world, from, to, torus = FALSE,
  allPairs = FALSE)

\S4method{NLdist}{NLworlds,matrix,SpatialPointsDataFrame}(world, from, to,
  torus = FALSE, allPairs = FALSE)

\S4method{NLdist}{NLworlds,SpatialPointsDataFrame,matrix}(world, from, to,
  torus = FALSE, allPairs = FALSE)


  \S4method{NLdist}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  from, to, torus = FALSE, allPairs = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{from}{Matrix (ncol = 2) with the first column \code{pxcor} and the
                second column \code{pycor} representing the coordinates of the
                patches from which the distances will be computed.

                SpatialPointsDataFrame created by \code{createTurtles()} or
                by \code{createOTurtles()} representing the turtles from which
                the distances will be computed.}

\item{to}{Matrix (ncol = 2) with the first column \code{pxcor} and the
                second column \code{pycor} representing the coordinates of the
                patches to which the distances will be computed.

                SpatialPointsDataFrame created by \code{createTurtles()} or
                by \code{createOTurtles()} representing the turtles to which
                the distances will be computed.

                Matrix (ncol = 2) with the first column \code{xcor} and the
                second column \code{ycor} representing the coordinates of the
                locations to which the distances will be computed.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}

\item{allPairs}{Logical. Only relevant if the number of agents/locations in
\code{from} and \code{to} is the same. If \code{FALSE}, the
distance between each agent/location in \code{from} with the
corresponding \code{to} is returned. If \code{TRUE}, a full
distance matrix is returned. Default is \code{allPairs = FALSE}.}
}
\value{
Numeric. Vector of distances if \code{from} and/or \code{to} contained
        one agent/location, or if \code{from} and \code{to} contained the same
        number of agents/locations and \code{allPairs = FALSE}.

        Matrix of distances between \code{from} (rows) and \code{to} (columns)
        if \code{from} and \code{to} are of different length, or of same length
        and \code{allPairs = TRUE}.
}
\description{
Report the distances betwenn agents \code{from} and agents or locations \code{to}.
}
\details{
Distances from/to a patch are measured from/to its center.

         If \code{torus = TRUE}, a distances around the sides of the world is
         reported only if smaller than the one across the world (i.e., as calculated
         with \code{torus = FALSE}).

         Coordinates of given \code{from} and \code{to} must be inside the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
NLdist(world = w1, from = patch(w1, 0, 0), to = patch(w1, c(1, 9), c(1, 9)))
NLdist(world = w1, from = patch(w1, 0, 0), to = patch(w1, c(1, 9), c(1, 9)), torus = TRUE)
t1 <- createTurtles(n = 2, coords = randomXYcor(w1, n = 2))
NLdist(world = w1, from = t1, to = patch(w1, c(1,9), c(1,9)), allPairs = TRUE)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#distance}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#distancexy}
}

