% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworldArray-class}
\alias{NLworldArray-class}
\title{The NLworldArray class}
\description{
This is similar to \code{NLworldStack}, but it is an s4 class extension of
\code{array}.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\seealso{
NLworldStack
}

