% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{NLworldArray}
\alias{NLworldArray}
\alias{NLworldArray,NLworldMatrix-method}
\title{Create a NLworldArray}
\usage{
NLworldArray(...)

\S4method{NLworldArray}{NLworldMatrix}(...)
}
\arguments{
\item{...}{NLworldMatrix objects.}
}
\value{
NLworldArray object.
}
\description{
Stack multiple NLworldMatrix data in an array.
}
\author{
Sarah Bauduin
}

