% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworldMatrix}
\alias{NLworldMatrix}
\alias{[}
\alias{[<-}
\title{The NLworldMatrix class}
\usage{
\S4method{[}{NLworldMatrix,numeric,numeric,ANY}(x, i, j, drop)

\S4method{[}{NLworldMatrix,missing,missing,ANY}(x, drop)

\S4method{[}{NLworldMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{NLworldMatrix,missing,missing,numeric}(x) <- value
}
\description{
This is similar to \code{NLworld}, but it is an s3 class extension of
\code{matrix}.
}
\details{
Careful: The methods \code{[]} and \code{[] <-} retrieve or assign values for
the patches (i.e., matrix cells) in the given order. When using \code{NLworlds}
objects, these methods retrive or assign values for the patches in the order
of the cell numbers as defined in \code{Raster*} objects, independently of the
given order of the patches coordinates inside the brackets.
When no patches coordinates are provided, the values retrieved or assigned
is done in the order of the cell numbers as defined in in \code{Raster*} objects.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\seealso{
NLworld
}

