% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\name{NLworldMatrix}
\alias{NLworldMatrix}
\title{The NLworldMatrix class}
\description{
This is similar to \code{NLworld}, but it is an s3 class extension of
\code{matrix}.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\seealso{
NLworld
}

