% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\name{NLworlds-class}
\alias{NLworlds}
\alias{NLworlds-class}
\title{The NLworlds class}
\description{
value <- matrix(data = value, nrow = 1, ncol = length(value))

The \code{NLworlds} class is the union of the \code{NLworld} and \code{NLworldStack}
classes. Mostly used for building function purposes.
}
\section{Slots}{

\describe{
\item{\code{members}}{NLworld, NLworldStack}

\item{\code{members}}{NLworld, NLworldStack}
}}
\author{
Sarah Bauduin

Sarah Bauduin
}

