% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworld-class.R
\docType{class}
\name{NLworldStack-class}
\alias{NLworldStack-class}
\title{The \code{NLworldStack} class}
\description{
Behaves the same as a \code{RasterStack} object except that it is a collection
of \code{NLworld} objects.
}
\author{
Sarah Bauduin
}

