% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{noPatches}
\alias{noPatches}
\title{No patches}
\usage{
noPatches(x)
}
\value{
An empty matrix (ncol = 2) with the first column \code{pxcor} and the
        second column \code{pycor}.
}
\description{
Reports an empty patch agentset.
}
\examples{
p1 <- noPatches()
nrow(p1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

