% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{noTurtles}
\alias{noTurtles}
\alias{noTurtles,missing-method}
\title{No turtles}
\usage{
noTurtles(x)

\S4method{noTurtles}{missing}()
}
\value{
An empty SpatialPointsDataFrame but with the turtle variables defined
        as when using \code{createTurtles()} or \code{createOTurtles()}.
}
\description{
Reports an empty turtle agentset.
}
\examples{
t1 <- noTurtles()
length(t1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

