% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{noTurtlesAM}
\alias{noTurtlesAM}
\alias{noTurtlesAM,missing-method}
\title{No turtles (AM)}
\usage{
noTurtlesAM(x)

\S4method{noTurtlesAM}{missing}()
}
\value{
AgentMatrix with the turtle variables defined
        as when using \code{createTurtles()} or \code{createOTurtles()} but
        with 0 individual.
}
\description{
Report an empty turtle agentset.
}
\examples{
t1 <- noTurtles()
count(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#no-turtles}
}

