% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{otherPatches}
\alias{otherPatches}
\alias{otherPatches,NLworld,matrix-method}
\alias{otherPatches,NLworldStack,matrix-method}
\title{Other patches}
\usage{
otherPatches(world, patches)

\S4method{otherPatches}{NLworld,matrix}(world, patches)

\S4method{otherPatches}{NLworldStack,matrix}(world, patches)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{patches}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches to be discarded.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates. Each row represents
        a patch and the order is the one of the cellnumbers as defined for a
        \code{Raster*} object.
}
\description{
Reports all patches coordinates \code{pxcor} and \code{pycor} except for the one(s)
with coordinates \code{pxcor} and \code{pycor} given.
}
\examples{
# Create a NLworld
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9) # 100 patches
other <- otherPatches(world = w1, patches = cbind(pxcor = 0, pycor = 0))
nrow(other) ## 99 patches left

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

