% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{otherTurtles}
\alias{otherTurtles}
\alias{otherTurtles,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\title{Other turtles}
\usage{
otherTurtles(turtles, except)

\S4method{otherTurtles}{SpatialPointsDataFrame,SpatialPointsDataFrame}(turtles,
  except)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles.}

\item{except}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) to remove from
\code{turtles}.}
}
\value{
A SpatialPointsDataFrame representing the \code{turtles} with the one(s)
        given as \code{except} removed.
}
\description{
Reports all turtles except for specific one(s).
}
\details{
Carefull: this function removes turtles only based on similar who numbers
         and breeds.
         No data is updated for the other turtles.
}
\examples{
t1 <- createTurtles(n = 10, coords = cbind(xcor = 0, ycor = 0))
t2 <- otherTurtles(turtles = t1, except = turtle(turtles = t1, who = 0))
length(t2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

