% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{patchDist}
\alias{patchDist}
\alias{patchDist,NLworld,matrix,numeric,numeric-method}
\alias{patchDist,NLworldStack,matrix,numeric,numeric-method}
\title{Patch at some distance}
\usage{
patchDist(world, agents, dist, head, torus = FALSE)

\S4method{patchDist}{NLworld,matrix,numeric,numeric}(world, agents, dist, head,
  torus = FALSE)

\S4method{patchDist}{NLworldStack,matrix,numeric,numeric}(world, agents, dist,
  head, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
calling patches.}

\item{dist}{Numeric. Distance from the caller. \code{dist} must be a single
value or of the length of \code{agents}.}

\item{head}{Numeric. Absolute angle from the caller. \code{head} must be a
single value or of the length of \code{agents}. Angle(s) are in
degrees with 0 being North.}

\item{torus}{Logical to determine if the \code{NLworld*} is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates. The order of the patches
        are the ones of the agents.
}
\description{
Reports the patch(es) coordinates \code{pxcor} and \code{pycor} at certain
distance and heading of the calling agent(s).
}
\details{
!!! Only implemented for the patches so far !!!

If \code{torus = FALSE} and the patch at distance \code{(dist)} and
         heading \code{head} of the caller is outside the world, \code{NA} is
         returned. Otherwise, if \code{torus = TRUE}, the patch coordinates from
         the wrapped world are returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
patchCorner <- patchDist(world = w1, agents = cbind(pxcor = 0, pycor = 0), dist = 1, head = 45)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

