% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{patchDistHead}
\alias{patchDistHead}
\alias{patchDistHead,NLworlds,SpatialPointsDataFrame,numeric,numeric-method}
\alias{patchDistHead,NLworlds,matrix,numeric,numeric-method}
\title{Patch at some distance and heading.}
\usage{
patchDistHead(world, agents, dist, head, torus = FALSE)

\S4method{patchDistHead}{NLworlds,matrix,numeric,numeric}(world, agents, dist,
  head, torus = FALSE)


  \S4method{patchDistHead}{NLworlds,SpatialPointsDataFrame,numeric,numeric}(world,
  agents, dist, head, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches from wich \code{dist} is computed.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles from wich
\code{dist} is computed.}

\item{dist}{Numeric. Distance from the agent. \code{dist} must be a single
value or of the length of \code{agents}.}

\item{head}{Numeric. Absolute angle from the agent. \code{head} must be a
single value or of the length of \code{agents}. Angle(s) are in
degrees with 0 being North.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates. The order of the patches
        followa the order of the agents.
}
\description{
Reports the patch(es) coordinates \code{pxcor} and \code{pycor} at certain
distance and heading of the agent(s).
}
\details{
If \code{torus = FALSE} and the patch at distance \code{dist} and
         heading \code{head} of the agent is outside the world's extent, \code{NA}
         is returned. If \code{torus = TRUE}, the patch coordinates from the
         wrapped world are returned.
         If \code{agents} are turtles, their heading is not taken into account; the
         given absolute heading \code{head} is used. To find a patch at certain
         distance from a turtle with the turtle's heading, look at \code{pacthAhead()},
         \code{pacthLeft()} or \code{pacthRight()}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
patchCorner <- patchDistHead(world = w1, agents = cbind(pxcor = 0, pycor = 0), dist = 1, head = 45)
t1 <- createTurtles(world = w1, n = 1, coords = cbind(xcor = 0, ycor = 0), heading = 315)
patchCorner <- patchDistHead(world = w1, agents = t1, dist = 1, head = 45)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

