% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{patchHere}
\alias{patchHere}
\alias{patchHere,NLworlds,SpatialPointsDataFrame-method}
\title{Patch here}
\usage{
patchHere(world, turtles)

\S4method{patchHere}{NLworlds,SpatialPointsDataFrame}(world, turtles)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles evolve onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the moving turtles.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates at the \code{turtles}
        location. The order of the patches follows the order of the \code{turtles}.
}
\description{
Reports the patch(es) coordinates under the turtle(s).
}
\details{
If a turtle is located outside of the world's extent, \code{NA} is returned
         for its patch coordinates.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10))
patchHere(world = w1, turtles = t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

