% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{patchRight}
\alias{patchRight}
\alias{patchRight,NLworlds,SpatialPointsDataFrame,numeric,numeric-method}
\title{Patch on the right}
\usage{
patchRight(world, turtles, dist, nDegrees, torus = FALSE)

\S4method{patchRight}{NLworlds,SpatialPointsDataFrame,numeric,numeric}(world,
  turtles, dist, nDegrees, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the turtles
move onto.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{dist}{Numeric. Distances from the \code{turtles} locations. Must be of
length 1 or of length \code{turtles}.}

\item{nDegrees}{Numeric. The number of degrees the turtle's heading should rotate
to the right to locate the patches. Must be of length 1 or of
length \code{turtles}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
Matrix (ncol = 2) with the first column \code{pxcor} and the second
        column \code{pycor} representing the patches coordinates at \code{dist}
        of the turtles locations and \code{nDegrees} to the right of their headings.
        The order of the patches follows the order of the \code{turtles}.
}
\description{
Report the patches coordinates \code{pxcor} and \code{pycor} at a given distances
to the right of the turtles.
}
\details{
If \code{nDegrees} is negative, the turtle would rotate to the left.
         If \code{dist} is negative, the turtle would move backward.

         If \code{torus = FALSE} and the patch at distance \code{dist} and
         heading \code{nDregrees} to the right of the turtles is outside the
         world's extent, \code{NA} are returned for the patch coordinate.
         If \code{torus = TRUE}, the patch coordinates from the wrapped world
         are returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 1, coords = cbind(xcor = 2, ycor = 2), heading = 90)
patchRight(world = w1, turtles = t1, dist = 2, nDegrees = 90)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#patch-lr-and-ahead}
}

