% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{patchSet}
\alias{patchSet}
\alias{patchSet,matrix-method}
\title{Patch set}
\usage{
patchSet(...)

\S4method{patchSet}{matrix}(...)
}
\arguments{
\item{...}{Matrices (ncol = 2) of patch(es) coordinates with the first column
\code{pxcor} and the second column \code{pycor}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates.
}
\description{
Reports a patch agentset as the the coordinates \code{pxcor} and \code{pycor}
of all patches in the inputs.
}
\details{
Duplicate patches among the inputs are removed in the return matrix.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
p1 <- patchAt(world = w1, agents = cbind(pxcor = 0, pycor = 0), dx = 1, dy = 1)
p2 <- patchDist(world = w1, agents = cbind(pxcor = 0, pycor = 0), dist = 3, head = 45)
p3 <- patchAt(world = w1, xcor = 4.3, ycor = 8)
set1 <- patchSet(p1, p2, p3)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

