% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{pExist}
\alias{pExist}
\alias{pExist,NLworld,numeric,numeric-method}
\alias{pExist,NLworldStack,numeric,numeric-method}
\title{Do the patches exist?}
\usage{
pExist(world, pxcor, pycor)

\S4method{pExist}{NLworld,numeric,numeric}(world, pxcor, pycor)

\S4method{pExist}{NLworldStack,numeric,numeric}(world, pxcor, pycor)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{pxcor}{Integers. Vector of \code{pxcor} coordinates. Must be of the same
length as \code{pycor}.}

\item{pycor}{Integers. Vector of \code{pycor} coordinates. Must be of the same
length as \code{pxcor}.}
}
\value{
Logicals. \code{TRUE} or \code{FALSE} if the patches exist inside the
        world's extent, in the order of the coordinates given.
}
\description{
Report \code{TRUE} if the patches exist in the world's extent, report
\code{FALSE} otherwise.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
pExist(world = w1, pxcor = -1, pycor = 2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

