% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\name{.projNowhere}
\alias{.projNowhere}
\title{Internal CRS usage}
\usage{
.projNowhere
}
\description{
This is used in various places when converting \code{agentMatrix} objects to
\code{SpatialPoints} for use with functions, such as \code{raster::buffer()}.
None of these functions results in any real geospatial information being conserved,
other than distance and angle  calculations. Buffering still occurs on the
unit-less World. This is therefore a way to allow \code{sp} and \code{raster}
functions to work \emph{as if a cell in the World was equal to 1 m}. The resulting
answers will be converted back to units of "cells", not "m".
The \code{CRS()} operation is relatively time consuming.
Having a single object usable throughout is \emph{much} faster.
}
\keyword{internal}
