% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot.R
\name{numLayers.worldArray}
\alias{numLayers.worldArray}
\alias{numLayers.worldMatrix}
\alias{layerNames,worldArray-method}
\alias{.identifyGrobToPlot,worldArray,.quickPlotGrob-method}
\title{Methods for \code{quickPlot}}
\usage{
\method{numLayers}{worldArray}(x)

\method{numLayers}{worldMatrix}(x)

\S4method{layerNames}{worldArray}(object)

\S4method{.identifyGrobToPlot}{worldArray,.quickPlotGrob}(toPlot, sGrob, takeFromPlotObj)
}
\arguments{
\item{x}{An object or list of objects.}

\item{object}{An object from which to extract the layer names.}

\item{toPlot}{The object to plot. Should be a single layer if from a multi-layer
object such as a \code{RasterStack}.}

\item{sGrob}{\code{quickPlot} grob object}

\item{takeFromPlotObj}{Logical. Should the data come from the argument passed
into \code{Plot} (\code{TRUE}), or from the (\code{.quickPlotEnv}) (\code{FALSE}).}
}
\value{
\code{numLayers} returns an integer representing the number of
layers in a \code{worldArray} or \code{worldMatrix} (which is always \code{1L})

\code{layerNames} returns an character vector representing the names
of the layers in a \code{worldArray}
}
\description{
These are required to create plotting methods to work with \pkg{quickPlot}.
}
