% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot.R
\name{.quickPlottables-class}
\alias{.quickPlottables-class}
\alias{quickPlottables}
\title{\code{quickPlot} classes}
\description{
\pkg{quickPlot} offers a type of plotting that is modular.
Users of NetLogoR may find this useful for simulation modeling.
We have put in place the required methods and imported the appropriate classes
to use the \code{quickPlot::Plot} function.
Users can still use \code{plot} from the \pkg{graphics} package, but it is not modular.
}
\details{
This adds \code{agentMatrix} to the \code{.quickPlottables}, \code{.quickObjects},
and \code{spatialObjects}.

This adds \code{worldMatrix} to the \code{.quickPlottables}, \code{.quickObjects},
\code{spatialObjects} and \code{griddedClasses}.
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.quickPlotObjects}} and \code{\link{.quickPlot}}}
}}

\seealso{
\code{\link{quickPlotClasses}}
}
\author{
Eliot McIntire
}
\keyword{internal}
