% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{randomPxcor}
\alias{randomPxcor}
\alias{randomPxcor,NLworlds,numeric-method}
\title{Random \code{pxcor}}
\usage{
randomPxcor(world, n)

\S4method{randomPxcor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{\code{NLworld} object.}

\item{n}{Integer. The number of random \code{pxcor} to generate.}
}
\value{
Integers.
}
\description{
Report random \code{pxcor} coordinates between a world's minPxcor and maxPxcor.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
pxcor10 <- randomPxcor(world = w1, n = 10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#random-pcor}
}

