% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{randomPycor}
\alias{randomPycor}
\alias{randomPycor,NLworlds,numeric-method}
\title{Random \code{pycor}}
\usage{
randomPycor(world, n)

\S4method{randomPycor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{n}{Integer. The number of random \code{pycor} to generate.}
}
\value{
Integers.
}
\description{
Report random \code{pycor} coordinates between minPycor and maxPycor.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
pycor10 <- randomPycor(world = w1, n = 10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#random-pcor}
}

