% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomXYcor}
\alias{randomXYcor}
\alias{randomXYcor,NLworldMs,numeric-method}
\title{Random turtles coordinates}
\usage{
randomXYcor(world, n)

\S4method{randomXYcor}{NLworldMs,numeric}(world, n)
}
\arguments{
\item{world}{NLworldMs object.}

\item{n}{Integer.}
}
\value{
Matrix (ncol = 2, nrow = \code{n}) with the first column "xcor" and the second
        column "ycor".
}
\description{
Report \code{n} random xcor and ycor coordinates within the \code{world}'s extent.
}
\examples{
w1 <- createNLworldMatrix(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = runif(25))
t1 <- createTurtlesAM(n = 10, coords = randomXYcor(world = w1, n = 10))
plot(world2raster(w1))
points(turtles2spdf(t1), col = of(agents = t1, var = "color"), pch = 16)


}
\author{
Sarah Bauduin
}

