% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{randPxcor}
\alias{randPxcor}
\alias{randPxcor,NLworld,numeric-method}
\alias{randPxcor,NLworldStack,numeric-method}
\title{Random pxcor}
\usage{
randPxcor(world, n)

\S4method{randPxcor}{NLworld,numeric}(world, n)

\S4method{randPxcor}{NLworldStack,numeric}(world, n)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{n}{Integer. Represent how many random pxcor needs to be generated.}
}
\value{
A vector of integer.
}
\description{
Reports random pxcor coordinate(s) between minPxcor and maxPxcor.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
pxcor10 <- randPxcor(world = w1, n = 10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

