% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{randPycor}
\alias{randPycor}
\alias{randPycor,NLworld,numeric-method}
\alias{randPycor,NLworldStack,numeric-method}
\title{Random pycor}
\usage{
randPycor(world, n)

\S4method{randPycor}{NLworld,numeric}(world, n)

\S4method{randPycor}{NLworldStack,numeric}(world, n)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{n}{Integer. Represent how many random pycor needs to be generated.}
}
\value{
A vector of integer.
}
\description{
Reports random pycor coordinate(s) between minPycor and maxPycor.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
pycor10 <- randPycor(world = w1, n = 10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

