% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{raster2world}
\alias{raster2world}
\alias{raster2world,RasterLayer,character-method}
\alias{raster2world,RasterStack,character-method}
\title{Convert a Raster* object into a NLworldMs object}
\usage{
raster2world(raster, method)

\S4method{raster2world}{RasterLayer,character}(raster, method)

\S4method{raster2world}{RasterStack,character}(raster, method)
}
\arguments{
\item{raster}{RasterLayer or RasterStack object.}

\item{method}{"ngb" or "bilinear" for the resample method}
}
\value{
NLworldMatrix or NLworldArray object depending on the input.
        Patches value are retained from the Raster* object.
}
\description{
Convert a RasterLayer object into a NLworldMatrix object or a RasterStack object
into a NLworldArray object.
}
\details{
See \code{help("NLworldMs-class")} for more details on the NLworlds
         classes.

         The \code{raster} is resampled to match the coordinates system and
         resolution of a NLworldMs using the nearest neighbor method. The
         extent will be bigger by 1 on the width and on the height.
}
\author{
Sarah Bauduin
}

