% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{resizeWorld}
\alias{resizeWorld}
\alias{resizeWorld,NLworlds,numeric,numeric,numeric,numeric,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworlds,numeric,numeric,numeric,numeric,missing-method}
\title{Resize a world}
\usage{
resizeWorld(world, minPxcor, maxPxcor, minPycor, maxPycor, turtles)


  \S4method{resizeWorld}{NLworlds,numeric,numeric,numeric,numeric,SpatialPointsDataFrame}(world,
  minPxcor, maxPxcor, minPycor, maxPycor, turtles)

\S4method{resizeWorld}{NLworlds,numeric,numeric,numeric,numeric,missing}(world,
  minPxcor, maxPxcor, minPycor, maxPycor)
}
\arguments{
\item{world}{NLworlds object.}

\item{minPxcor}{Integer. Minimum pxcor for the patches (world's left border).}

\item{maxPxcor}{Integer. Maximum pxcor for the patches (world's right border).}

\item{minPycor}{Integer. Minimum pycor for the patches (world's bottom border).}

\item{maxPycor}{Integer. Maximum pycor for the patches (world's top border).}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}
}
\value{
NLworld object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patch value are \code{NA}.
}
\description{
Change the size of the \code{world} and cleat the patches and given \code{turtles}.
}
\details{
The given \code{world} and \code{turtles} are deleted, as well as the
         patches values contained in the \code{world}.
}
\examples{
w1 <- createNLworld()
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, world = w1)
w2 <- resizeWorld(world = w1, minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9, turtles = t1)
w2[] <- runif(length(w2))
plot(w1) # does not work, 'w1' does not exist anymore
plot(w2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#resize-world}
}

