% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\name{show,worldArray-method}
\alias{show,worldArray-method}
\alias{show,worldMatrix-method}
\title{Key base R functions for \code{worldNLR} classes}
\usage{
\S4method{show}{worldArray}(object)

\S4method{show}{worldMatrix}(object)
}
\arguments{
\item{object}{An \code{agentMatrix} object.}
}
\description{
Slight modifications from the default versions.
}
