% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{SpatialPoints2}
\alias{SpatialPoints2}
\title{SpatialPoints2}
\usage{
SpatialPoints2(coords, proj4string = CRS(as.character(NA)), bbox = NULL)
}
\arguments{
\item{coords}{needs documentation}

\item{proj4string}{needs documentation}

\item{bbox}{needs documentation}
}
\description{
needs documentation
}

