% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassesUnused.R
\docType{methods}
\name{SpatialPointsDataTable}
\alias{SpatialPointsDataTable}
\alias{SpatialPointsDataTable,ANY-method}
\title{Create a SpatialPointsDataTable}
\usage{
SpatialPointsDataTable(coords, data, coords.nrs = numeric(0),
  proj4string = CRS(NA_character_), match.ID, bbox = NULL)

\S4method{SpatialPointsDataTable}{ANY}(coords, data, coords.nrs = numeric(0),
  proj4string = CRS(NA_character_), match.ID, bbox = NULL)
}
\arguments{
\item{coords}{description needed}

\item{data}{description needed}

\item{coords.nrs}{description needed}

\item{proj4string}{description needed}

\item{match.ID}{description needed}

\item{bbox}{description needed}
}
\description{
Description needed
}

