% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{sprout}
\alias{sprout}
\alias{sprout,NLworlds,matrix,numeric-method}
\title{Sprout}
\usage{
sprout(world, patches, n, turtles, breed, heading, color)

\S4method{sprout}{NLworlds,matrix,numeric}(world, patches, n, turtles, breed,
  heading, color)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles are created on.}

\item{patches}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) on which the new turtles are created. \code{nrow(patches)}
must be equal to 1 or to \code{n}.}

\item{n}{Integer. Number of new turtle(s) to create.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the existing turtles
in the world.}

\item{...}{Additional arguments (see details).}
}
\value{
A SpatialPointsDataFrame with the created turtles.
}
\description{
Create new turtles on specific patch(es).
}
\details{
The additional arguments to be passed on are some of the function
         \code{createTurtles()} which are: heading, breed and color. If not
         provided, these arguments take the default value as defined in
         \code{createTurtles()}.
         If \code{turtles} is provided, the new created turtles are added to
         the \code{turtles} when returned. The who number of the created turtles
         therefore follow the ones from the \code{turtles}. If no \code{turtles}
         is provided, a new SpatialPointsDataFrame is created and the who number
         starts at 0.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- sprout(world = w1, patches = cbind(pxcor = 2, pycor = 2), n = 3)
t2 <- sprout(world = w1, patches = cbind(pxcor = 3, pycor = 3), n = 3, turtles = t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

