% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{sproutAM}
\alias{sproutAM}
\alias{sproutAM,numeric,matrix-method}
\title{Sprout new turtles (AM)}
\usage{
sproutAM(n, patches, breed, heading, color, turtles)

\S4method{sproutAM}{numeric,matrix}(n, patches, breed, heading, color, turtles)
}
\arguments{
\item{n}{Integer.}

\item{patches}{Matrix (ncol = 2) with the first column "pxcor" and the second
column "pycor" representing the patches coordinates.}

\item{breed}{Character. Vector of "breed" names. Must be of length 1 or of length
\code{n}. If missing, \code{breed = "turtle"} for all turtles.}

\item{heading}{Numeric. Vector of values between 0 and 360. Must be of length 1 or
of length \code{n}. If missing, a random heading is assigned to
each turtle.}

\item{color}{Character. Vector of color names. Must be of length \code{n}.
If missing, colors are assigned using the function \code{rainbow(n)}.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}
}
\value{
AgentMatrix including the new
        sprouted turtles.
}
\description{
Create \code{n} new turtles on specific \code{patches}.
}
\details{
\code{nrow(patches)} must be equal to 1 or to \code{n}.

         If \code{turtles} is provided, the new turtles are added to
         the \code{turtles} when returned. The who numbers of the sprouted turtles
         therefore follow the ones from the \code{turtles}. If no \code{turtles}
         is provided, a new agentMatrix is created and the who numbers
         start at 0.
}
\examples{
t1 <- sproutAM(patches = cbind(pxcor = 2, pycor = 2), n = 3)
t2 <- sproutAM(patches = cbind(pxcor = 3, pycor = 3), n = 3, turtles = t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#sprout}
}

