% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-classes.R
\docType{methods}
\name{stackWorlds}
\alias{stackWorlds}
\alias{stackWorlds,NLworldMatrix-method}
\title{Stack worlds}
\usage{
stackWorlds(...)

\S4method{stackWorlds}{NLworldMatrix}(...)
}
\arguments{
\item{...}{NLworldMatrix objects.}
}
\value{
NLworldArray object.
}
\description{
Stack multiple NLworldMatrix into a NLworldArray.
}
\details{
The NLworldMatrix objects must all have the same extents.
}
\examples{
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 1:25)
w2 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 25:1)
w3 <- stackWorlds(w1, w2)
plot(world2raster(w3))

}
\author{
Sarah Bauduin
}

