% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{subHeadings}
\alias{subHeadings}
\alias{subHeadings,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{subHeadings,SpatialPointsDataFrame,numeric-method}
\alias{subHeadings,numeric,SpatialPointsDataFrame-method}
\alias{subHeadings,numeric,numeric-method}
\title{Substract headings}
\usage{
subHeadings(angle1, angle2, range360 = FALSE)

\S4method{subHeadings}{numeric,numeric}(angle1, angle2, range360 = FALSE)

\S4method{subHeadings}{SpatialPointsDataFrame,numeric}(angle1, angle2,
  range360 = FALSE)

\S4method{subHeadings}{numeric,SpatialPointsDataFrame}(angle1, angle2,
  range360 = FALSE)

\S4method{subHeadings}{SpatialPointsDataFrame,SpatialPointsDataFrame}(angle1,
  angle2, range360 = FALSE)
}
\arguments{
\item{angle1}{SpatialPointsDataFrame created by \code{createTurtles()} or by
                \code{createOTurtles()} representing the turtles with their
                headings.

                Numeric.}

\item{angle2}{SpatialPointsDataFrame created by \code{createTurtles()} or by
                \code{createOTurtles()} representing turtles with their
                headings.

                Numeric.

                \code{angle2} must be of length 1 or of length \code{angle1}.}

\item{range360}{Logical. Determine if the returned values are between 0째 and
360째 (\code{range360 = TRUE}) or between -180째 and 180째
(\code{range360 = FALSE}). Default is \code{range360 = FALSE}.}
}
\value{
Numeric. Vector of values representing the smallest angles in degrees
        by which \code{angle1} could be rotated to produce \code{angle2}
        (i.e., the target heading). Note: this is the opposite as in NetLogo where
        angle1 is the target.
}
\description{
Compute the difference between headings.
}
\details{
Positive values mean clockwise rotation, negative value mean counterclockwise.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createOTurtles(n = 10, world = w1)
subHeadings(angle1 = t1, angle2 = 0)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#subtract-headings}
}

