% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{tExist}
\alias{tExist}
\alias{tExist,SpatialPointsDataFrame,numeric,character-method}
\alias{tExist,SpatialPointsDataFrame,numeric,missing-method}
\title{Do the turtles exist?}
\usage{
tExist(turtles, who, breed)

\S4method{tExist}{SpatialPointsDataFrame,numeric,missing}(turtles, who)

\S4method{tExist}{SpatialPointsDataFrame,numeric,character}(turtles, who, breed)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{who}{Integer. Vector of the "who" numbers of the turtles to check for existence.}

\item{breed}{Characters. Vector of "breed" names for the turtles to check
for existence. Must be of length 1 or of length \code{n}.
If missing, there is
no distinction based upon "breed".}
}
\value{
Logical. Vector of \code{TRUE} or \code{FALSE} if the turtles with
        the given \code{who} numbers and potentially given \code{breed} exist or not
        in the given \code{turtles} agentset.
}
\description{
Report \code{TRUE} if a turtle exists inside the \code{turtles} agentset, report
\code{FALSE} otherwise.
}
\details{
If \code{breed} is provided, the turtle with the given \code{who} number
         AND given \code{breed} must exists inside \code{turtles} for \code{TRUE}
         to be returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10), breed = c(rep("sheep", 5), rep("wolf", 5)))
tExist(turtles = t1, who = 3, breed = "sheep")
tExist(turtles = t1, who = 9, breed = "sheep")
tExist(turtles = t1, who = c(3, 9))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#member}
}

