% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{tExists}
\alias{tExists}
\alias{tExists,SpatialPointsDataFrame,numeric,character-method}
\alias{tExists,SpatialPointsDataFrame,numeric,missing-method}
\title{Do the turtles exist?}
\usage{
tExists(turtles, who, breed)

\S4method{tExists}{SpatialPointsDataFrame,numeric,missing}(turtles, who)

\S4method{tExists}{SpatialPointsDataFrame,numeric,character}(turtles, who,
  breed)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtle agentset.}

\item{who}{Integers. The who numbers of the turtles to check.}

\item{breed}{Characters. To specify the breed of turtles to check
for existence. Must be of length 1 or of length \code{n}. If
\code{breed} is of length \code{n} then the order of breeds
should follow the order of \code{who}. If missing, there is
no distinction based upon breed to check existence.}
}
\value{
Logicals. Vector of \code{TRUE} or \code{FALSE} if the turtles with
        the given who numbers and potentially given \code{breed} exists or not
        in the given \code{turtles}.
}
\description{
Report \code{TRUE} if the turtles exist inside the turtle agenset, report
\code{FALSE} otherwise.
}
\details{
If \code{breed} is provided, the turtle with the given \code{who} number
         AND given \code{breed} must exists inside \code{turtles} for \code{TRUE}
         to be returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10), breed = c(rep("sheep", 5), rep("wolf", 5)))
tExists(turtles = t1, who = 3, breed = "sheep")
tExists(turtles = t1, who = 9, breed = "sheep")
tExists(turtles = t1, who = c(3, 9))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

