% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{towards}
\alias{towards}
\alias{towards,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,SpatialPointsDataFrame,matrix-method}
\alias{towards,NLworlds,matrix,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,matrix,matrix-method}
\title{Direction towards}
\usage{
towards(world, agents1, agents2, torus = FALSE)

\S4method{towards}{NLworlds,matrix,matrix}(world, agents1, agents2,
  torus = FALSE)

\S4method{towards}{NLworlds,SpatialPointsDataFrame,matrix}(world, agents1,
  agents2, torus = FALSE)

\S4method{towards}{NLworlds,matrix,SpatialPointsDataFrame}(world, agents1,
  agents2, torus = FALSE)


  \S4method{towards}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  agents1, agents2, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{agents1}{Matrix (ncol = 2) with the first column \code{pxcor} and the
             second column \code{pycor} representing the coordinates of the
             patches from which the directions are computed.

             SpatialPointsDataFrame created by \code{createTurtles()} or by
             \code{createOTurtles()} representing the turtles from which the
             directions are computed.}

\item{agents2}{Matrix (ncol = 2) with the first column \code{pxcor} and the
             second column \code{pycor} representing the coordinates of the
             patches towards which the directions are computed.

             SpatialPointsDataFrame created by \code{createTurtles()} or by
             \code{createOTurtles()} representing the turtles towards which the
             directions are be computed.

             Matrix (ncol = 2) with the first column \code{x} and the second
             column \code{y} representing the coordinates of the locations
             towards which the directions are computed.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
Numeric. Vector of angles in degrees of length equal to the largest
        number of agents/locations between \code{agents1} and \code{agents2}.
}
\description{
Report the directions of \code{agents1} towards agents or locations \code{agents2}.
}
\details{
\code{agents1} and \code{agents2} must be of the same length or if different, one
         of the two has be of length 1.

         If \code{torus = TRUE} and the distance from one \code{agents1} to
         its corresponding agent/location \code{agents2} is smaller around the
         sides of the world than across it, then the direction to the agent/location
         going around the sides of the world is reported.

         The direction from a patch to its location returns 0, the direction from
         a turtle to its location returns the turtle's heading.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
towards(world = w1, agents1 = patches(w1), agents2 = cbind(x = 0, y = 0))
t1 <- createTurtles(n = 10, world = w1)
towards(world = w1, agents1 = t1, agents2 = cbind(x = 0, y = 0))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#towards}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#towardsxy}
}

