% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtleSet}
\alias{turtleSet}
\alias{turtleSet,SpatialPointsDataFrame-method}
\title{Create a turtle agenset}
\usage{
turtleSet(...)

\S4method{turtleSet}{SpatialPointsDataFrame}(...)
}
\arguments{
\item{...}{SpatialPointsDataFrame objects created by \code{createTurtles()} or
by \code{createOTurtles()} representing turtles.}
}
\value{
SpatialPointsDataFrame with all the turtles provided in the inputs.
}
\description{
Report a turtle agentset containing all turtles provided in the inputs.
}
\details{
This functions does not affect turtles coordinates and variables.
         Therefore there may be multiple turtles with the same variable (e.g.,
         who number, color, ...) . This must be taken care of prior or later
         using this function.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10), breed = "sheep")
t2 <- createTurtles(n = 2, coords = randomXYcor(w1, n = 2), breed = "wolf")
t3 <- createTurtles(n = 1, coords = randomXYcor(w1, n = 1), breed = "sheperd")
tAll <- turtleSet(t1, t2, t3)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#turtle-set}
}

