% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtlesOn}
\alias{turtlesOn}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,character-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,missing-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,matrix,character-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,matrix,missing-method}
\title{Turtles on}
\usage{
turtlesOn(world, turtles, agents, breed)

\S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,matrix,missing}(world,
  turtles, agents)

\S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,matrix,character}(world,
  turtles, agents, breed)


  \S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,missing}(world,
  turtles, agents)


  \S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,character}(world,
  turtles, agents, breed)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles evolve onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles among which the
one(s) on the same patch(es) as the \code{agents} is/are going to
be reported.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) where to look for turtles.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing turtles whose patch location
is/are the patch(es) where to look for turtles.
Patch(es) or turtle(s) location must be inside the world's extent.}

\item{breed}{String of characters. To specify the breed(s) of turtles to report.
If missing, there is no distinction based upon breed to report turtle(s).}
}
\value{
A SpatialPointsDataFrame representing the turtle(s) among \code{turtles}
        which are located on the \code{agent} patch(es) and of the given \code{breed}
        if specified.
}
\description{
Reports the turtle(s) at some patch(es) or on the same patch(es) as some turtle(s).
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(100)
t1 <- createTurtles(n = 500, coords = randomXYcor(world = w1, n = 500))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")

t2 <- turtlesOn(world = w1, turtles = t1, agents = patch(world = w1, xcor = 2, ycor = 2))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

