% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{withMax}
\alias{withMax}
\alias{withMax,NLworld,matrix,missing-method}
\alias{withMax,NLworldStack,matrix,character-method}
\title{With maximum}
\usage{
withMax(world, agents, pVar)

\S4method{withMax}{NLworld,matrix,missing}(world, agents)

\S4method{withMax}{NLworldStack,matrix,character}(world, agents, pVar)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches to be evaluated.}

\item{pVar}{If the world is a \code{NLworldStack}, pVar is the name (characters)
of the layer used for evaluating the maximum patch value.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the coordinates for the patches with the maximum
        value among the agents.
}
\description{
Reports the coordinates \code{pxcor} and \code{pycor} of the patches which
have their variable equals to the maximum value.
}
\details{
!!! Only implemented for patches so far !!!
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
pMax <- withMax(world = w1, agents = patches(world = w1))
plot(w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

