% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{world2raster}
\alias{world2raster}
\alias{world2raster,NLworldArray-method}
\alias{world2raster,NLworldMatrix-method}
\title{Convert a NLworldMs object into a Raster* object}
\usage{
world2raster(world)

\S4method{world2raster}{NLworldMatrix}(world)

\S4method{world2raster}{NLworldArray}(world)
}
\arguments{
\item{world}{NLworldMatrix or NLworldArray object}
}
\value{
RasterLayer or RasterStack object depending on the input.
        Patches value are retained from the NLworldMs object.
}
\description{
Convert a NLworldMatrix object into a RasterLayer object or a
NLworldArray object into a RasterStack object
}
\author{
Sarah Bauduin
}

