% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{worldHeight}
\alias{worldHeight}
\alias{worldHeight,NLworld-method}
\title{NLworld height}
\usage{
worldHeight(world)

\S4method{worldHeight}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the height of the \code{NLworld} in patch number.
}
\details{
It equals \code{NLworld@maxPycor - NLworld@minPycor + 1}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
worldHeight(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

