% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-classes.R
\name{worldNLR-class}
\alias{worldNLR}
\alias{worldNLR-class}
\title{The worldNLR class}
\description{
The \code{worldNLR} class is the union of the \code{worldMatrix} and \code{worldArray}
classes. Mostly used for building function purposes.
}
\section{Slots}{

\describe{
\item{\code{members}}{worldMatrix, worldArray}
}}
\author{
Sarah Bauduin, and Eliot McIntire
}

