% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{worldWidth}
\alias{worldWidth}
\alias{worldWidth,NLworld-method}
\title{NLworld width}
\usage{
worldWidth(world)

\S4method{worldWidth}{NLworld}(world)
}
\arguments{
\item{world}{A \code{NLworld} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the width of the \code{NLworld} in patch number.
}
\details{
It equals \code{NLworld@maxPxcor - NLworld@minPxcor + 1}
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
worldWidth(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

