% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent-classes.R, R/Classes.R
\docType{class}
\name{agentClasses-class}
\alias{agentClasses-class}
\alias{agentClasses}
\title{A meta class for \code{agentMatrix} and \code{SpatialPointsDataFrame}}
\description{
Both these types can be used by NetLogoR to describe turtle agents.

Both these types can be used by NetLogoR to describe turtle agents.
}
\author{
Eliot McIntire

Eliot McIntire
}
