% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{cbind}
\alias{cbind}
\alias{cbind.agentMatrix}
\alias{rbind}
\alias{rbind.agentMatrix}
\title{Combine R Objects by Rows or Columns}
\usage{
\method{cbind}{agentMatrix}(..., deparse.level)

\method{rbind}{agentMatrix}(..., deparse.level = 1)
}
\arguments{
\item{...}{Two \code{agentMatrix} objects.}

\item{deparse.level}{See \code{\link[base]{cbind}}.}
}
\value{
An \code{agentMatrix} object.
}
\description{
Take a sequence of \code{agentMatrix} arguments and combine by columns or rows, respectively.
This will take the coordinates of the first argument and remove the coordinates
of the second object.
}
