% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{show,agentMatrix-method}
\alias{show,agentMatrix-method}
\alias{length,agentMatrix-method}
\alias{nrow,agentMatrix-method}
\alias{head}
\alias{head.agentMatrix}
\alias{tail}
\alias{tail.agentMatrix}
\title{Key base R functions for \code{agentMatrix} class}
\usage{
\S4method{show}{agentMatrix}(object)

\S4method{length}{agentMatrix}(x)

\S4method{nrow}{agentMatrix}(x)

\method{head}{agentMatrix}(x, n = 6L, ...)

\method{tail}{agentMatrix}(x, n = 6L, ...)
}
\arguments{
\item{object}{An \code{agentMatrix} object.}

\item{x}{An \code{agentMatrix} object.}

\item{n}{documentation needed}

\item{...}{documentation needed}
}
\description{
Slight modifications from the default versions.
}
