% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{show,agentMatrix-method}
\alias{show,agentMatrix-method}
\alias{length,agentMatrix-method}
\alias{nrow,agentMatrix-method}
\alias{head}
\alias{head.agentMatrix}
\alias{tail}
\alias{tail.agentMatrix}
\title{Key base R functions for \code{agentMatrix} class}
\usage{
\S4method{show}{agentMatrix}(object)

\S4method{length}{agentMatrix}(x)

\S4method{nrow}{agentMatrix}(x)

\method{head}{agentMatrix}(x, n = 6L, ...)

\method{tail}{agentMatrix}(x, n = 6L, ...)
}
\arguments{
\item{object}{An \code{agentMatrix} object.}

\item{x}{An \code{agentMatrix} object}

\item{n}{an integer vector of length up to dim(x) (or 1, for non-dimensioned objects).}

\item{...}{arguments to be passed to or from other methods (currently, none used).}
}
\value{
\code{show} returns an invisible \code{NULL}.
\code{length} returns a non-negative integer of length 1,
except for vectors of more than 2^31 - 1 elements, when it returns a double.
\code{nrow} returns an integer of length 1 or \code{NULL}.

An \code{agentMatrix} object, like \code{x}, but generally smaller.
}
\description{
Slight modifications from the default versions.
}
