% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{[}
\alias{[}
\alias{cbind}
\alias{cbind.agentMatrix}
\alias{head}
\alias{head.agentMatrix}
\alias{tail}
\alias{tail.agentMatrix}
\title{Methods for agentMatrix class objects}
\usage{
\S4method{nrow}{agentMatrix}(x)

head.agentMatrix(x, n = 6L, ...)

tail.agentMatrix(x, n = 6L, ...)

cbind.agentMatrix(..., deparse.level)

\S4method{length}{agentMatrix}(x)
}
\description{
These are NetLogoR specific methods.
}

