% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{agentMatrix}
\alias{agentMatrix}
\title{agentMatrix class}
\usage{
agentMatrix(coords, ..., coords.nrs = numeric(0),
  proj4string = CRS(as.character(NA)), match.ID, bbox = NULL)
}
\description{
This is incomplete.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(world = w1, n = 10, agent = TRUE)

library(microbenchmark)
N = 1e5
microbenchmark(createTurtles(world = w1, n = N, agent = TRUE),
              createTurtles(world = w1, n = N, agent = FALSE))
}
\author{
Eliot McIntire
}

