% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{agentMatrix}
\alias{[}
\alias{[<-}
\alias{agentMatrix}
\alias{agentMatrix,matrix-method}
\alias{agentMatrix,missing-method}
\alias{cbind}
\alias{head}
\alias{head.agentMatrix}
\alias{tail}
\alias{tail.agentMatrix}
\title{agentMatrix class}
\usage{
agentMatrix(..., coords)

\S4method{agentMatrix}{matrix}(..., coords)

\S4method{agentMatrix}{missing}(..., coords)

\S4method{[}{agentMatrix,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,missing,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,numeric,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,missing,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{agentMatrix,missing,numeric,numeric}(x, j) <- value

\S4method{[}{agentMatrix,numeric,missing,numeric}(x, i) <- value

\S4method{[}{agentMatrix,numeric,numeric,character}(x, i, j) <- value

\S4method{[}{agentMatrix,missing,numeric,character}(x, j) <- value

\S4method{[}{agentMatrix,missing,character,character}(x, j) <- value

\S4method{[}{agentMatrix,numeric,character,character}(x, i, j) <- value

\S4method{==}{agentMatrix,character}(e1, e2)

\S4method{[}{agentMatrix,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{$}{agentMatrix}(x, name)

\S4method{show}{agentMatrix}(object)

\S4method{nrow}{agentMatrix}(x)

head.agentMatrix(x, n = 6L, ...)

tail.agentMatrix(x, n = 6L, ...)

cbind(..., deparse.level = 1)

\S4method{cbind}{agentMatrix}(..., deparse.level = 1)

\S4method{length}{agentMatrix}(x)
}
\arguments{
\item{...}{Vectors or a data.frame or a matrix of extra columns to add to the coordinates}

\item{coords}{A matrix with 2 columns representing x and y coordinates}
}
\description{
This is incomplete.

This is a fast alternative to the SpatialPointsDataFrame. It is meant to replace
that functionality, though there are not as many methods (yet). The object is primarily
a numeric matrix. Any character column passed to ... will be converted to a numeric,
using \code{as.factor} internally, and stored as a numeric. Methods using this class
will automatically convert character queries to the correct numeric alternative.
}
\examples{
newAgent <- new("agentMatrix",
      coords=cbind(pxcor=c(1,2,5),pycor=c(3,4,6)),
      char = letters[c(1,2,6)],
      nums2 = c(4.5, 2.6, 2343),
      char2 = LETTERS[c(4,24,3)],
      nums = 5:7)

# compare speeds -- about 5x faster
microbenchmark(times = 499,
  spdf={
     SpatialPointsDataFrame(coords=cbind(pxcor=c(1,2,5),pycor=c(3,4,6)),
     data = data.frame(
       char = letters[c(1,2,6)],
       nums2 = c(4.5, 2.6, 2343),
       char2 = LETTERS[c(4,24,3)],
       nums = 5:7))},
  agentMat = { agentMatrix(
     coords=cbind(pxcor=c(1,2,5),
     pycor=c(3,4,6)),
     char = letters[c(1,2,6)],
     nums2 = c(4.5, 2.6, 2343),
     char2 = LETTERS[c(4,24,3)],
     nums = 5:7)},
  agentMatDirect = { new("agentMatrix",
     coords=cbind(pxcor=c(1,2,5),
     pycor=c(3,4,6)),
     char = letters[c(1,2,6)],
     nums2 = c(4.5, 2.6, 2343),
     char2 = LETTERS[c(4,24,3)],
     nums = 5:7)})

agentMatrix()
newAgent <- agentMatrix(
      coords=cbind(pxcor=c(1,2,5),pycor=c(3,4,6)),
      char = letters[c(1,2,6)],
      nums2 = c(4.5, 2.6, 2343),
      char2 = LETTERS[c(4,24,3)],
      nums = 5:7)

w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1), val = runif(count(patches(w1))))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))


}
\author{
Eliot McIntire

Eliot McIntire
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#clear-turtles}
}

