% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\docType{methods}
\name{bbox}
\alias{bbox}
\title{Extract or set bounding box}
\usage{
.bbox(obj)

\S4method{.bbox}{agentMatrix}(obj)

.bbox(obj) <- value

\S4method{.bbox}{agentMatrix,matrix}(obj) <- value

\S4method{.bbox}{worldNLR}(obj)

\S4method{.bbox}{SpatExtent}(obj)
}
\arguments{
\item{obj}{object deriving from class "agentMatrix",
or for \code{bbox} and \code{extent}, a "worldMatrix" or "worldArray"}

\item{value}{2x2 matrix representing the bounding box. See \code{sp::bbox}.}
}
\description{
These are methods for classes in NetLogoR, i.e., \code{agentMatrix}, \code{worldMatrix},
and \code{worldArray}.
}
\examples{
newAgent <- agentMatrix(
      coords = cbind(pxcor = c(1, 2, 5), pycor = c(3, 4, 6)),
      char = letters[c(1, 2, 6)],
      nums2 = c(4.5, 2.6, 2343),
      char2 = LETTERS[c(4, 24, 3)],
      nums = 5:7)
bbox(newAgent)
extent(newAgent)
coordinates(newAgent)
}
\seealso{
\code{\link[=extent]{extent()}}, \code{\link[=coordinates]{coordinates()}}, \code{sp::bbox}
}
