% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{extent,worldNLR-method}
\alias{extent,worldNLR-method}
\alias{extent,agentMatrix-method}
\alias{.bboxCoords}
\alias{bbox,agentMatrix-method}
\alias{bbox<-}
\alias{bbox<-,agentMatrix,matrix-method}
\alias{bbox,worldNLR-method}
\title{Bounding box and extent methods for NetLogoR classes}
\usage{
\S4method{extent}{worldNLR}(x, ...)

\S4method{extent}{agentMatrix}(x, ...)

.bboxCoords(coords)

\S4method{bbox}{agentMatrix}(obj)

bbox(obj) <- value

\S4method{bbox}{agentMatrix,matrix}(obj) <- value

\S4method{bbox}{worldNLR}(obj)
}
\arguments{
\item{x}{An object from which to get the extent}

\item{...}{Passed to \code{quickPlot::extent}, but is ignored here and there.}

\item{coords}{xy coordinates for all cells, e.g., produced by \code{raster::coordinates}.}

\item{obj}{object deriving from class "agentMatrix" or "worldMatrix" or "worldArray"}

\item{value}{2x2 matrix representing the bounding box. See \code{sp::bbox}.}
}
\value{
\code{bbox} returns a two-column matrix; the first column has the minimum,
the second the maximum values; rows represent the spatial dimensions.
\code{extent} returns an \code{extent} object.

two-column matrix; the first column has the minimum, the second the maximum values;
rows represent the spatial dimensions. See \code{sp::bbox}.
}
\description{
Same as \code{sp::bbox} and \code{raster::extent}.

These are methods for classes in NetLogoR, i.e., \code{agentMatrix}, \code{worldMatrix},
and \code{worldArray}.
}
