% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{bk}
\alias{bk}
\alias{bk,NLworlds,SpatialPointsDataFrame,numeric-method}
\title{Move backward}
\usage{
bk(world, turtles, step, torus = FALSE, out = TRUE)

\S4method{bk}{NLworlds,SpatialPointsDataFrame,numeric}(world, turtles, step,
  torus = FALSE, out = TRUE)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the
turtles move onto.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{step}{Numeric. Distances by which the turtles will move backward.
Must be of length 1 or of length \code{turtles}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}

\item{out}{Logical. Determine if turtles should move at all when
\code{torus = FALSE} and their ending position will be outside of
the world's extent. Default is \code{out = TRUE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated
        coordinates and updated data for their previous coordinates "prevX" and
        "prevY".
}
\description{
Move the turtles backward of the turtles' heading direction.
}
\details{
If the \code{NLworlds} object is wrapped (\code{torus = TRUE}) and the
         distance to move leads a turtle outside of the world's extent, it is
         relocated on the other side of the world, inside the world's extent. If
         \code{torus = FALSE} and \code{out = TRUE}, the turtle moves past the
         world's extent. If \code{torus = FALSE} and \code{out = FALSE}, the
         turtle does not move at all. In the event that a turtle does not move,
         its previous coordinates are still updated with its position before
         running \code{bk()} (i.e., its current position).

         If a given \code{step} value is negative, then the turtle moves
         forward.

         This function is equivalent to the forward function with the distances
         to move equal to \code{step * -1}.

         The turtles' heading are not affected by the function (i.e., the turtles
         do not head backward).
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createOTurtles(world = w1, n = 10)

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1")

t1 <- fd(world = w1, turtles = t1, step = 2)
Plot(t1, addTo ="w1")
t1 <- bk(world = w1, turtles = t1, step = 1)
Plot(t1, addTo ="w1")
t1 <- fd(world = w1, turtles = t1, step = 0.5)
Plot(t1, addTo ="w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#back}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#jump}
}

