% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{canMove}
\alias{canMove}
\alias{canMove,NLworlds,SpatialPointsDataFrame,numeric-method}
\title{Can the turtles move?}
\usage{
canMove(world, turtles, dist)

\S4method{canMove}{NLworlds,SpatialPointsDataFrame,numeric}(world, turtles,
  dist)
}
\arguments{
\item{world}{NLworlds object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{dist}{Numeric. Vector of distances to move. Must
be of length 1 or of length \code{turtles}.}
}
\value{
Logical. Vector of length \code{turtles}.
}
\description{
Report \code{TRUE} if a turtle can move the given distance without leaving
the \code{world}'s extent, report \code{FALSE} otherwise.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
canMove(world = w1, turtles = t1, dist = 1:10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#can-move}
}

