% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{canMove}
\alias{canMove}
\alias{canMove,NLworlds,SpatialPointsDataFrame,numeric-method}
\title{Can the turtle move?}
\usage{
canMove(world, turtles, step)

\S4method{canMove}{NLworlds,SpatialPointsDataFrame,numeric}(world, turtles,
  step)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the
turtles move onto.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtles to
evaluate.}

\item{step}{Numeric. The distances the turtles would move. Must be of
length 1 or of length \code{turtles}.}
}
\value{
Logicals. Vector of length \code{turtles}.
}
\description{
Report \code{TRUE} or \code{FALSE} if the turtles can move the given distances
without leaving the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
canMove(world = w1, turtles = t1, step = 1:10)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#can-move}
}

