% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{cellFromPxcorPycor}
\alias{cellFromPxcorPycor}
\alias{cellFromPxcorPycor,NLworld,numeric,numeric-method}
\alias{cellFromPxcorPycor,NLworldStack,numeric,numeric-method}
\title{Cells numbers from patches coordinates}
\usage{
cellFromPxcorPycor(world, pxcor, pycor)

\S4method{cellFromPxcorPycor}{NLworld,numeric,numeric}(world, pxcor, pycor)

\S4method{cellFromPxcorPycor}{NLworldStack,numeric,numeric}(world, pxcor, pycor)
}
\arguments{
\item{world}{NLworlds object.}

\item{pxcor}{Integer. Vector of patches pxcor coordinates. Must be of length 1
or of the same length as \code{pycor}.}

\item{pycor}{Integer. Vector of patches pycor coordinates. Must be of length 1
or of the same length as \code{pxcor}.}
}
\value{
Numeric. Vector of cells number.
}
\description{
Report the cells numbers as defined for a Raster* object given the patches
coordinates \code{pxcor} and \code{pycor}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
cellFromPxcorPycor(world = w1, pxcor = 0, pycor = 9)


}
\author{
Sarah Bauduin
}

