% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworld-class.R
\docType{methods}
\name{cellFromPxcorPycor}
\alias{cellFromPxcorPycor}
\alias{cellFromPxcorPycor,NLworld,numeric,numeric-method}
\alias{cellFromPxcorPycor,NLworldStack,numeric,numeric-method}
\title{Cells number from patches coordinates}
\usage{
cellFromPxcorPycor(world, pxcor, pycor)

\S4method{cellFromPxcorPycor}{NLworld,numeric,numeric}(world, pxcor, pycor)

\S4method{cellFromPxcorPycor}{NLworldStack,numeric,numeric}(world, pxcor, pycor)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{pxcor}{A vector of \code{pxcor} coordinates.}

\item{pycor}{A vector of \code{pycor} coordinates.}
}
\value{
A vector of cells number.
}
\description{
Report the cells number as defined for a \code{Raster*} object with the patches
coordinates \code{pxcor} and \code{pycor}.
}
\examples{
# Create a world.
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
cellFromPxcorPycor(world = w1, pxcor = 0, pycor = 9)

}
\author{
Sarah Bauduin
}

