% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{clearPatches}
\alias{clearPatches}
\alias{clearPatches,NLworld-method}
\alias{clearPatches,NLworldArray-method}
\alias{clearPatches,NLworldMatrix-method}
\alias{clearPatches,NLworldStack-method}
\title{Clear world's patches}
\usage{
clearPatches(world)

\S4method{clearPatches}{NLworld}(world)

\S4method{clearPatches}{NLworldStack}(world)

\S4method{clearPatches}{NLworldMatrix}(world)

\S4method{clearPatches}{NLworldArray}(world)
}
\arguments{
\item{world}{NLworlds or NLworldMs object.}
}
\value{
NLworld or NLworldMatrix object with \code{NA} values for all patches.
}
\description{
Reset all patches values to \code{NA}.
}
\details{
The name of the layer for the NLWorld is set to \code{""}.
}
\examples{
w1 <- createNLworld()
w1 <- NLset(world = w1, agents = patches(w1), val = runif(NLcount(patches(w1))))
w1Val <- of(world = w1, agents = patches(w1))
summary(w1Val)

w1 <- clearPatches(w1)
w1Val <- of(world = w1, agents = patches(w1))
summary(w1Val)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#clear-patches}
}

