% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{clearPatches}
\alias{clearPatches}
\alias{clearPatches,NLworld-method}
\alias{clearPatches,NLworldStack-method}
\title{Clear patches}
\usage{
clearPatches(world)

\S4method{clearPatches}{NLworld}(world)

\S4method{clearPatches}{NLworldStack}(world)
}
\arguments{
\item{world}{A \code{NLworld} or \code{NLworldStack} object.}
}
\value{
An empty \code{NLworld} object.
}
\description{
Reset all patches value to \code{NA}.
}
\details{
Similar as setValues(world, NA).
         The name of the layer is reset to "".
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
w1[] <- runif(n = 1089)
w1_val <- values(w1)
summary(w1_val)

w1 <- clearPatches(world = w1)
w1_val <- values(w1)
summary(w1_val)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

