% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{clearTurtles}
\alias{clearTurtles}
\alias{clearTurtles,SpatialPointsDataFrame-method}
\title{Clear turtles}
\usage{
clearTurtles(turtles)

\S4method{clearTurtles}{SpatialPointsDataFrame}(turtles)
}
\arguments{
\item{turtles}{A \code{SpatialPointsDataFrame} object representing the turtles.}
}
\description{
Kill all \code{turtles}.
}
\details{
Remove the \code{turtles} object from the R environment.
}
\examples{
# Create one turtle.
t1 <- SpatialPointsDataFrame(coords = matrix(c(1,2), nrow = 1, ncol = 2), data = data.frame(NA))
t1
clearTurtles(turtles = t1)
t1 # does not work

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

