% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{convertNLworld}
\alias{convertNLworld}
\alias{convertNLworld,RasterLayer-method}
\alias{convertNLworld,RasterStack-method}
\title{Convert a raster to a \code{NLworlds} object}
\usage{
convertNLworld(raster)

\S4method{convertNLworld}{RasterLayer}(raster)

\S4method{convertNLworld}{RasterStack}(raster)
}
\arguments{
\item{raster}{\code{RasterLayer} or a \code{RasterStack} object.}
}
\value{
\code{NLworld} or a \code{NLworldStack} object depending on the input.
        Patches value are retained from the \code{Raster*} object.
}
\description{
Convert a \code{RasterLayer} object into a \code{NLworld} object or a \code{RasterStack}
into a {NLworldStack} object.
}
\details{
See \code{help("NLworlds-class")} for more details on the \code{NLworlds}
         classes.

         The \code{Raster*} is resampled to match the coordinates system and
         resolution of a \code{NLworlds} using the nearest neighbor method. The
         extent will be bigger by 1 on the width and on the height.
}
\examples{
r <- raster(nrows = 21, ncols = 21, xmn = 0, ymn = 0, res = 5)
r[] <- runif(648)
plot(r)
extent(r)
world <- convertNLworld(raster = r)
plot(world)
extent(world)
minPxcor(world)
maxPxcor(world)
minPycor(world)
maxPycor(world)


}
\author{
Sarah Bauduin
}

