% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{convertNLworld}
\alias{convertNLworld}
\alias{convertNLworld,RasterLayer-method}
\alias{convertNLworld,RasterStack-method}
\title{Convert to a \code{NLworld}}
\usage{
convertNLworld(raster)

\S4method{convertNLworld}{RasterLayer}(raster)

\S4method{convertNLworld}{RasterStack}(raster)
}
\arguments{
\item{raster}{A \code{RasterLayer} or a \code{RasterStack} object.}
}
\value{
A \code{NLworld} or \code{NLworldStack} object depending on the input.
        Patches value are retained from the \code{Raster*} object.
}
\description{
Convert a \code{Raster*} object into a \code{NLworld}.
}
\details{
See \code{help("NLworld")} for more details on the \code{NLworld}.
}
\examples{
r <- raster(system.file("external/test.grd", package="raster")) # from the raster package
plot(r)
maxPxcor(r) # does not work
world <- convertNLworld(raster = r)
plot(world)
maxPxcor(world) # works

}
\author{
Sarah Bauduin
}

