% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{coordinates,agentDataTable-method}
\alias{coordinates,agentDataTable-method}
\title{Extract coordinates from agentDataTable}
\usage{
\S4method{coordinates}{agentDataTable}(obj, ...)
}
\description{
This is incomplete
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(world = w1, n = 10, agent = TRUE)
coordinates(t1)

library(microbenchmark)
N = 1e4
coords <-  cbind(xcor=runif(N, xmin(w1), xmax(w1)),
                 ycor = runif(N, ymin(w1), ymax(w1)))
turtlesDT <- createTurtles(coords = coords, n = N, agent = TRUE)
turtlesDF <- createTurtles(coords = coords, n = N, agent = FALSE)
microbenchmark(coordinates(turtlesDT), coordinates(turtlesDF))
microbenchmark(coordinates(turtlesDT) <- coords)
}

