% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{coordinates,agentMatrix-method}
\alias{coordinates,agentMatrix-method}
\title{Set spatial coordinates}
\usage{
\S4method{coordinates}{agentMatrix}(obj, ...)
}
\arguments{
\item{obj}{documentation needed}

\item{...}{additional arguments that may be used by particular methods

 description needed}
}
\description{
Set spatial coordinates
}
