% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{coordinates,agentMatrix-method}
\alias{coordinates,agentMatrix-method}
\title{Spatial accessors and setters for NetLogoR classes}
\usage{
\S4method{coordinates}{agentMatrix}(obj, ...)
}
\arguments{
\item{obj}{object deriving from class "agentMatrix"}

\item{...}{additional arguments that may be used by particular methods}
}
\value{
\code{coordinates} returns a matrix of coordinates of the \code{obj}.
}
\description{
Spatial accessors and setters for NetLogoR classes
}
\seealso{
\code{\link[=bbox]{bbox()}}, \code{\link[=extent]{extent()}}
}
