% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\name{coordinates,agentMatrix-method}
\alias{coordinates,agentMatrix-method}
\title{Set spatial coordinates}
\usage{
\S4method{coordinates}{agentMatrix}(obj, ...)
}
\arguments{
\item{obj}{an \code{AgentMatrix} object}

\item{...}{additional arguments that may be used by particular methods}
}
\value{
usually an object of class \code{SpatialPointsDataFrame}; if the coordinates set cover
the full set of variables in object, an object of class \code{SpatialPoints} is returned.
See \code{\link[sp:coordinates]{sp::coordinates()}}.
}
\description{
Set spatial coordinates
}
