% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{createNLworld}
\alias{createNLworld}
\alias{createNLworld,missing,missing,missing,missing-method}
\alias{createNLworld,numeric,numeric,numeric,numeric-method}
\title{Create a NLworld}
\usage{
createNLworld(minPxcor = -16, maxPxcor = 16, minPycor = -16,
  maxPycor = 16)

\S4method{createNLworld}{numeric,numeric,numeric,numeric}(minPxcor = -16,
  maxPxcor = 16, minPycor = -16, maxPycor = 16)

\S4method{createNLworld}{missing,missing,missing,missing}()
}
\arguments{
\item{minPxcor}{\code{pxcor} for patches on the left border of the NLworld.
Default value = \code{-16}, as in NetLogo.}

\item{maxPxcor}{\code{pxcor} for patches on the right border of the NLworld.
Default value = \code{16}, as in NetLogo.}

\item{minPycor}{\code{pycor} for patches at the bottom of the NLworld.
Default value \code{-16}, as in NetLogo.}

\item{maxPycor}{\code{pycor} for patches at the top of the NLworld.
Default value \code{16}, as in NetLogo.}
}
\value{
A NLworld object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patches value are NA.
}
\description{
Create an empty grid of patches of class \code{NLworld}.
}
\details{
See \code{help("NLworld")} for more details on the NLworld.
}
\examples{
# Create a square world of 25 patches.
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
# Give the patches random values between 0 and 10.
w1[] <- runif(n = 25, min = 0, max = 10)
plot(w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

