% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{createNLworld}
\alias{createNLworld}
\alias{createNLworld,missing,missing,missing,missing-method}
\alias{createNLworld,numeric,numeric,numeric,numeric-method}
\title{Create a world}
\usage{
createNLworld(minPxcor = -16, maxPxcor = 16, minPycor = -16,
  maxPycor = 16)

\S4method{createNLworld}{numeric,numeric,numeric,numeric}(minPxcor = -16,
  maxPxcor = 16, minPycor = -16, maxPycor = 16)

\S4method{createNLworld}{missing,missing,missing,missing}()
}
\arguments{
\item{minPxcor}{Integer. Minimum pxcor for the patches (world's left border).}

\item{maxPxcor}{Integer. Maximum pxcor for the patches (world's right border).}

\item{minPycor}{Integer. Minimum pycor for the patches (world's bottom border).}

\item{maxPycor}{Integer. Maximum pycor for the patches (world's top border).}
}
\value{
NLworld object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patches value are \code{NA}.
}
\description{
Create an empty grid of patches of class NLworld.
}
\details{
If no parameters value are provided, default values are: \code{minPxcor = -16},
         \code{maxPxcor = 16}, \code{minPycor = -16}, and \code{maxPycor = 16}.

         See \code{help("NLworld-class")} for more details on the NLworld class.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
plot(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

