% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{createOTurtles}
\alias{createOTurtles}
\alias{createOTurtles,numeric,NLworlds-method}
\title{Create ordered turtles}
\usage{
createOTurtles(n, world, breed, color)

\S4method{createOTurtles}{numeric,NLworlds}(n, world, breed, color)
}
\arguments{
\item{n}{Integer. The number of new turtles to create.}

\item{world}{\code{NLworlds} object.}

\item{breed}{Characters. Either of length 1 or \code{n}. If missing,
\code{breed = "turtle"} for all turtles.}

\item{color}{Characters. Vector of length \code{n}. If missing, colors are
assigned using the function \code{rainbow(n)}.}
}
\value{
SpatialPointsDataFrame of length \code{n} with the columns for the
        dataframe being: "who", "heading", "prevX", "prevY", "breed", and "color".
}
\description{
Create \code{n} new turtles at the center of the world with their headings evenly
distributed.
}
\details{
The first turtle \code{who = 0} has its heading set to 0.

         The identity of the turtles is defined by their "who" number. This
         numbering starts at 0 and increments by 1.

         The coordinates from the previous time step are stored in "prevX" and
         "prevY". The initial values are \code{NA}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createOTurtles(n = 10, world = w1)

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1") # automatically uses color column in SpatialPointsDataFrame

t1 <- fd(world = w1, turtles = t1, dist = 1)
Plot(t1, addTo ="w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#create-ordered-turtles}
}

