% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\name{createTurtles}
\alias{createTurtles}
\alias{createTurtles,numeric,matrix,missing-method}
\alias{createTurtles,numeric,missing,ANY-method}
\title{Create \code{turtles}}
\usage{
createTurtles(n, coords, world, heading, breed, color)

\S4method{createTurtles}{numeric,matrix,missing}(n, coords, world, heading, breed, color)

\S4method{createTurtles}{numeric,missing,ANY}(n, coords, world, heading, breed, color)
}
\arguments{
\item{n}{Integer.}

\item{coords}{Matrix (\code{ncol} = 2) with the first column \code{xcor} and the second
column \code{ycor} representing the \code{turtles} initial locations.
\code{nrow(coords)} must be equal to 1 or to \code{n}.
Given coordinates must be inside the \code{world}'s extent. If missing,
\code{turtles} are put in the center of the \code{world}.}

\item{world}{\code{WorldMatrix} or \code{worldArray} object.}

\item{heading}{Numeric. Vector of values between 0 and 360. Must be of length 1 or
of length \code{n}. If missing, a random \code{heading} is assigned to
each \code{turtle}.}

\item{breed}{Character. Vector of \code{breed} names. Must be of length 1 or of length
\code{n}. If missing, \code{breed = "turtle"} for all \code{turtles}.}

\item{color}{Character. Vector of \code{color} names. Must be of length \code{n}.
If missing, \code{colors} are assigned using the function \code{rainbow(n)}.}
}
\value{
\code{AgentMatrix} object of length \code{n} with data for the
\code{turtles} being: \code{xcor}, \code{ycor}, \code{who}, \code{heading}, \code{prevX}, \code{prevY},
\code{breed}, and \code{color}.
}
\description{
Create \code{n} moving \code{agents} with a set of defined variables.
}
\details{
If \code{coords} is provided, \code{world} must not be provided.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     The identity of the `turtles` is defined by their `who` number. This
     numbering starts at 0 and increments by 1.

     The coordinates from the previous time step are stored in `prevX` and
     `prevY`. The initial values are `NA`.
}\if{html}{\out{</div>}}
}
\examples{
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4,
                          data = runif(25))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
plot(w1)
points(t1, col = of(agents = t1, var = "color"), pch = 16)


}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#create-turtles}
}
\author{
Sarah Bauduin
}
