% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{die}
\alias{die}
\alias{die,SpatialPointsDataFrame,numeric-method}
\title{Die}
\usage{
die(turtles, who)

\S4method{die}{SpatialPointsDataFrame,numeric}(turtles, who)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()}.}

\item{who}{Numeric. The who number(s) of the turtle(s) that die(s).}
}
\value{
A SpatialPointsDataFrame of length equal to \code{length(turtles) - length(who)}.
}
\description{
Kill the turtle(s).
}
\details{
The who numbers of the other turtles are maintained.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
length(t1)
t1 <- die(turtles = t1, who = c(2, 3, 4))
length(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

