% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{diffuse}
\alias{diffuse}
\alias{diffuse,NLworld,missing,numeric,numeric-method}
\alias{diffuse,NLworldStack,character,numeric,numeric-method}
\title{Diffuse}
\usage{
diffuse(world, pVar, share, nNeighbors)

\S4method{diffuse}{NLworld,missing,numeric,numeric}(world, share, nNeighbors)

\S4method{diffuse}{NLworldStack,character,numeric,numeric}(world, pVar, share,
  nNeighbors)
}
\arguments{
\item{world}{A \code{NLworld} or \code{NLworldStack} object.}

\item{pVar}{If the world is a \code{NLworldStack}, pVar is the name
(characters) of the layer used for the diffusion.}

\item{share}{A value between 0 and 1 representing the portion of the patch
value to be diffused among its neighbors.}

\item{nNeighbors}{4 or 8. Represent the number of neihgbor patches involved in
the diffusion process.}
}
\value{
A \code{NLworld} or \code{NLworldStack} object with updated values.
}
\description{
Each patch gives an equal share to its neighbor patches of a portion of its value.
}
\details{
What is given is lost for the patches.
         Patches on the sides of the world have less than 4 or 8 neighbors.
         Each neighbor still gets 1/4 or 1/8 of the shared amount and the diffusing
         patch keeps the leftover.
}
\examples{
# Create a world of 25 patches.
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
# Give the patches random values between 0 and 10.
w1[] <- runif(n = 25, min = 0, max = 10)
plot(w1)
# Diffuse 50\% of each patch value to its 8 neighbors
w2 <- diffuse(world = w1, share = 0.5, nNeighbors = 8)
plot(w2)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

