% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{diffuse}
\alias{diffuse}
\alias{diffuse,NLworld,missing,numeric,numeric-method}
\alias{diffuse,NLworldStack,character,numeric,numeric-method}
\title{Diffuse values in a world}
\usage{
diffuse(world, pVar, share, nNeighbors)

\S4method{diffuse}{NLworld,missing,numeric,numeric}(world, share, nNeighbors)

\S4method{diffuse}{NLworldStack,character,numeric,numeric}(world, pVar, share,
  nNeighbors)
}
\arguments{
\item{world}{NLworlds object.}

\item{pVar}{Character. If the \code{world} is a NLworldStack object, \code{pVar}
is the name of the layer to use to define the patches values.
\code{pVar} must not be provided if the \code{world} is a NLworld object.}

\item{share}{Numeric. Value between 0 and 1 representing the portion of
the patches values to be diffused among the neighbors.}

\item{nNeighbors}{Integer: 4 or 8. Represents the number of neighbor patches
considered.}
}
\value{
NLworlds object with patches values updated.
}
\description{
Each patch gives an equal share of a portion of its value to its neighbor patches.
}
\details{
What is given is lost for the patches.

         Patches on the sides of the \code{world} have less than 4 or 8 neighbors.
         Each neighbor still gets 1/4 or 1/8 of the shared amount and the diffusing
         patch keeps the leftover.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
plot(w1)
# Diffuse 50\% of each patch value to its 8 neighbors
w2 <- diffuse(world = w1, share = 0.5, nNeighbors = 8)
plot(w2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#diffuse}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#diffuse4}
}

