% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{diffuse}
\alias{diffuse}
\alias{diffuse,NLworld,missing,numeric,numeric-method}
\alias{diffuse,NLworldArray,character,numeric,numeric-method}
\alias{diffuse,NLworldMatrix,missing,numeric,numeric-method}
\alias{diffuse,NLworldStack,character,numeric,numeric-method}
\title{Diffuse values in a world}
\usage{
diffuse(world, pVar, share, nNeighbors, torus = FALSE)

\S4method{diffuse}{NLworld,missing,numeric,numeric}(world, share, nNeighbors,
  torus)

\S4method{diffuse}{NLworldStack,character,numeric,numeric}(world, pVar, share,
  nNeighbors, torus = FALSE)

\S4method{diffuse}{NLworldMatrix,missing,numeric,numeric}(world, share,
  nNeighbors, torus)

\S4method{diffuse}{NLworldArray,character,numeric,numeric}(world, pVar, share,
  nNeighbors, torus = FALSE)
}
\arguments{
\item{world}{NLworlds or NLworldMs object.}

\item{pVar}{Character. If the \code{world} is a NLworldStack object, \code{pVar}
is the name of the layer to use to define the patches values.
\code{pVar} must not be provided if the \code{world} is a NLworld object.}

\item{share}{Numeric. Value between 0 and 1 representing the portion of
the patches values to be diffused among the neighbors.}

\item{nNeighbors}{Integer: 4 or 8. Represents the number of neighbor patches
considered.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
NLworlds or NLworldMs object with patches values updated.
}
\description{
Each patch gives an equal share of a portion of its value to its neighbor patches.
}
\details{
What is given is lost for the patches.

         If \code{torus = TRUE}, all patches have \code{nNeighbors} patches around them, which
         some may be on the other sides of the \code{world}. If \code{torus = FALSE},
         patches located on the edges of the \code{world} have less than \code{nNeighbors} patches around them.
         However, each neighbor still gets 1/4 or 1/8 of the shared amount and the diffusing
         patch keeps the leftover.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1), val = runif(count(patches(w1))))
plot(w1)
# Diffuse 50\% of each patch value to its 8 neighbors
w2 <- diffuse(world = w1, share = 0.5, nNeighbors = 8)
plot(w2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#diffuse}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#diffuse4}
}

