% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\name{dx}
\alias{dx}
\alias{dx,agentMatrix,numeric-method}
\alias{dx,agentMatrix,missing-method}
\title{x-increment}
\usage{
dx(turtles, dist = 1)

\S4method{dx}{agentMatrix,numeric}(turtles, dist = 1)

\S4method{dx}{agentMatrix,missing}(turtles)
}
\arguments{
\item{turtles}{\code{AgentMatrix} object representing the moving \code{agents}.}

\item{dist}{Numeric. Vector of distances the \code{turtles} would have to
move forward to
compute the increment values. Must be of length 1 or of length
\code{turtles}. The default value is \code{dist = 1}.}
}
\value{
Numeric. Vector of length \code{turtles}.
}
\description{
Report the amount by which the \code{turtles}' coordinates \code{xcor} would change
if the \code{turtles} were
to move forward the given distances with their current \code{headings}.
}
\details{
Report the sine of the \code{turtles}' \code{heading} multiplied by the \code{dist}
values. Heading 0 is north and angles are calculated in degrees in a
clockwise manner.
}
\examples{
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createOTurtles(world = w1, n = 10)
dx(turtles = t1)

}
\references{
Wilensky, U. 1999. NetLogo. https://www.netlogo.org.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://docs.netlogo.org/dictionary.html#dxy}
}
\author{
Sarah Bauduin
}
