% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentMatrix-Class-methods.R
\docType{methods}
\name{extent,worldNLR-method}
\alias{extent,worldNLR-method}
\alias{extent,agentMatrix-method}
\title{Bounding box and extent methods for NetLogoR classes}
\usage{
\S4method{extent}{worldNLR}(x, ...)

\S4method{extent}{agentMatrix}(x, ...)
}
\arguments{
\item{x}{object deriving from class "agentMatrix",
or a \code{worldMatrix} or \code{worldArray}}

\item{...}{Ignored.}
}
\value{
\code{bbox} returns a two-column matrix; the first column has the minimum,
the second the maximum values; rows represent the spatial dimensions.
\code{extent} returns an \code{SpatExtent} object from the package \code{terra}.
}
\description{
Same as \code{terra::ext()}, but for \code{NetLogoR} classes.
}
\seealso{
\code{\link[terra:ext]{terra::ext()}}
}
