% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-classes.R, R/Classes.R
\docType{methods}
\name{[}
\alias{$,agentMatrix-method}
\alias{[}
\alias{[,NLworldArray,missing,missing,ANY-method}
\alias{[,NLworldArray,numeric,numeric,ANY-method}
\alias{[,NLworldMatrix,missing,missing,ANY-method}
\alias{[,NLworldMatrix,numeric,numeric,ANY-method}
\alias{[,agentMatrix,logical,missing,ANY-method}
\alias{[,agentMatrix,missing,character,ANY-method}
\alias{[,agentMatrix,missing,missing,missing-method}
\alias{[,agentMatrix,missing,numeric,ANY-method}
\alias{[,agentMatrix,numeric,character,ANY-method}
\alias{[,agentMatrix,numeric,missing,ANY-method}
\alias{[,agentMatrix,numeric,numeric,ANY-method}
\alias{[<-}
\alias{[<-,NLworldArray,missing,missing,matrix-method}
\alias{[<-,NLworldArray,numeric,numeric,matrix-method}
\alias{[<-,NLworldMatrix,missing,missing,ANY-method}
\alias{[<-,NLworldMatrix,numeric,numeric,ANY-method}
\alias{[<-,agentMatrix,missing,character,character-method}
\alias{[<-,agentMatrix,missing,numeric,character-method}
\alias{[<-,agentMatrix,missing,numeric,numeric-method}
\alias{[<-,agentMatrix,numeric,character,character-method}
\alias{[<-,agentMatrix,numeric,character,data.frame-method}
\alias{[<-,agentMatrix,numeric,missing,numeric-method}
\alias{[<-,agentMatrix,numeric,numeric,character-method}
\alias{[<-,agentMatrix,numeric,numeric,numeric-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{NLworldMatrix,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NLworldMatrix,missing,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NLworldMatrix,numeric,numeric,ANY}(x, i, j) <- value

\S4method{[}{NLworldMatrix,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{NLworldArray,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NLworldArray,missing,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NLworldArray,numeric,numeric,matrix}(x, i, j) <- value

\S4method{[}{NLworldArray,missing,missing,matrix}(x, i, j) <- value

\S4method{[}{agentMatrix,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,logical,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,missing,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,missing,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,numeric,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,missing,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{agentMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{agentMatrix,missing,numeric,numeric}(x, i, j) <- value

\S4method{[}{agentMatrix,numeric,missing,numeric}(x, i, j) <- value

\S4method{[}{agentMatrix,numeric,character,data.frame}(x, i, j) <- value

\S4method{[}{agentMatrix,numeric,numeric,character}(x, i, j) <- value

\S4method{[}{agentMatrix,missing,numeric,character}(x, i, j) <- value

\S4method{[}{agentMatrix,missing,character,character}(x, i, j) <- value

\S4method{[}{agentMatrix,numeric,character,character}(x, i, j) <- value

\S4method{$}{agentMatrix}(x, name)
}
\arguments{
\item{x}{A \code{NLworld} object from which to extract element(s) or
in which to replace element(s).}

\item{i}{Indices specifying elements to extract or replace.}

\item{j}{see \code{i}.}

\item{...}{other named arguments}

\item{drop}{not implemented}

\item{value}{Any R object}

\item{name}{documentation needed}
}
\description{
Operators acting on vectors, matrices, arrays and lists to extract or replace parts.
}
\note{
Extract methods for agentMatrix class will generally maintain the \code{agentMatrix} class.
This means that there will still be coordinates, character columns represented as numerics etc.
\code{$} is for extracting the raw columns and does not maintain the \code{agentMatrix} class.
}

