% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{face}
\alias{face}
\alias{face,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{face,NLworlds,SpatialPointsDataFrame,matrix-method}
\title{Face}
\usage{
face(world, turtles, to, torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,matrix}(world, turtles, to,
  torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  turtles, to, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtle(s) from which
the heading will be modified.}

\item{to}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) towards which the turtles' heading will be set
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) towards which
the turtles' heading will be set
A matrix (ncol = 2) with the first column \code{x} and the second
column \code{y} representing the coordinates of the location(s)
towards which the heading will be set
\code{to} must be of length 1 or of the same length as \code{from}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A SpatialPointsDataFrame representing the turtles with updated headings.
}
\description{
Set the turtles' heading towards some agent(s) or location(s) [x,y].
}
\details{
If \code{torus = TRUE} and the distance from one turtle \code{from} to
         its corresponding agent or location \code{to} is smaller around the
         sides of the world than across it, then the heading to the agent or location
         going around the sides of the world is given to the turtle.
         There is no change in heading for a turtle towards itself or its own location.
         This function is similar to setting the agents' heading to the results of
         the \code{towards()} function.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(25)
t1 <- createTurtles(n = 10, coords = cbind(pxcor = randomXcor(world = w1, n = 10),
                                           pycor = randomYcor(world = w1, n = 10)))
plot(w1)
points(t1, pch = 16, col = t1@data$color)
t1 <- face(world = w1, turtles = t1, to = cbind(x = 0, y = 0))
t1 <- fd(world = w1, turtles = t1, step = 0.5)
points(t1, pch = 16, col = t1@data$color)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

