% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{face}
\alias{face}
\alias{face,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{face,NLworlds,SpatialPointsDataFrame,matrix-method}
\title{Face directions}
\usage{
face(world, turtles, to, torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,matrix}(world, turtles, to,
  torus = FALSE)

\S4method{face}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  turtles, to, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtles to modify
their heading.}

\item{to}{Matrix (ncol = 2) with the first column \code{pxcor} and the
               second column \code{pycor} representing the coordinates of the
               patches towards which the turtles' heading is set

               SpatialPointsDataFrame created by \code{createTurtles()} or by
               \code{createOTurtles()} representing the turtles towards which
               the turtles' heading is set.

               Matrix (ncol = 2) with the first column \code{x} and the second
               column \code{y} representing the coordinates of the locations
               towards which the heading is set.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated headings.
}
\description{
Set the turtles' heading towards some agents or locations.
}
\details{
\code{to} must be of length 1 or of length \code{turtles}.

         If \code{torus = TRUE} and the distance from one turtle to
         its corresponding agent/location \code{to} is smaller around the
         sides of the world than across it, then the heading to the agent/location
         going around the sides of the world is given to the turtle.

         There is no change in the heading when the turtle faces itself or its own location.

         This function is similar to setting the agents' heading to the results of
         \code{towards()}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearplot()
Plot(w1)
Plot(t1, addTo = "w1")

t1 <- face(world = w1, turtles = t1, to = cbind(x = 0, y = 0))
t1 <- fd(world = w1, turtles = t1, dist = 0.5)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#face}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#facexy}
}

