% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{fastCbind}
\alias{fastCbind}
\title{Faster cbind for 2 matrices}
\usage{
fastCbind(a, b)
}
\arguments{
\item{a}{matrix}

\item{b}{matrix}
}
\description{
This is a drop-in replacement for cbind IF it is 2 matrices.
It is at least 2x faster for small matrices. It becomes
slower for larger matrices so should only be used for small
matrices (smaller than 10,000 x 9)
}
\author{
Eliot McIntire
}

