% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{hatch}
\alias{hatch}
\alias{hatch,SpatialPointsDataFrame,numeric,numeric,character-method}
\alias{hatch,SpatialPointsDataFrame,numeric,numeric,missing-method}
\title{Hatch}
\usage{
hatch(turtles, who, n, breed)

\S4method{hatch}{SpatialPointsDataFrame,numeric,numeric,character}(turtles, who,
  n, breed)

\S4method{hatch}{SpatialPointsDataFrame,numeric,numeric,missing}(turtles, who,
  n)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} containing the parent turtle.}

\item{who}{Numeric. The who number of the parent turtle.}

\item{n}{Numeric. The number of new turtle(s) to create.}

\item{breed}{Character. The breed for the turtle(s) to be created. If missing,
the created turtle(s) is/are of the same breed as the parent turtle.}
}
\value{
A SpatialPointsDataFrame of length equal to \code{length(turtles) + n}.
}
\description{
Create new turtle(s) from a parent turtle.
}
\details{
The created turtle(s) inherit(s) of all the data from the parent turtle,
         except for the breed if specified otherwise, and for the who number.
         The who numbers of the turtles created take on following the highest
         who number among the turtles.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(world = w1, n = 10)
length(t1)
t1 <- hatch(turtles = t1, who = 0, n = 2)
length(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

