% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{hatch}
\alias{hatch}
\alias{hatch,SpatialPointsDataFrame,numeric,numeric,character-method}
\alias{hatch,SpatialPointsDataFrame,numeric,numeric,missing-method}
\title{Hatch new turtles}
\usage{
hatch(turtles, who, n, breed)

\S4method{hatch}{SpatialPointsDataFrame,numeric,numeric,character}(turtles, who,
  n, breed)

\S4method{hatch}{SpatialPointsDataFrame,numeric,numeric,missing}(turtles, who,
  n)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{who}{Integer. The "who" number of the parent turtle.}

\item{n}{Integer.}

\item{breed}{Character. Vector of "breed" names. Must be of length 1 or of length
\code{n}. If missing,
the created turtles are of the same "breed" as the parent turtle.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with the new
        hatched ones.
}
\description{
Create \code{n} new turtles from a parent turtle.
}
\details{
The parent turtle must be contained in the \code{turtles}.

         The created turtles inherit of all the data from the parent turtle,
         except for the "breed", if specified otherwise, and for the "who" numbers.
         The "who" numbers of the turtles created take on following the highest
         "who" number among the \code{turtles}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
length(t1)
t1 <- hatch(turtles = t1, who = 0, n = 2)
length(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#hatch}
}

