% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{home}
\alias{home}
\alias{home,NLworlds,SpatialPointsDataFrame,character-method}
\title{Home}
\usage{
home(world, turtles, home)

\S4method{home}{NLworlds,SpatialPointsDataFrame,character}(world, turtles, home)
}
\arguments{
\item{world}{A \code{NLworld*} object, representing the world in which the
turtles move onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the agents to move.}

\item{home}{Character. Can take one of the following options to define where
is home for the turtles.
\code{home = "home0"} will place the turtles at the location
\code{xcor = 0, ycor = 0}.
\code{home = "center"} will place the turtles at the center of
the world.
\code{home = "pCorner"} will place the turtles at the center of
the patch located in the left bottom corner of the world.
\code{home = "corner"} will place the turtles at the left bottom
corner of the world.}
}
\value{
A SpatialPointsDataFrame with updated coordinates.
}
\description{
Move the turtles back "home".
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(25)
t1 <- createTurtles(world = w1, n = 10, coords = cbind(xcor = runif(10, 0, 4), ycor = runif(10, 0, 4)))
plot(w1)
points(t1, pch = 16, col = t1@data$color)
t1 <- home(world = w1, turtles = t1, home = "pCorner")
points(t1, pch = 16, col = t1@data$color)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

