% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{layoutCircle}
\alias{layoutCircle}
\alias{layoutCircle,NLworld,SpatialPointsDataFrame,numeric-method}
\title{Layout turtles on a circle}
\usage{
layoutCircle(world, turtles, radius, torus = FALSE)

\S4method{layoutCircle}{NLworld,SpatialPointsDataFrame,numeric}(world, turtles,
  radius, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the
turtles move onto.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles.}

\item{radius}{Numeric. Radius of the cercle.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated coordinates
        and updated data for their previous coordinates "prevX" and "prevY".
}
\description{
Arrange the turtles in a circle centered on the world with the given radius.
}
\details{
The turtles point outwards.

         If the \code{NLworlds} object is wrapped (\code{torus = TRUE}) and the
         \code{radius} values lead a turtle outside of the world's extent, it is
         relocated on the other side of the world, inside the world's extent. If
         \code{torus = FALSE} and \code{out = TRUE}, the turtle are located past
         the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(100)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")

t2 <- layoutCircle(world = w1, turtles = t1, radius = 3)
Plot(t2, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#layout-circle}
}

