% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{left}
\alias{left}
\alias{left,SpatialPointsDataFrame,numeric-method}
\title{Rotate to the left}
\usage{
left(turtles, nDegrees)

\S4method{left}{SpatialPointsDataFrame,numeric}(turtles, nDegrees)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtles to rotate.}

\item{nDegrees}{Numeric. The number of degrees by which to rotate the turtles'
heading to the left. Must be of length 1 or of length \code{turtles}.}
}
\value{
SpatialPointsDataFrame representing the turtles with updated headings.
}
\description{
Rotate the turtles's heading to the left.
}
\details{
If \code{nDegrees} is negative, the turtle rotate to the right.
}
\examples{
t1 <- createTurtles(n = 10, world = w1)
of(t1, tVarName = "heading")
t1 <- left(turtles = t1, nDegrees = 180)
of(t1, tVarName = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#left}
}

