% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{maxOneOf}
\alias{maxOneOf}
\alias{maxOneOf,SpatialPointsDataFrame,missing,character-method}
\alias{maxOneOf,matrix,NLworld,missing-method}
\alias{maxOneOf,matrix,NLworldStack,character-method}
\title{One agent with maximum}
\usage{
maxOneOf(agents, world, varName)

\S4method{maxOneOf}{matrix,NLworld,missing}(agents, world)

\S4method{maxOneOf}{matrix,NLworldStack,character}(agents, world, varName)

\S4method{maxOneOf}{SpatialPointsDataFrame,missing,character}(agents, varName)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column \code{pxcor} and the
               second column \code{pycor} representing the coordinates for the
               patches to evaluate.

               SpatialPointsDataFrame created by \code{createTurtles()} or by
               \code{createOTurtles()} representing the turtles to evaluate.}

\item{world}{\code{NLworlds} object. Only needed if \code{agents} are patches.
Must not be provided if \code{agents} are turtles.}

\item{varName}{Characters. The name of the variable to evaluate for the maximum value.
If \code{agents} are patches and the \code{world} is a \code{NLworld}
object, \code{varName} must not be provided. If \code{agents} are
patches and the \code{world} is a \code{NLworldStack} object,
\code{varName} refers to the layer used for evaluating patches
values. If \code{agents} are turtles, \code{varName} is one of
the turtles' variable. \code{varName} can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
Matrix (ncol = 2, nrow = 1) with the first column \code{pxcor} and
        the second column \code{pycor} representing the coordinates for the patch
        (or one of the patches) with the maximum value among the \code{agents}.

        SpatialPointsDataFrame of length 1 representing the turtle (or one of
        the turtles) which has its \code{varName} equals to the maximum value
        among the \code{agents}.
}
\description{
Report one patch coordinates \code{pxcor} and \code{pycor} which has its
variable equals to the maximum value.
Report one turtle which has its variable equals to the maximum value.
}
\details{
If there are several \code{agents}. with the maximum value, one is chosen
         randomly. To access to all agents with their variable equal to the maximum
         value, use \code{withMax()}.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
plot(w1)
pMax <- maxOneOf(agents = patches(w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10), heading = sample(1:3, size = 10, replace= TRUE))
t2 <- maxOneOf(agents = t1, varName = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-one-of}
}

