% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{maxOneOf}
\alias{maxOneOf}
\alias{maxOneOf,SpatialPointsDataFrame,missing,character-method}
\alias{maxOneOf,agentMatrix,missing,character-method}
\alias{maxOneOf,matrix,NLworld,missing-method}
\alias{maxOneOf,matrix,NLworldArray,character-method}
\alias{maxOneOf,matrix,NLworldMatrix,missing-method}
\alias{maxOneOf,matrix,NLworldStack,character-method}
\title{One agent with maximum}
\usage{
maxOneOf(agents, world, var)

\S4method{maxOneOf}{matrix,NLworld,missing}(agents, world)

\S4method{maxOneOf}{matrix,NLworldStack,character}(agents, world, var)

\S4method{maxOneOf}{SpatialPointsDataFrame,missing,character}(agents, var)

\S4method{maxOneOf}{matrix,NLworldMatrix,missing}(agents, world)

\S4method{maxOneOf}{matrix,NLworldArray,character}(agents, world, var)

\S4method{maxOneOf}{agentMatrix,missing,character}(agents, var)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{world}{NLworlds object.}

\item{var}{Character. The name of the selected \code{agents} variable.
If \code{agents} are patches and the \code{world} is a
\code{NLworld} object, \code{var} must not be provided. If
\code{agents} are patches and the \code{world} is a \code{NLworldStack}
object, \code{var} is the name of the layer to use to define the patches
values. If \code{agents} are turtles, \code{var} is one of
the turtles' variable and can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
Matrix (ncol = 2, nrow = 1) with the first column "pxcor" and
        the second column "pycor" representing the coordinates of the patch
        (or of one of the patches) among the \code{agents} which has its variable
        equals to the maximum value
        among the \code{agents}, or

        SpatialPointsDataFrame of length 1 representing the turtle (or one of
        the turtles) among the \code{agents} which has its variable \code{var}
        equals to the maximum value
        among the \code{agents}.
}
\description{
Report one patch or one turtle among \code{agents} which has its variable equals
to the maximum value.
}
\details{
\code{world} must not be provided if \code{agents} are turtles.

         If there are several patches or turtles among \code{agents} with their
         variable equal to the maximum
         value, one is chosen randomly. To access to all patches or turtles among
         \code{agents} which have their variable equal
         to the maximum value, use \code{withMax()}.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1),
          val = sample(1:5, size = NLcount(patches(w1)), replace = TRUE))
plot(w1)
p1 <- maxOneOf(agents = patches(w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10),
                    heading = sample(1:3, size = 10, replace= TRUE))
t2 <- maxOneOf(agents = t1, var = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-one-of}
}

