% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\name{maxOneOf}
\alias{maxOneOf}
\alias{maxOneOf,matrix,worldMatrix,missing-method}
\alias{maxOneOf,matrix,worldArray,character-method}
\alias{maxOneOf,agentMatrix,missing,character-method}
\title{One \code{agent} with maximum}
\usage{
maxOneOf(agents, world, var)

\S4method{maxOneOf}{matrix,worldMatrix,missing}(agents, world)

\S4method{maxOneOf}{matrix,worldArray,character}(agents, world, var)

\S4method{maxOneOf}{agentMatrix,missing,character}(agents, var)
}
\arguments{
\item{agents}{Matrix (\code{ncol} = 2) with the first column \code{pxcor} and the second
              column \code{pycor} representing the \code{patches} coordinates, or

              \code{AgentMatrix} object representing the moving \code{agents}.}

\item{world}{\code{WorldMatrix} or \code{worldArray} object.}

\item{var}{Character. The name of the selected \code{agents} variable.
If \code{agents} are \code{patches} and the \code{world} is a
\code{worldMatrix} object, \code{var} must not be provided. If
\code{agents} are \code{patches} and the \code{world} is a \code{worldArray}
object, \code{var} is the name of the layer to use to define the \code{patches}
values. If \code{agents} are \code{turtles}, \code{var} is one of
the \code{turtles}' variable and can be equal to \code{xcor},
\code{ycor}, any of the variables created when \code{turtles} were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
Matrix (\code{ncol} = 2, \code{nrow} = 1) with the first column \code{pxcor} and
        the second column \code{pycor} representing the coordinates of the \code{patch}
        (or of one of the \code{patches}) among the \code{agents} which has its variable
        equals to the maximum value
        among the \code{agents}, or

        \code{AgentMatrix} of length 1 representing the \code{turtle} (or one of
        the \code{turtles}) among the \code{agents} which has its variable \code{var}
        equals to the maximum value
        among the \code{agents}.
}
\description{
Report one \code{patch} or one \code{turtle} among \code{agents} which has its variable equals
to the maximum value.
}
\details{
\code{world} must not be provided if \code{agents} are turtles.

         If there are several \code{patches} or \code{turtles} among \code{agents} with their
         variable equal to the maximum
         value, one is chosen randomly. To access to all \code{patches} or \code{turtles} among
         \code{agents} which have their variable equal
         to the maximum value, use \code{withMax()}.
}
\examples{
# Patches
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4,
                          data = sample(1:5, size = 25, replace = TRUE))
plot(w1)
p1 <- maxOneOf(agents = patches(w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10),
                    heading = sample(1:3, size = 10, replace= TRUE))
t2 <- maxOneOf(agents = t1, var = "heading")


}
\references{
Wilensky, U. 1999. NetLogo. \url{http://ccl.northwestern.edu/netlogo/}.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-one-of}
}
\author{
Sarah Bauduin
}
