% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPxcor}
\alias{maxPxcor}
\alias{maxPxcor,NLworldMs-method}
\title{Maximum pxcor}
\usage{
maxPxcor(world)

\S4method{maxPxcor}{NLworldMs}(world)
}
\arguments{
\item{world}{NLworldMs object.}
}
\value{
Integer.
}
\description{
Report the patches maximum pxcor in the \code{world}.
}
\examples{
w1 <- createNLworldMatrix()
maxPxcor(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-pcor}
}

