% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPxcor}
\alias{maxPxcor}
\alias{maxPxcor,RasterLayer-method}
\alias{maxPxcor,RasterStack-method}
\title{Maximum pxcor}
\usage{
maxPxcor(worldRaster)

\S4method{maxPxcor}{RasterLayer}(worldRaster)

\S4method{maxPxcor}{RasterStack}(worldRaster)
}
\arguments{
\item{worldRaster}{A \code{Raster} or \code{RasterStack} object.}
}
\value{
A numeric value
}
\description{
Report the maximum x-coordinate for patches.
}
\examples{
# Create a world with the default settings.
world <- createWorld()
maxPxcor(worldRaster = world)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

