% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\name{maxPxcor}
\alias{maxPxcor}
\alias{maxPxcor,worldNLR-method}
\title{Maximum \code{pxcor}}
\usage{
maxPxcor(world)

\S4method{maxPxcor}{worldNLR}(world)
}
\arguments{
\item{world}{\code{WorldMatrix} or \code{worldArray} object.}
}
\value{
Integer.
}
\description{
Report the patches maximum \code{pxcor} in the \code{world}.
}
\examples{
w1 <- createWorld()
maxPxcor(w1)


}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-pcor}
}
\author{
Sarah Bauduin
}
