% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPycor}
\alias{maxPycor}
\alias{maxPycor,NLworld-method}
\alias{maxPycor,NLworldMs-method}
\alias{maxPycor,NLworldStack-method}
\title{Maximum pycor}
\usage{
maxPycor(world)

\S4method{maxPycor}{NLworld}(world)

\S4method{maxPycor}{NLworldStack}(world)

\S4method{maxPycor}{NLworldMs}(world)
}
\arguments{
\item{world}{NLworlds or NLworldMs object.}
}
\value{
Integer.
}
\description{
Report the patches maximum pycor in the \code{world}.
}
\examples{
w1 <- createNLworld()
maxPycor(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#max-pcor}
}

