% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{maxPycor}
\alias{maxPycor}
\alias{maxPycor,NLworld-method}
\alias{maxPycor,NLworldStack-method}
\title{Maximum pycor}
\usage{
maxPycor(world)

\S4method{maxPycor}{NLworld}(world)

\S4method{maxPycor}{NLworldStack}(world)
}
\arguments{
\item{world}{A \code{NLworld*} object representing the world.}
}
\value{
A numeric value
}
\description{
Report the patches maximum \code{pycor} in a \code{NLworld*}.
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
maxPycor(world = w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

