% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\name{minPxcor}
\alias{minPxcor}
\alias{minPxcor,worldNLR-method}
\title{Minimum \code{pxcor}}
\usage{
minPxcor(world)

\S4method{minPxcor}{worldNLR}(world)
}
\arguments{
\item{world}{\code{WorldMatrix} or \code{worldArray} object.}
}
\value{
Integer.
}
\description{
Report the patches minimum \code{pxcor} in the \code{world}.
}
\examples{
w1 <- createWorld()
minPxcor(w1)

}
\references{
Wilensky, U. 1999. NetLogo. https://www.netlogo.org.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://docs.netlogo.org/dictionary.html#min-pcor}
}
\author{
Sarah Bauduin
}
