% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{minPycor}
\alias{minPycor}
\alias{minPycor,NLworldMs-method}
\title{Minimum pycor}
\usage{
minPycor(world)

\S4method{minPycor}{NLworldMs}(world)
}
\arguments{
\item{world}{NLworldMs object.}
}
\value{
Integer.
}
\description{
Report the patches minimum pycor in the \code{world}.
}
\examples{
w1 <- createWorld()
minPycor(w1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#min-pcor}
}

