% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\name{minPycor}
\alias{minPycor}
\alias{minPycor,worldNLR-method}
\title{Minimum \code{pycor}}
\usage{
minPycor(world)

\S4method{minPycor}{worldNLR}(world)
}
\arguments{
\item{world}{\code{WorldMatrix} or \code{worldArray} object.}
}
\value{
Integer.
}
\description{
Report the patches minimum \code{pycor} in the \code{world}.
}
\examples{
w1 <- createWorld()
minPycor(w1)


}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#min-pcor}
}
\author{
Sarah Bauduin
}
