% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{moveTo}
\alias{moveTo}
\alias{moveTo,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{moveTo,SpatialPointsDataFrame,matrix-method}
\title{Move to}
\usage{
moveTo(turtles, agents)

\S4method{moveTo}{SpatialPointsDataFrame,matrix}(turtles, agents)

\S4method{moveTo}{SpatialPointsDataFrame,SpatialPointsDataFrame}(turtles,
  agents)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{agents}{Matrix (ncol = 2) with the first column \code{pxcor} and the
               second column \code{pycor} representing the coordinates of the
               patches where the \code{turtles} move to.

               SpatialPointsDataFrame created by \code{createTurtles()} or
               by \code{createOTurtles()} representing the turtles whose
               locations the \code{turtles} move to.}
}
\value{
SpatialPointsDataFrame representing the moving turtles with updated
        coordinates and updated data for their previous coordinates "prevX"
        and "prevY".
}
\description{
Move the turtles to the \code{agents} locations.
}
\details{
The number of \code{agents} must be equal to 1 or of
         length \code{turtles}.

         The turtle's headings are not affected with this function.

         If a turtle is moving to a patch location, it will be located at
         the patch center.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 5, coords = randomXYcor(w1, n = 5))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1")

t1 <- moveTo(turtles = t1, agents = turtle(t1, who = 0))
Plot(t1, addTo ="w1")
t1 <- moveTo(turtles = t1, agents = patch(w1, 9, 9))
Plot(t1, addTo ="w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#move-to}
}

