% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{moveTo}
\alias{moveTo}
\alias{moveTo,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{moveTo,SpatialPointsDataFrame,matrix-method}
\alias{moveTo,agentMatrix,matrix-method}
\title{Move to}
\usage{
moveTo(turtles, agents)

\S4method{moveTo}{SpatialPointsDataFrame,matrix}(turtles, agents)

\S4method{moveTo}{SpatialPointsDataFrame,SpatialPointsDataFrame}(turtles,
  agents)

\S4method{moveTo}{agentMatrix,matrix}(turtles, agents)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated coordinates
        and updated data for their previous coordinates "prevX" and "prevY".
}
\description{
Move the \code{turtles} to the \code{agents}' locations.
}
\details{
The number of \code{agents} must be equal to 1 or to
         length \code{turtles}.

         The \code{turtle}'s headings are not affected with this function.

         If a turtle is moving to a patch location, it will be located at
         the patch center.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1 <- set(world = w1, agents = patches(w1), val = runif(NLcount(patches(w1))))
t1 <- createTurtles(n = 5, coords = randomXYcor(w1, n = 5))

#clearPlot()
#Plot(w1)
#Plot(t1, addTo ="w1")

t1 <- moveTo(turtles = t1, agents = turtle(t1, who = 0))
#Plot(t1, addTo ="w1")
t1 <- moveTo(turtles = t1, agents = patch(w1, 9, 9))
#Plot(t1, addTo ="w1")

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#move-to}
}

