% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{neighbors}
\alias{neighbors}
\alias{neighbors,NLworlds,SpatialPointsDataFrame,numeric-method}
\alias{neighbors,NLworlds,matrix,numeric-method}
\title{Find neighbors patches}
\usage{
neighbors(world, agents, nNeighbors, torus = FALSE)

\S4method{neighbors}{NLworlds,matrix,numeric}(world, agents, nNeighbors,
  torus = FALSE)

\S4method{neighbors}{NLworlds,SpatialPointsDataFrame,numeric}(world, agents,
  nNeighbors, torus = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{agents}{Matrix (ncol = 2) with the first column \code{pxcor} and
                  the second column \code{pycor} representing the patches
                  coordinates for which neighbors will be reported.

                  SpatialPointsDataFrame created by \code{createTurtles()} or
                  by \code{createOTurtles()} representing the turtles around
                  which the neighbors patches will be reported.}

\item{nNeighbors}{Integer: 4 or 8. The number of neighbor patches to identify.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
List with each item being the patches coordinates of the neighbors for
        each agent. The list items follows the order of the \code{agents}.
}
\description{
Report the 4 or 8 surrounding patches (neighbors) around patches or turtles.
}
\details{
The patch or turtle around which the neighbors are identified is not
         returned.

         If \code{torus = FALSE}, \code{agents} located on the edges of the world
         have less than \code{nNeighbors} patches. If \code{torus = FALSE}, agents
         located on the egdes of the world all have \code{nNeighbors} patches which
         some may be on the other side of the world.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
neighbors(world = w1, agents = patch(w1, c(0,9), c(0,7)), nNeighbors = 8)
t1 <- createTurtles(n = 3, coords = randomXYcor(w1, n = 3))
neighbors(world = w1, agents = t1, nNeighbors = 4)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#neighbors}
}

