% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{neighbors}
\alias{neighbors}
\alias{neighbors,NLworlds,SpatialPointsDataFrame,numeric-method}
\alias{neighbors,NLworlds,matrix,numeric-method}
\title{Neighbors}
\usage{
neighbors(world, agents, nNeighbors, torus = FALSE)

\S4method{neighbors}{NLworlds,matrix,numeric}(world, agents, nNeighbors,
  torus = FALSE)

\S4method{neighbors}{NLworlds,SpatialPointsDataFrame,numeric}(world, agents,
  nNeighbors, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and
the second column \code{pycor} representing the patch(es)
coordinates for which neighbors will be reported.
A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtles around
which the neighbors patches will be reported.}

\item{nNeighbors}{4 or 8 for the number of neighbor patches.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A list with each item being the patches coordinates of the neighbors for
        each of the agents. The list items follows the order of the agents.
}
\description{
Reports the 4 or 8 surrounding patches (neighbors) around patch(es) or turtle(s).
}
\details{
The patch around which the neighbors are identified if \code{agents}
         are patches or the patch on which the turtle is located if \code{agents}
         are turtles is not reported.
         If \code{torus = FALSE}, \code{agents} located on the edges of the world
         have less than \code{nNeighbors} patches. If \code{torus = FALSE}, agents
         located on the egdes of the world all have \code{nNeighbors} patches which
         some may be on the other side of the world.
}
\examples{
# Create a NLworld
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
neighbors(world = w1, agents = cbind(pxcor = c(0, 4), pycor = c(0, 6)), nNeighbors = 8)
t1 <- createTurtles(world = w1, n = 3, coords = cbind(xcor = randomXcor(world = w1, n = 3),
                                                      ycor = randomYcor(world = w1, n = 3)))
neighbors(world = w1, agents = t1, nNeighbors = 4)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

