% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{package}
\name{NetLogoR-package}
\alias{NetLogoR}
\alias{NetLogoR-package}
\title{The \code{NetLogoR} package}
\description{
The 3 specific R examples can be opened here, but it's recommended to make copies of these
for editing:
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{exPath <- file.path(system.file(package = "NetLogoR", "examples")
file.edit(exPath, "Ants", "Ants.R"))
file.edit(exPath, "Butterfly", "Butterfly-1.R"))
file.edit(exPath, "Wolf-Sheep-Predation", "Wolf-Sheep-Predation.R"))
}\if{html}{\out{</div>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://netlogor.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/NetLogoR/}
  \item \url{https://groups.google.com/g/netlogor}
  \item Report bugs at \url{https://github.com/PredictiveEcology/NetLogoR/issues}
}

}
\author{
\strong{Maintainer}: Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})

Authors:
\itemize{
  \item Sarah Bauduin \email{sarahbauduin@hotmail.fr} (\href{https://orcid.org/0000-0002-3252-5894}{ORCID})
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

Other contributors:
\itemize{
  \item His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
