% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{package}
\name{NetLogoR-package}
\alias{NetLogoR}
\alias{NetLogoR-package}
\title{The \code{NetLogoR} package}
\description{
The suggested package \pkg{fastshp} can be installed with
\code{install.packages("fastshp", repos = "https://rforge.net", type = "source")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://netlogor.predictiveecology.org}
  \item Report bugs at \url{https://github.com/PredictiveEcology/NetLogoR/issues}
}

}
\author{
\strong{Maintainer}: Sarah Bauduin \email{sarahbauduin@hotmail.fr} (0000-0002-3252-5894)

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@canada.ca} (0000-0002-6914-8316)
  \item Alex M Chubaty \email{alex.chubaty@gmail.com} (0000-0001-7146-8135)
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
