% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{package}
\name{NetLogoR-package}
\alias{NetLogoR}
\alias{NetLogoR-package}
\title{The \code{NetLogoR} package}
\description{
The suggested package \pkg{fastshp} can be installed with
\code{install.packages("fastshp", repos = "https://rforge.net", type = "source")}.
The examples included with the package, are located in the R package "examples" folder,
which can be found at \code{system.file(package = "NetLogoR", "examples")}. The 3 specific
R examples can be opened here:
\code{file.edit(file.path(system.file(package = "NetLogoR", "examples"), "Ants", "Ants.R"))},
\code{file.edit(file.path(system.file(package = "NetLogoR", "examples"), "Butterfly", "Butterfly-1.R"))},
or
\code{file.edit(file.path(system.file(package = "NetLogoR", "examples"), "Wolf-Sheep-Predation", "Wolf-Sheep-Predation.R"))}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://netlogor.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/NetLogoR/}
  \item Report bugs at \url{https://github.com/PredictiveEcology/NetLogoR/issues}
}

}
\author{
\strong{Maintainer}: Sarah Bauduin \email{sarahbauduin@hotmail.fr} (\href{https://orcid.org/0000-0002-3252-5894}{ORCID})

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
  \item Alex M Chubaty \email{alex.chubaty@gmail.com} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
