% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{package}
\name{NetLogoR-package}
\alias{NetLogoR}
\alias{NetLogoR-package}
\title{The \code{NetLogoR} package}
\description{
The suggested package \pkg{fastshp} can be installed with
\code{install.packages("fastshp", repos = "http://rforge.net", type = "source")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://netlogor.predictiveecology.org}
  \item Report bugs at \url{https://github.com/PredictiveEcology/NetLogoR/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alexander.chubaty@canada.ca}

Authors:
\itemize{
  \item Sarah Bauduin \email{sarahbauduin@hotmail.fr}
  \item Eliot J B McIntire \email{eliot.mcintire@canada.ca}
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
