% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{NLany}
\alias{NLany}
\alias{NLany,matrix-method}
\title{Any}
\usage{
NLany(agents)

\S4method{NLany}{matrix}(agents)
}
\arguments{
\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patches to be evaluated.}
}
\value{
Logical. \code{TRUE} if there is at least one patch coordinates in the
        \code{agents}, return \code{FALSE} otherwise.
}
\description{
Reports \code{TRUE} if the given agentset \code{agents} is non empty, or \code{FALSE}
otherwise.
}
\details{
!!! Only implemented for patches so far !!!
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
p1 <- noPatches()
p2 <- patch(world = w1, xcor = 0, ycor = 0)
NLany(p1)
NLany(p2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

