% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{NLany}
\alias{NLany}
\alias{NLany,SpatialPointsDataFrame-method}
\alias{NLany,matrix-method}
\title{Any agents?}
\usage{
NLany(agents)

\S4method{NLany}{matrix}(agents)

\S4method{NLany}{SpatialPointsDataFrame}(agents)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column \code{pxcor} and the
              second column \code{pycor} representing the coordinates of the
              patches to evaluate.

              SpatialPointsDataFrame created by \code{createTurtles()} or by
              \code{createOTurtles()} representing the turtles to evaluate.}
}
\value{
Logical. \code{TRUE} if there is at least one patch or one turtle in the
        \code{agents}, \code{FALSE} otherwise.
}
\description{
Report \code{TRUE} if \code{agents} is non empty, or \code{FALSE} otherwise.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
p1 <- noPatches()
p2 <- patch(w1, 0, 0)
NLany(p1)
NLany(p2)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
t2 <- noTurtles()
NLany(t1)
NLany(t2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#any}
}

