% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\name{NLcount}
\alias{NLcount}
\alias{count}
\alias{NLcount,matrix-method}
\title{Count agents}
\usage{
NLcount(agents)

\S4method{NLcount}{matrix}(agents)
}
\arguments{
\item{agents}{Matrix (\code{ncol} = 2) with the first column \code{pxcor} and the second
column \code{pycor} representing the \code{patches} coordinates, or

\if{html}{\out{<div class="sourceCode">}}\preformatted{          `AgentMatrix` object representing the moving `agents`.
}\if{html}{\out{</div>}}}
}
\value{
Integer.
}
\description{
Report the number of patches or turtles inside \code{agents}.
}
\examples{
# Patches
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
p1 <- patches(w1)
NLcount(p1) # 25 patches

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
NLcount(t1) # 10 turtles


}
\references{
Wilensky, U. 1999. NetLogo. \url{http://ccl.northwestern.edu/netlogo/}.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#count}
}
\author{
Sarah Bauduin
}
