% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{NLstack}
\alias{NLstack}
\alias{NLstack,NLworld-method}
\title{Creating a \code{NLworldStack}}
\usage{
NLstack(...)

\S4method{NLstack}{NLworld}(...)
}
\arguments{
\item{...}{\code{NLworld} objects.}
}
\value{
\code{NLworldStack} object with the \code{NLworld} stacked as layers.
}
\description{
Stacking several \code{NLworld} together to obtain a single \code{NlworldStack}
which contains different layers for the different \code{NLworld} values.
}
\examples{
# Create 2 NLworld with the default settings (same extent) but different values.
w1 <- createNLworld()
w1[] <- runif(length(w1))
w2 <- createNLworld()
w2[] <- runif(length(w2))
# Stack them together.
w3 <- NLstack(w1, w2)

library(SpaDES)
clearPlot()
Plot(w3)


}
\author{
Sarah Bauduin
}

