% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworld-class.R
\docType{methods}
\name{NLstack}
\alias{NLstack}
\alias{NLstack,NLworld,NLworld-method}
\title{Creating a \code{NLworldStack}}
\usage{
NLstack(world1, world2)

\S4method{NLstack}{NLworld,NLworld}(world1, world2)
}
\arguments{
\item{world1}{A \code{NLworld} object.}

\item{world2}{A \code{NLworld} object.}
}
\value{
A \code{NLworldStack} object with the \code{NLworld} stacked as layers.
}
\description{
Stacking several \code{NLworld} together to obtain a single \code{NlworldStack}
which contains different layers for the different \code{NLworld} values.
}
\examples{
# Create 2 worlds with the default settings but different values.
w1 <- createNLworld()
w2 <- createNLworld()
w1[] <- runif(n = 1089)
w2[] <- runif(n = 1089)
# Stack the 2 world together.
w3 <- NLstack(w1, w2)

library(SpaDES)
Plot(w3)

}
\author{
Sarah Bauduin
}

