% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{methods}
\name{NLstack}
\alias{NLstack}
\alias{NLstack,NLworld-method}
\title{Create a NLworldStack}
\usage{
NLstack(...)

\S4method{NLstack}{NLworld}(...)
}
\arguments{
\item{...}{NLworld objects.}
}
\value{
NLworldStack object with the NLworld stacked as layers.
}
\description{
Stack multiple NLworld objects.
}
\details{
The NLworld objects must have the same extents and cannot be empty
         (i.e., the patches values must be different than \code{NA}).
}
\examples{
w1 <- createNLworld()
w1 <- set(world = w1, agents = patches(w1), val = runif(count(patches(w1))))
w2 <- createNLworld()
w2 <- set(world = w2, agents = patches(w2), val = runif(count(patches(w2))))
w3 <- NLstack(w1, w2)

#clearPlot()
#Plot(w3)

}
\author{
Sarah Bauduin
}

