% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{NLwith}
\alias{NLwith}
\alias{NLwith,NLworld,matrix,missing,numeric-method}
\alias{NLwith,NLworldStack,matrix,character,numeric-method}
\title{NLwith}
\usage{
NLwith(world, agents, pVar, val)

\S4method{NLwith}{NLworld,matrix,missing,numeric}(world, agents, val)

\S4method{NLwith}{NLworldStack,matrix,character,numeric}(world, agents, pVar,
  val)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patches to be evaluated.}

\item{pVar}{If the world is a \code{NLworldStack}, pVar is the name (characters)
of the layer used for evaluating the conditional value.}

\item{val}{Numeric or character depending on the variable class.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the coordinates of the patches with their value
        equal to \code{val}.
}
\description{
Reports the coordinates \code{pxcor} and \code{pycor} of the patches with their
variable equals to a specific value.
}
\details{
!!! Only implemented for patches so far !!!
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
p2 <- NLwith(world = w1, agents = patches(world = w1), val = 2)
plot(w1)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

