% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworld-class}
\alias{NLworld}
\alias{NLworld-class}
\alias{[}
\alias{[<-}
\title{The NLworld class}
\usage{
\S4method{[}{NLworld,numeric,numeric,ANY}(x, i, j, drop)

\S4method{[}{NLworld,numeric,numeric,ANY}(x, i, j) <- value
}
\description{
A \code{NLworld} object is a grid composed of squared patches (i.e., pixels)
that behaves mostly the same as a \code{RasterLayer} object.
Patches have two coordinates \code{pxcor} and \code{pycor}, representing the
location of their center. \code{pxcor} and \code{pycor} are always integer
and increment by 1. \code{pxcor} increases as you move right and \code{pycor}
increases as you move up.  \code{pxcor} and \code{pycor} can be negative if
there are patches to the left or below the patch \code{[pxcor = 0, pycor = 0]}.
}
\details{
When creating a \code{NLworld} object, its extent is equal to \code{xmin = minPxcor - 0.5},
\code{xmax = maxPxcor + 0.5}, \code{ymin = minPycor - 0.5}, and \code{ymax = maxPycor + 0.5}.
The number of patches created is then equal to
\code{((maxPxcor - minPxcor) + 1) * ((maxPycor - minPycor) + 1)}.

\code{[]} can be used to extract values from a \code{NLworld} object by using
the patch coordinates \code{[pxcor, pyxor]}. When multiple coordinates are provided,
the order of the values returned matches the order of the cell numbers as defined
for a \code{RasterLayer}. \code{[] <-} can be used to replace patch values in a
\code{NLworld} object. Similarly, when replacing values of several patches, the
values should be given in the order of the cells numbers as defined for a \code{RasterLayer}.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

