% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworld-class.R
\docType{class}
\name{NLworld-class}
\alias{NLworld-class}
\title{The \code{NLworld} class}
\description{
Behaves the same as a \code{RasterLayer} object except:
}
\details{
A \code{NLworld} is a grid composed of squared patches of resolution 1.
Patches have two coordinates \code{pxcor} and \code{pycor}.
\code{pxcor} and \code{pycor} are always integer.
\code{pxcor} and \code{pycor} represent the raster coordinates in the bottom left
corner of the patches.
When creating a \code{NLworld}, the extent of the \code{NLworld} is of
\code{xmin = minPxcor}, \code{xmax = maxPxcor}, \code{ymin = minPycor}, and
\code{ymax = maxPycor}. The number of patches is equal to
\code{((maxPxcor - minPxcor) + 1) * ((maxPycor - minPycor) + 1)}.
\code{pxcor} can be seen as a column number and \code{pycor} can be seen as a
row number but increasing as you move up. However, \code{pxcor} and \code{pycor}
can be negative if there are patches to the left or below the patch \code{[0,0]}.
The use of \code{[]} to extract \code{Raster*} cells by row and column numbers has
been redefined to extract \code{NLworld} patches using the patches' coordinates
\code{[pxcor,pyxor]}.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

