% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworldArray}
\alias{NLworldArray}
\alias{[}
\alias{[<-}
\title{The NLworldArray class}
\usage{
\S4method{[}{NLworldArray,numeric,numeric,ANY}(x, i, j, drop)

\S4method{[}{NLworldArray,missing,missing,ANY}(x, drop)

\S4method{[}{NLworldArray,numeric,numeric,matrix}(x, i, j) <- value

\S4method{[}{NLworldArray,missing,missing,matrix}(x) <- value
}
\description{
This is similar to \code{NLworldStack}, but it is an s4 class extension of
\code{array}.
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
\seealso{
NLworldStack
}

