% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-classes.R
\docType{methods}
\name{NLworldArray}
\alias{NLworldArray}
\alias{NLworldArray,NLworldMatrix-method}
\title{Create a NLworldArray}
\usage{
NLworldArray(...)

\S4method{NLworldArray}{NLworldMatrix}(...)
}
\arguments{
\item{...}{NLworldMatrix objects.}
}
\value{
NLworldArray object.
}
\description{
Stack multiple NLworldMatrix data in an array.
}
\details{
The NLworldMatrix objects must all have the same extents.
}
\examples{
w1 <- createNLworldMatrix(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 1:25)
w2 <- createNLworldMatrix(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = 25:1)
w3 <- NLworldArray(w1, w2)
plot(world2raster(w3))

}
\author{
Sarah Bauduin
}

