% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\name{NLworldMs-class}
\alias{NLworldMs}
\alias{NLworldMs-class}
\title{The NLworldMs class}
\description{
The \code{NLworldMs} class is the union of the \code{NLworldMatrix} and \code{NLworldArray}
classes. Mostly used for building function purposes.
}
\section{Slots}{

\describe{
\item{\code{members}}{NLworldMatrix, NLworldArray}
}}
\author{
Sarah Bauduin, and Eliot McIntire
}

