% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworld-class.R
\name{NLworlds-class}
\alias{NLworlds}
\alias{NLworlds-class}
\title{The \code{NLworlds} class}
\description{
This class is the union of the \code{NLworld} and \code{NLworldStack} classes.
}
\section{Slots}{

\describe{
\item{\code{members}}{NLworld, NLworldStack}
}}
\author{
Sarah Bauduin
}

