% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\name{NLworlds-class}
\alias{NLworlds}
\alias{NLworlds-class}
\title{The \code{NLworlds} class}
\description{
The \code{NLworlds} class is the union of the \code{NLworld} and \code{NLworldStack}
classes. Mostly used for building function purposes.
}
\section{Slots}{

\describe{
\item{\code{members}}{NLworld, NLworldStack}
}}
\author{
Sarah Bauduin

Sarah Bauduin
}

