% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworldStack-class}
\alias{NLworldStack}
\alias{NLworldStack-class}
\alias{[}
\alias{[<-}
\title{The \code{NLworldStack} class}
\usage{
\S4method{[}{NLworldStack,numeric,numeric}(x, i, j, drop)

\S4method{[}{NLworldStack,numeric,numeric,matrix}(x, i, j) <- value

\S4method{[}{NLworldStack,numeric,numeric,numeric}(x, i, j) <- value
}
\description{
A \code{NLworldStack} object is similar to a \code{RasterStack} object, it is
a collection of \code{NLworld} objects with the same extent.
}
\author{
Sarah Bauduin
}

