% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLworlds-class.R
\docType{class}
\name{NLworldStack-class}
\alias{NLworldStack}
\alias{NLworldStack-class}
\title{The NLworldStack class}
\description{
A \code{NLworldStack} object is similar to a \code{RasterStack} object, it is
a collection of \code{NLworld} objects with the same extent.
}
\author{
Sarah Bauduin
}

