% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{nOf}
\alias{nOf}
\alias{nOf,SpatialPointsDataFrame,numeric-method}
\alias{nOf,matrix,numeric-method}
\title{N random agents}
\usage{
nOf(agents, n)

\S4method{nOf}{matrix,numeric}(agents, n)

\S4method{nOf}{SpatialPointsDataFrame,numeric}(agents, n)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{n}{Integer.}
}
\value{
Matrix (ncol = 2, nrow = \code{n}) with the first column "pxcor"
        and the second  column "pycor" representing the coordinates of the
        selected patches from \code{agents}, or

        SpatialPointsDataFrame of length \code{n} representing the turtles
        selected from \code{agents}.
}
\description{
Report \code{n} patches or turtles randomly selected among \code{agents}.
}
\details{
\code{n} must be less or equal the number of patches or turtles in \code{agents}.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
pSelect <- nOf(agents = patches(w1), n = 5)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
tSelect <- nOf(agents = t1, n = 2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#n-of}
}

