% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{noPatches}
\alias{noPatches}
\alias{noPatches,missing-method}
\title{No patches}
\usage{
noPatches(x)

\S4method{noPatches}{missing}()
}
\value{
Matrix (ncol = 2, nrow = 0) with the first column "pxcor" and the
        second column "pycor".
}
\description{
Report an empty patch agentset.
}
\examples{
p1 <- noPatches()
nrow(p1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#no-patches}
}

