% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\name{noPatches}
\alias{noPatches}
\title{No \code{patches}}
\usage{
noPatches()
}
\value{
Matrix (\code{ncol} = 2, \code{nrow} = 0) with the first column \code{pxcor} and the
second column \code{pycor}.
}
\description{
Report an empty \code{patch} \code{agentset}.
}
\examples{
p1 <- noPatches()
NLcount(p1)

}
\references{
Wilensky, U. 1999. NetLogo. \url{http://ccl.northwestern.edu/netlogo/}.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#no-patches}
}
\author{
Sarah Bauduin
}
