% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{noTurtles}
\alias{noTurtles}
\alias{noTurtles,missing-method}
\title{No turtles}
\usage{
noTurtles(x)

\S4method{noTurtles}{missing}()
}
\arguments{
\item{x}{documentation needed}
}
\value{
SpatialPointsDataFrame with the turtle variables defined
        as when using \code{createTurtles()} or \code{createOTurtles()} but
        of length equals to 0.
}
\description{
Report an empty turtle agentset.
}
\examples{
t1 <- noTurtles()
NLcount(t1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#no-turtles}
}

