% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\name{noTurtles}
\alias{noTurtles}
\title{No turtles}
\usage{
noTurtles()
}
\value{
AgentMatrix with the turtle variables defined as when using
        \code{createTurtles()} but with 0 turtles.
}
\description{
Report an empty turtle agentset.
}
\examples{
t1 <- noTurtles()
NLcount(t1)

}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#no-turtles}
}
\author{
Sarah Bauduin
}
