% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\name{noTurtles}
\alias{noTurtles}
\title{No \code{turtles}}
\usage{
noTurtles()
}
\value{
\code{AgentMatrix} with the \code{turtle} variables defined as when using
\code{createTurtles()} but with 0 \code{turtle}.
}
\description{
Report an empty \code{turtle} \code{agentset}.
}
\examples{
t1 <- noTurtles()
NLcount(t1)

}
\references{
Wilensky, U. 1999. NetLogo. https://www.netlogo.org.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://docs.netlogo.org/dictionary.html#no-turtles}
}
\author{
Sarah Bauduin
}
