% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{of}
\alias{of}
\alias{of,SpatialPointsDataFrame,character-method}
\title{Values of a turtles variable}
\usage{
of(turtles, tVar)

\S4method{of}{SpatialPointsDataFrame,character}(turtles, tVar)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{tVar}{Character. The name of one turtles' variable to report. Can be equal to
\code{"coords"}, \code{"xcor"},
\code{"ycor"}, \code{"who"}, \code{"heading"}, \code{"prevCoords"},
\code{"prevX"}, \code{"prevY"}, \code{"breed"}, \code{"color"} or
any of the variables created with \code{turtlesOwn()}.}
}
\value{
Vector of \code{tVar} values for the \code{turtles}. The class depends
        of the variable class. The order of the vector follows the order
        of the \code{turtles}.
}
\description{
Report the \code{turtles} values for the requested variable.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
of(turtles = t1, tVar = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#of}
}

