% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{other}
\alias{other}
\alias{other,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{other,matrix,matrix-method}
\title{Other}
\usage{
other(agents, except)

\S4method{other}{matrix,matrix}(agents, except)

\S4method{other}{SpatialPointsDataFrame,SpatialPointsDataFrame}(agents, except)
}
\arguments{
\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patches to evaluate.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles to evaluate.}

\item{except}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) to remove from the \code{agents}.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) to remove from
the \code{agents}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second
        column \code{pycor} representing the patches in \code{agents} with
        the patch(es) in \code{except} removed.
        A SpatialPointsDataFrame representing the turtles in \code{agents} with
        the turtle(s) in \code{except} removed.
}
\description{
Reports all \code{agents} except specific one(s).
}
\details{
Both \code{agents} and \code{except} must be of the same class (e.g., both
         patches or both turtles).
         Carefull: this function removes turtles only based on similar who numbers
         and breeds.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9) # 100 patches
p1 <- other(agents = patches(world = w1), except = cbind(pxcor = 0, pycor = 0))
nrow(p1)

# Turtles
t1 <- createTurtles(n = 10, coords = cbind(xcor = 0, ycor = 0))
t2 <- other(agents = t1, except = turtle(turtles = t1, who = 0))
length(t2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

