% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{patch}
\alias{patch}
\alias{patch,NLworld,numeric,numeric-method}
\alias{patch,NLworldStack,numeric,numeric-method}
\title{Patch}
\usage{
patch(world, xcor, ycor)

\S4method{patch}{NLworld,numeric,numeric}(world, xcor, ycor)

\S4method{patch}{NLworldStack,numeric,numeric}(world, xcor, ycor)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{xcor}{A vector of \code{xcor} coordinates.}

\item{ycor}{A vector of \code{ycor} coordinates.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patch(es) coordinates. Each row represents
        a patch and the order is the one given in the coordinates \code{xcor} and
        \code{ycor}.
}
\description{
Reports the patch(es) coordinates \code{pxcor} and \code{pycor} at the given
\code{xcor} and \code{ycor} coordinates.
}
\details{
This function is equivalent to round the values \code{xcor} and \code{ycor}.
         If \code{xcor} or \code{ycor} are outside the world's extent, \code{NA, NA} is
         returned for the patch coordinates.
}
\examples{
# Create a NLworld
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
patch(world = w1, xcor = c(8.9, 5), ycor = c(-0.1, 12.4))

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

