% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{patch}
\alias{patch}
\alias{patch,NLworlds,numeric,numeric-method}
\title{Patch}
\usage{
patch(world, x, y, duplicate = FALSE, torus = FALSE, out = FALSE)

\S4method{patch}{NLworlds,numeric,numeric}(world, x, y, duplicate = FALSE,
  torus = FALSE, out = FALSE)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{x}{Numeric. Vector of \code{x} coordinates. Must be of same
length as \code{y}.}

\item{y}{Numeric. Vector of \code{y} coordinates. Must be of same
length as \code{x}.}

\item{duplicate}{Logical. If more than one location \code{[x, y]}
fall into the same patch and \code{duplicate == TRUE}, the
patch coordinates are returned the number of times the locations.
If \code{duplicate == FALSE}, the patch coordinates
are only returned once.
Default is \code{duplicate == FALSE}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}

\item{out}{Logical to determine if the coordinates for the patches outside of the
world should be returned. Default is \code{out = FALSE}.}
}
\value{
Matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates at \code{[x, y]}.
}
\description{
Report the patches coordinates \code{[pxcor, pycor]} at the given
\code{[x, y]} locations.
}
\details{
If a location \code{[x, y]} is outside of the world's extent and
         \code{torus = FALSE} and \code{out = FALSE}, no patch coordinates are returned;
         if \code{torus = FALSE} and \code{out = TRUE}, \code{NA} are returned;
         if \code{torus = TRUE}, the patch coordinates from a wrapped world are
         returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
patch(world = w1, x = c(0, 9.1, 8.9, 5, 5.3), y = c(0, 0, -0.1, 12.4, 12.4))
patch(world = w1, x = c(0, 9.1, 8.9, 5, 5.3), y = c(0, 0, -0.1, 12.4, 12.4), duplicate = TRUE)
patch(world = w1, x = c(0, 9.1, 8.9, 5, 5.3), y = c(0, 0, -0.1, 12.4, 12.4), torus = TRUE)
patch(world = w1, x = c(0, 9.1, 8.9, 5, 5.3), y = c(0, 0, -0.1, 12.4, 12.4), torus = TRUE, duplicate = TRUE)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#patch}
}

