% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-functions.R
\docType{methods}
\name{patchAt}
\alias{patchAt}
\alias{patchAt,NLworld,matrix,numeric,numeric-method}
\alias{patchAt,NLworldStack,matrix,numeric,numeric-method}
\title{Patch at}
\usage{
patchAt(world, agents, dx, dy, torus = FALSE)

\S4method{patchAt}{NLworld,matrix,numeric,numeric}(world, agents, dx, dy,
  torus = FALSE)

\S4method{patchAt}{NLworldStack,matrix,numeric,numeric}(world, agents, dx, dy,
  torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworld*} object.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
calling patches.}

\item{dx}{Numeric. Distance to east from the caller. If \code{dx} is negative, the
distance to the west is computed. \code{dx} must be a single value or
of the length of \code{agents}.}

\item{dy}{Numeric. Distance to the north from the caller. If \code{dy} is negative,
the distance to the south is computed. \code{dy} must be a single value or
of the length of \code{agents}.}

\item{torus}{Logical to determine if the \code{NLworld*} is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates. The order of the patches
        are the ones of the agents.
}
\description{
Reports the patch(es) coordinates \code{pxcor} and \code{pycor} at \code{(dx, dy)}
distance of the calling agent(s).
}
\details{
!!! Only implemented for the patches so far !!!

If \code{torus = FALSE} and the patch at distance \code{(dx, dy)}
         of the caller is outside the world, \code{NA} is returned. Otherwise,
         if \code{torus = TRUE}, the patch coordinates from the wrapped world are
         returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
patchCorner <- patchAt(world = w1, agents = cbind(pxcor = 0, pycor = 0), dx = 1, dy = 1)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

