% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-functions.R
\docType{methods}
\name{patches}
\alias{patches}
\alias{patches,NLworld-method}
\alias{patches,NLworldStack-method}
\title{All the patches in a world}
\usage{
patches(world)

\S4method{patches}{NLworld}(world)

\S4method{patches}{NLworldStack}(world)
}
\arguments{
\item{world}{\code{NLworlds} object.}
}
\value{
Matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the patches coordinates. The order of the patches
        follows the order of the cellnumbers as defined for a \code{Raster*}.
}
\description{
Report the coordinates \code{pxcor} and \code{pycor} for all patches in a
\code{NLworlds}.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9) # 100 patches
allPatches <- patches(world = w1)
nrow(allPatches)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#patches}
}

