% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{patchRight}
\alias{patchRight}
\alias{patchRight,NLworlds,SpatialPointsDataFrame,numeric,numeric-method}
\title{Patch right}
\usage{
patchRight(world, turtles, dist, nDegrees, torus = FALSE)

\S4method{patchRight}{NLworlds,SpatialPointsDataFrame,numeric,numeric}(world,
  turtles, dist, nDegrees, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles evolve onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{dist}{Numeric. Distance from the turtles. \code{dist} must be a single
value or of the length of \code{turtles}.}

\item{nDegrees}{Numeric. The number of degrees the turtle's heading should rotate
to the right to locate the patch. Must be of length 1 or of the
same length as the \code{turtles}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second
        column \code{pycor} representing the patches coordinates at \code{dist}
        of the turtles location and \code{nDegrees} to the right of their heading.
        The order of the patches follows the order of the \code{turtles}.
}
\description{
Reports the patch(es) coordinates at a given distance to the right of the turtle(s).
}
\details{
If \code{nDegrees} is negative, the turtle would rotate to the left.
         If \code{dist} is negative, the turtle would move backward.
         If a turtle is located outside of the world's extent, \code{NA} is
         returned for its patch coordinates.
         If \code{torus = FALSE} and the patch at distance \code{dist} and
         heading \code{nDregrees} to the right of the turtles is outside the
         world's extent, \code{NA} is returned. If \code{torus = TRUE}, the
         patch coordinates from the wrapped world are returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(world = w1, n = 1, coords = cbind(xcor = 2, ycor = 2), heading = 90)
patchRight(world = w1, turtles = t1, dist = 2, nDegrees = 90)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

