% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{data}
\name{.projNowhere}
\alias{.projNowhere}
\title{Internal CRS usage}
\format{An object of class \code{CRS} of length 1.}
\usage{
.projNowhere
}
\description{
This is used in various places when converting \code{agentMatrix} objects to
\code{SpatialPoints} for use with functions, such as \code{\link[raster]{buffer}}.
The \code{CRS()} operation is relatively time consuming.
Having a single object usable throughout is \emph{much} faster.
}
\keyword{internal}
