% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLogoR-package.R
\docType{data}
\name{projNowhere}
\alias{projNowhere}
\title{Internal CRS usage
This is used in various places when converting \code{agentMatrix} objects to \code{SpatialPoints}
for use with functions, such as \code{\link[raster]{buffer}}. The \code{CRS()} operation is
relatively time consuming. Having a single object useable throughout is MUCH faster.}
\format{An object of class \code{CRS} of length 1.}
\usage{
projNowhere
}
\description{
Internal CRS usage
This is used in various places when converting \code{agentMatrix} objects to \code{SpatialPoints}
for use with functions, such as \code{\link[raster]{buffer}}. The \code{CRS()} operation is
relatively time consuming. Having a single object useable throughout is MUCH faster.
}
\keyword{datasets}
