% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomXcor}
\alias{randomXcor}
\alias{randomXcor,NLworlds,numeric-method}
\title{Random xcor}
\usage{
randomXcor(world, n)

\S4method{randomXcor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{NLworlds object.}

\item{n}{Integer.}
}
\value{
Numeric. Vector of length \code{n} of xcor coordinates.
}
\description{
Report \code{n} random xcor coordinates within the \code{world}'s extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10,coords = cbind(xcor = randomXcor(world = w1, n = 10),
                                          ycor = randomYcor(world = w1, n = 10)))
library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#random-cor}
}

