% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomXYcor}
\alias{randomXYcor}
\alias{randomXYcor,NLworldMs,numeric-method}
\alias{randomXYcor,NLworlds,numeric-method}
\title{Random turtles coordinates}
\usage{
randomXYcor(world, n)

\S4method{randomXYcor}{NLworlds,numeric}(world, n)

\S4method{randomXYcor}{NLworldMs,numeric}(world, n)
}
\arguments{
\item{world}{NLworlds object.}

\item{n}{Integer.}
}
\value{
Matrix (ncol = 2, nrow = \code{n}) with the first column "xcor" and the second
        column "ycor".
}
\description{
Report \code{n} random xcor and ycor coordinates within the \code{world}'s extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1), val = runif(count(patches(w1))))
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10))

#clearPlot()
#Plot(w1)
#Plot(t1, addTo = "w1")

}
\author{
Sarah Bauduin
}

