% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomXYcor}
\alias{randomXYcor}
\alias{randomXYcor,NLworlds,numeric-method}
\title{Generante random turtles coordinates}
\usage{
randomXYcor(world, n)

\S4method{randomXYcor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{n}{Integer. The number of set of coordinates to generate.}
}
\value{
Matrix (ncol = 2, nrow = n) with the first column \code{xcor} and the second
        column \code{ycor}.
}
\description{
Report random xcor and ycor coordinates inside the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}

