% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomXYcor}
\alias{randomXYcor}
\alias{randomXYcor,NLworlds,numeric-method}
\title{Random xY cor}
\usage{
randomXYcor(world, n)

\S4method{randomXYcor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{n}{Numeric. The number of set of coordinates to generate.}
}
\value{
A matrix (ncol = 2) with the first column \code{xcor} and the second
        column \code{ycor} with nrow = \code{n}.
}
\description{
Reports random xcor and ycor coordinates inside the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10))
w1[] <- runif(25)

library(Spades)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}

