% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomYcor}
\alias{randomYcor}
\alias{randomYcor,NLworlds,numeric-method}
\title{Generate random ycor}
\usage{
randomYcor(world, n)

\S4method{randomYcor}{NLworlds,numeric}(world, n)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{n}{Integer. The number of ycor coordinates to generate.}
}
\value{
Numeric. Vector of ycor coordinate values of length \code{n}.
}
\description{
Report random ycor coordinates inside the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, coords = cbind(xcor = randomXcor(world = w1, n = 10),
                                           ycor = randomYcor(world = w1, n = 10)))
library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#random-cor}
}

