% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomYcor}
\alias{randomYcor}
\alias{randomYcor,NLworldMs,numeric-method}
\title{Random ycor}
\usage{
randomYcor(world, n)

\S4method{randomYcor}{NLworldMs,numeric}(world, n)
}
\arguments{
\item{world}{NLworldMs object.}

\item{n}{Integer.}
}
\value{
Numeric. Vector of length \code{n} of ycor coordinates.
}
\description{
Report \code{n} random ycor coordinates within the \code{world}'s extent.
}
\examples{
w1 <- createNLworldMatrix(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = runif(25))
t1 <- createTurtlesAM(n = 10, coords = cbind(xcor = randomXcor(world = w1, n = 10),
                                           ycor = randomYcor(world = w1, n = 10)))
plot(world2raster(w1))
points(turtles2spdf(t1), col = of(agents = t1, var = "color"), pch = 16)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#random-cor}
}

