% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{randomYcor}
\alias{randomYcor}
\title{Random ycor}
\usage{
randomYcor(world, n)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{n}{Numeric. The number of ycor coordinates to generate.}
}
\value{
A vector of ycor coordinate values of length \code{n}.
}
\description{
Reports random ycor coordinate(s) inside the world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, coords = cbind(xcor = randomXcor(world = w1, n = 10),
                                           ycor = randomYcor(world = w1, n = 10)))
w1[] <- runif(25)
plot(w1)
points(t1, pch = 16, col = t1@data$color)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

