% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{resizeWorld}
\alias{resizeWorld}
\alias{resizeWorld,NLworlds,numeric,numeric,numeric,numeric,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworlds,numeric,numeric,numeric,numeric,missing-method}
\title{Resize a world}
\usage{
resizeWorld(world, minPxcor, maxPxcor, minPycor, maxPycor, turtles)


  \S4method{resizeWorld}{NLworlds,numeric,numeric,numeric,numeric,SpatialPointsDataFrame}(world,
  minPxcor, maxPxcor, minPycor, maxPycor, turtles)

\S4method{resizeWorld}{NLworlds,numeric,numeric,numeric,numeric,missing}(world,
  minPxcor, maxPxcor, minPycor, maxPycor)
}
\arguments{
\item{world}{\code{NLworlds} object.}

\item{minPxcor}{Integer. Minimum \code{pxcor} for the patches (left border)
for the new world.}

\item{maxPxcor}{Integer. Maximum \code{pxcor} for the patches (right border)
for the new world.}

\item{minPycor}{Integer. Minimum \code{pycor} for the patches (bottom) for
the new world.}

\item{maxPycor}{Integer. Maximum \code{pycor} for the patches (top) for
the new world.}

\item{turtles}{SpatialPointsDataFrame object representing all the turtles
to remove.}
}
\value{
\code{NLworld} object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patch value are \code{NA}.
}
\description{
Change the size of a \code{NLworlds}. Previous patches and turtles are lost.
}
\details{
The given \code{NLworlds} and \code{turtles} are deleted, as well as the
         patches values contained in the \code{NLworlds}.
}
\examples{
w1 <- createNLworld()
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 10, world = w1)
w2 <- resizeWorld(world = w1, minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9, turtles = t1)
w2[] <- runif(n = 100)
plot(w1) # does not work, w1 does not exist anymore
plot(w2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#resize-world}
}

