% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{resizeWorld}
\alias{resizeWorld}
\alias{resizeWorld,NLworld,missing,missing,missing,missing,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworld,missing,missing,missing,missing,missing-method}
\alias{resizeWorld,NLworld,numeric,numeric,numeric,numeric,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworld,numeric,numeric,numeric,numeric,missing-method}
\alias{resizeWorld,NLworldStack,missing,missing,missing,missing,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworldStack,missing,missing,missing,missing,missing-method}
\alias{resizeWorld,NLworldStack,numeric,numeric,numeric,numeric,SpatialPointsDataFrame-method}
\alias{resizeWorld,NLworldStack,numeric,numeric,numeric,numeric,missing-method}
\title{Resize NLworld}
\usage{
resizeWorld(world, minPxcor, maxPxcor, minPycor, maxPycor, turtles)


  \S4method{resizeWorld}{NLworld,numeric,numeric,numeric,numeric,SpatialPointsDataFrame}(world,
  minPxcor, maxPxcor, minPycor, maxPycor, turtles)

\S4method{resizeWorld}{NLworld,numeric,numeric,numeric,numeric,missing}(world,
  minPxcor, maxPxcor, minPycor, maxPycor)


  \S4method{resizeWorld}{NLworld,missing,missing,missing,missing,SpatialPointsDataFrame}(world,
  turtles)

\S4method{resizeWorld}{NLworld,missing,missing,missing,missing,missing}(world)


  \S4method{resizeWorld}{NLworldStack,numeric,numeric,numeric,numeric,SpatialPointsDataFrame}(world,
  minPxcor, maxPxcor, minPycor, maxPycor, turtles)


  \S4method{resizeWorld}{NLworldStack,numeric,numeric,numeric,numeric,missing}(world,
  minPxcor, maxPxcor, minPycor, maxPycor)


  \S4method{resizeWorld}{NLworldStack,missing,missing,missing,missing,SpatialPointsDataFrame}(world,
  turtles)


  \S4method{resizeWorld}{NLworldStack,missing,missing,missing,missing,missing}(world)
}
\arguments{
\item{world}{A \code{NLworld} or \code{NLworldStack} object representing the current world.}

\item{minPxcor}{\code{pxcor} for patches on the left border of the new \code{NLworld}
Default value = \code{-16}, as in NetLogo.}

\item{maxPxcor}{\code{pxcor} for patches on the right border of the new \code{NLworld}
Default value = \code{16}, as in NetLogo.}

\item{minPycor}{\code{pycor} for patches at the bottom of the new \code{NLworld}
Default value \code{-16}, as in NetLogo.}

\item{maxPycor}{\code{pycor} for patches at the top of the new \code{NLworld}
Default value \code{16}, as in NetLogo.}

\item{turtles}{A \code{SpatialPointsDataFrame} object representing the turtles.}
}
\value{
A \code{NLworld} object composed of \code{(maxPxcor - minPxcor + 1) * (maxPycor - minPycor + 1)}
        patches. Patch value are \code{NA}.
}
\description{
Change the size of the \code{NLworld}. Previous patches and turtles are lost.
}
\details{
By using \code{resizeWorld}, the previous \code{NLworld}, patches and
         turtles are deleted.
}
\examples{
# Create a world with the default settings.
w1 <- createNLworld()
w1[] <- runif(n = 1089)
# Create one turtle
t1 <- SpatialPointsDataFrame(coords = matrix(c(1,2), nrow = 1, ncol = 2), data = data.frame(NA))
w2 <- resizeWorld(world = w1, minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9, turtles = t1)
w2[] <- runif(n = 100)
plot(w1) # does not work, w1 does not exist anymore
plot(w2)

}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

