% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{right}
\alias{right}
\alias{right,SpatialPointsDataFrame,numeric-method}
\title{Rotate to the right}
\usage{
right(turtles, angle)

\S4method{right}{SpatialPointsDataFrame,numeric}(turtles, angle)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the turtles to rotate.}

\item{angle}{Numeric. Angles in degrees by which to rotate the turtles'
heading to the right Must be of length 1 or of length \code{turtles}.}
}
\value{
SpatialPointsDataFrame representing the turtles with updated headings.
}
\description{
Rotate the turtles's heading to the right.
}
\details{
If \code{angle} is negative, the turtle rotates to the left.
}
\examples{
t1 <- createTurtles(n = 10, world = w1)
of(t1, tVar = "heading")
t1 <- right(turtles = t1, angle = 180)
of(t1, tVar = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#right}
}

