% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{set}
\alias{set}
\alias{set,NLworld,missing,matrix,missing-method}
\alias{set,NLworldArray,missing,matrix,character-method}
\alias{set,NLworldMatrix,missing,matrix,missing-method}
\alias{set,NLworldStack,missing,matrix,character-method}
\alias{set,missing,SpatialPointsDataFrame,SpatialPointsDataFrame,character-method}
\title{Set an agents variable}
\usage{
set(world, turtles, agents, var, val)

\S4method{set}{NLworld,missing,matrix,missing}(world, agents, val)

\S4method{set}{NLworldStack,missing,matrix,character}(world, agents, var, val)


  \S4method{set}{missing,SpatialPointsDataFrame,SpatialPointsDataFrame,character}(turtles,
  agents, var, val)

\S4method{set}{NLworldMatrix,missing,matrix,missing}(world, agents, val)

\S4method{set}{NLworldArray,missing,matrix,character}(world, agents, var, val)
}
\arguments{
\item{world}{NLworlds object.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{var}{Character. Vector of the name of the selected \code{agents} variables.
If \code{agents} are patches and the \code{world} is a
\code{NLworld} object, \code{var} must not be provided. If
\code{agents} are patches and the \code{world} is a \code{NLworldStack}
object, \code{var} is the name of the layers to use to define the patches
values. If \code{agents} are turtles, \code{var} is some of
the turtles' variable and can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}

\item{val}{Numeric or character. Vector of length 1 or length \code{count(agents)}
           if \code{length(var) == 1}, or

           Matrix or Dataframe (ncol = \code{length(var)}, nrow = \code{count(agents)}).
           Columns must be in the same order as \code{var}.}
}
\value{
NLworlds object with the values \code{val} assigned to the patches variables \code{var}
        for the \code{agents}, or

        SpatialPointsDataFrame representing the \code{turtles} with
        the values \code{val} assigned to the variables \code{var} for the \code{agents}.
}
\description{
Assign values to the \code{agents} for their selected variables.
}
\details{
If \code{agents} are patches, \code{world} must be provided and \code{turtles}
         must not be provided. If \code{agents} are turtles, \code{turtles} must be
         provided and \code{world} must not be provided.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1), val = 1)
w1 <- set(world = w1, agents = patch(w1, 0, 4), val = 0) # set the patch[0,0] to 0
of(world = w1, agents = patches(w1))

t1 <- createTurtles(n = 3, world = w1, heading = 0)
t2 <- set(turtles = t1, agents = turtle(t1, who = 0), var = "heading", val = 180) # set the heading of the turtle 0 to 180
of(agents = t2, var = "heading") # c(180, 0, 0)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#set}
}

