% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{setXY}
\alias{setXY}
\alias{setXY,SpatialPointsDataFrame,numeric,numeric,NLworlds,logical-method}
\alias{setXY,SpatialPointsDataFrame,numeric,numeric,missing,ANY-method}
\title{Set turtles' locations}
\usage{
setXY(turtles, xcor, ycor, world, torus = FALSE)

\S4method{setXY}{SpatialPointsDataFrame,numeric,numeric,missing,ANY}(turtles,
  xcor, ycor, torus)


  \S4method{setXY}{SpatialPointsDataFrame,numeric,numeric,NLworlds,logical}(turtles,
  xcor, ycor, world, torus = FALSE)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{xcor}{Numeric. Vector of x coordinates. Must be of length 1 or
of length \code{turtles}.}

\item{ycor}{Numeric. Vector of y coordinates. Must be of length 1 or
of length \code{turtles}.}

\item{world}{NLworlds object.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
SpatialPointsDataFrame representing the \code{turtles} with updated coordinates
        and updated data for their previous coordinates "prevX" and "prevY"
}
\description{
Set the turtles \code{xcor} and \code{ycor} coordinates.
}
\details{
\code{world} must be provided if only if \code{torus = TRUE}.

         If the given coordinates \code{[xcor, ycor]}
         are located outside of the \code{world}'s extent and \code{torus = TRUE},
         then the coordinates assigned to the turtle
         are the ones from a wrapped \code{word}; if \code{torus = FALSE}, the turtle
         is located outside of the \code{world}'s extent with the given coordinates.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 5, coords = randomXYcor(w1, n = 5))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1")

t1 <- setXY(turtles = t1, xcor = 1:5, ycor = 1:5)
Plot(t1, addTo ="w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#setxy}
}

