% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{setXY}
\alias{setXY}
\alias{setXY,NLworlds,SpatialPointsDataFrame,numeric,numeric-method}
\title{Set xcor and ycor}
\usage{
setXY(world, turtles, xcor, ycor, torus = FALSE)

\S4method{setXY}{NLworlds,SpatialPointsDataFrame,numeric,numeric}(world,
  turtles, xcor, ycor, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles evolve onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving turtles.}

\item{xcor}{Numeric. X coordinates for the turles. Must be of length 1 or
of the same length as the \code{turtles}.}

\item{ycor}{Numeric. X coordinates for the turles. Must be of length 1 or
of the same length as the \code{turtles}}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A SpatialPointsDataFrame representing the moving turtles with updated
        coordinates and updated data for their previous coordinates (i.e.,
        \code{turtles@data$prevX} and \code{turtles@data$prevY}).
}
\description{
Set the turtles \code{xcor} and \code{ycor} coordinates.
}
\details{
If \code{torus = TRUE} and the given coordinates \code{xcor} or \code{ycor}
         are located outside of the world's extent, then the coordinates assigned
         are the ones from a wrapped word. If \code{torus = FALSE}, the turtle
         is then located outside of the world's extent with the given coordinates.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(100)
t1 <- createTurtles(n = 5, coords = cbind(xcor = randomXcor(world = w1, n = 5),
                                          ycor = randomYcor(world = w1, n = 5)))
library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo ="w1")

t1 <- setXY(world = w1, turtles = t1, xcor = 1:5, ycor = 1:5)
Plot(t1, addTo ="w1")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

