% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\name{setXY}
\alias{setXY}
\alias{setXY,agentMatrix,numeric,numeric,missing,ANY-method}
\alias{setXY,agentMatrix,numeric,numeric,worldNLR,logical-method}
\title{Set \code{turtles}' locations}
\usage{
setXY(turtles, xcor, ycor, world, torus = FALSE)

\S4method{setXY}{agentMatrix,numeric,numeric,missing,ANY}(turtles, xcor, ycor, torus)

\S4method{setXY}{agentMatrix,numeric,numeric,worldNLR,logical}(turtles, xcor, ycor, world, torus = FALSE)
}
\arguments{
\item{turtles}{\code{AgentMatrix} object representing the moving \code{agents}.}

\item{xcor}{Numeric. Vector of \code{x} coordinates. Must be of length 1 or
of length \code{turtles}.}

\item{ycor}{Numeric. Vector of \code{y} coordinates. Must be of length 1 or
of length \code{turtles}.}

\item{world}{\code{WorldMatrix} or \code{worldArray} object.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
\code{AgentMatrix} representing the \code{turtles} with updated coordinates
and updated data for their previous coordinates \code{prevX} and \code{prevY}.
}
\description{
Set the \code{turtles} \code{xcor} and \code{ycor} coordinates.
}
\details{
\code{world} must be provided only if \code{torus = TRUE}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     If the given coordinates `[xcor, ycor]`
     are located outside of the `world`'s extent and `torus = TRUE`,
     then the coordinates assigned to the `turtle`
     are the ones from a wrapped `word`; if `torus = FALSE`, the `turtle`
     is located outside of the `world`'s extent with the given coordinates.
}\if{html}{\out{</div>}}
}
\examples{
w1 <- createWorld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9,
                          data = runif(100))
t1 <- createTurtles(n = 5, coords = randomXYcor(w1, n = 5))
plot(w1)
points(t1, col = of(agents = t1, var = "color"), pch = 16)

t1 <- setXY(turtles = t1, xcor = 1:5, ycor = 1:5)
points(t1, col = of(agents = t1, var = "color"), pch = 16)


}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#setxy}
}
\author{
Sarah Bauduin
}
