% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{sortOn}
\alias{sortOn}
\alias{sortOn,SpatialPointsDataFrame,missing,character-method}
\alias{sortOn,agentMatrix,missing,character-method}
\alias{sortOn,matrix,NLworld,missing-method}
\alias{sortOn,matrix,NLworldArray,character-method}
\alias{sortOn,matrix,NLworldMatrix,missing-method}
\alias{sortOn,matrix,NLworldStack,character-method}
\title{Sort agents}
\usage{
sortOn(agents, world, var)

\S4method{sortOn}{matrix,NLworld,missing}(agents, world)

\S4method{sortOn}{matrix,NLworldStack,character}(agents, world, var)

\S4method{sortOn}{SpatialPointsDataFrame,missing,character}(agents, var)

\S4method{sortOn}{matrix,NLworldMatrix,missing}(agents, world)

\S4method{sortOn}{matrix,NLworldArray,character}(agents, world, var)

\S4method{sortOn}{agentMatrix,missing,character}(agents, var)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{world}{NLworlds object.}

\item{var}{Character. The name of the selected \code{agents} variable.
If \code{agents} are patches and the \code{world} is a
\code{NLworld} object, \code{var} must not be provided. If
\code{agents} are patches and the \code{world} is a \code{NLworldStack}
object, \code{var} is the name of the layer to use to define the patches
values. If \code{agents} are turtles, \code{var} is one of
the turtles' variable and can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
Matrix (ncol = 2) with the first column "pxcor" and the second column
        "pycor" representing the coordinates of the patches, sorted according to
        their values, if \code{agents}
        are patches, or

        SpatialPointsDataFrame representing the turtles, sorted according
        to their \code{var} values, if \code{agents} are
        turtles.
}
\description{
Report the \code{agents} sorted according to their value.
}
\details{
\code{world} must not be provided if \code{agents} are turtles.

         The sorting of the \code{agents} is done in a increasing order.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- set(world = w1, agents = patches(w1),
          val = sample(1:5, size = count(patches(w1)), replace = TRUE))
plot(w1)
p1 <- sortOn(agents = patches(w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
sortdHeadingst1 <- sortOn(agents = t1, var = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#sort-on}
}

