% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{spdf2turtles}
\alias{spdf2turtles}
\alias{spdf2turtles,SpatialPointsDataFrame-method}
\title{Convert a SPDF object into a agentMatrix object}
\usage{
spdf2turtles(spdf)

\S4method{spdf2turtles}{SpatialPointsDataFrame}(spdf)
}
\arguments{
\item{spdf}{}
}
\value{
agentMatrix
}
\description{
Convert a SPDF object into a agentMatrix object
}
\author{
Sarah Bauduin
}

