% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{subHeadings}
\alias{subHeadings}
\alias{subHeadings,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{subHeadings,SpatialPointsDataFrame,numeric-method}
\alias{subHeadings,numeric,SpatialPointsDataFrame-method}
\alias{subHeadings,numeric,numeric-method}
\title{substract headings}
\usage{
subHeadings(heading1, heading2, range360 = FALSE)

\S4method{subHeadings}{numeric,numeric}(heading1, heading2, range360 = FALSE)

\S4method{subHeadings}{SpatialPointsDataFrame,numeric}(heading1, heading2,
  range360 = FALSE)

\S4method{subHeadings}{numeric,SpatialPointsDataFrame}(heading1, heading2,
  range360 = FALSE)

\S4method{subHeadings}{SpatialPointsDataFrame,SpatialPointsDataFrame}(heading1,
  heading2, range360 = FALSE)
}
\arguments{
\item{heading1}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing turtle(s) with its/their
heading(s).
A numeric vector.}

\item{heading2}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing turtle(s) with its/their
heading(s).
A numeric vector.
\code{heading2} must be of length 1 or of length \code{heading1}.}

\item{range360}{Logical. Determine if the returned value(s) are between 0째 and
360째 (\code{range360 = TRUE}) or between -180째 and 180째
(\code{range360 = FALSE}). Default is \code{range360 = FALSE}.}
}
\value{
A numeric vector of values representing the smallest angle in degrees
        by which \code{heading1} could be rotated to produce \code{heading2}
        (i.e., the target heading). Note: this is the opposite as in NetLogo where
        heading1 is the target.
}
\description{
Compute the difference between \code{heading1} and \code{heading2}.
}
\details{
Positive values mean clockwise rotation, negative value mean counterclockwise.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createOTurtles(n = 10, world = w1)
subHeadings(heading1 = t1, heading2 = 0)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

