% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldNLR-classes-methods.R
\docType{methods}
\name{[[}
\alias{[[}
\alias{[[,worldArray,ANY,ANY-method}
\alias{[[<-}
\alias{[[<-,worldArray,ANY,ANY,ANY-method}
\alias{$}
\alias{$,worldArray-method}
\title{Subsetting for worldArray class}
\usage{
\S4method{[[}{worldArray,ANY,ANY}(x, i)

\S4method{[[}{worldArray,ANY,ANY,ANY}(x, i) <- value

\S4method{$}{worldArray}(x, name)
}
\arguments{
\item{x}{A \code{worldArray} object.}

\item{i}{Index number or layer name specifying a subset of layer(s)
from the worldArray.}

\item{value}{A replacement worldMatrix layer for one of the current layers in the
worldArray.}

\item{name}{Layer name, normally without back ticks, unless has symbols.}
}
\description{
These function similarly to \code{[[} for \code{RasterStack} objects
}
