% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{towards}
\alias{towards}
\alias{towards,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,SpatialPointsDataFrame,matrix-method}
\alias{towards,NLworlds,matrix,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,matrix,matrix-method}
\title{Directions towards}
\usage{
towards(world, agents, agents2, torus = FALSE)

\S4method{towards}{NLworlds,matrix,matrix}(world, agents, agents2,
  torus = FALSE)

\S4method{towards}{NLworlds,SpatialPointsDataFrame,matrix}(world, agents,
  agents2, torus = FALSE)

\S4method{towards}{NLworlds,matrix,SpatialPointsDataFrame}(world, agents,
  agents2, torus = FALSE)


  \S4method{towards}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  agents, agents2, torus = FALSE)
}
\arguments{
\item{world}{NLworlds object.}

\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{agents2}{Matrix (ncol = 2) with the first column "pxcor" and the second
               column "pycor" representing the patches coordinates, or

               SpatialPointsDataFrame created by \code{createTurtles()} or
               by \code{createOTurtles()} representing the moving agents, or

               Matrix (ncol = 2) with the first column "x and the second column
               "y" representing locations coordinates.}

\item{torus}{Logical to determine if the \code{world} is wrapped. Default is
\code{torus = FALSE}.}
}
\value{
Numeric. Vector of angles in degrees of length equal to the largest
        number of agents/locations between \code{agents} and \code{agents2}.
}
\description{
Report the directions of \code{agents} towards \code{agents2}.
}
\details{
\code{agents} and \code{agents2} must have the same number of agents/locations
         or if different, one of them must have only one agent/location.

         If \code{torus = TRUE} and the distance from one \code{agents} to
         its corresponding \code{agents2} is smaller around the
         sides of the \code{world} than across it, then the direction to \code{agents2}
         going around the sides of the \code{world} is returned.

         The direction from a patch to its location returns 0; the direction from
         a turtle to its location returns the turtle's heading.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
towards(world = w1, agents = patches(w1), agents2 = cbind(x = 0, y = 0))
t1 <- createTurtles(n = 10, world = w1)
towards(world = w1, agents = t1, agents2 = cbind(x = 0, y = 0))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#towards}

         \url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#towardsxy}
}

