% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{towards}
\alias{towards}
\alias{towards,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,SpatialPointsDataFrame,matrix-method}
\alias{towards,NLworlds,matrix,SpatialPointsDataFrame-method}
\alias{towards,NLworlds,matrix,matrix-method}
\title{Towards}
\usage{
towards(world, from, to, torus = FALSE)

\S4method{towards}{NLworlds,matrix,matrix}(world, from, to, torus = FALSE)

\S4method{towards}{NLworlds,SpatialPointsDataFrame,matrix}(world, from, to,
  torus = FALSE)

\S4method{towards}{NLworlds,matrix,SpatialPointsDataFrame}(world, from, to,
  torus = FALSE)


  \S4method{towards}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame}(world,
  from, to, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object.}

\item{from}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) from which the heading will be computed.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) from which the
heading will be computed.}

\item{to}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patch(es) to which the heading will be computed.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtle(s) to which the
heading will be computed.
A matrix (ncol = 2) with the first column \code{x} and the second
column \code{y} representing the coordinates of the location(s) to
which the heading will be computed.
\code{to} must be of length 1 or of the same length as \code{from}.}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A vector of angles in degrees of the length of \code{from}.
}
\description{
Reports the heading from agent(s) towards other agent(s) or location(s) [x,y].
}
\details{
If \code{torus = TRUE} and the distance from one agent \code{from} to
         its corresponding agent or location \code{to} is smaller around the
         sides of the world than across it, then the heading to the agent or location
         going around the sides of the world is reported.
         The heading from an agent to itself or its own location will return 0.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
towards(world = w1, from = patches(world = w1), to = cbind(x = 0, y = 0))
t1 <- createTurtles(n = 10, world = w1)
towards(world = w1, from = t1, to = cbind(x = 0, y = 0))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

