% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtle}
\alias{turtle}
\alias{turtle,SpatialPointsDataFrame,numeric,character-method}
\alias{turtle,SpatialPointsDataFrame,numeric,missing-method}
\title{Select turtles}
\usage{
turtle(turtles, who, breed)

\S4method{turtle}{SpatialPointsDataFrame,numeric,missing}(turtles, who)

\S4method{turtle}{SpatialPointsDataFrame,numeric,character}(turtles, who, breed)
}
\arguments{
\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or
by \code{createOTurtles()} representing the moving agents.}

\item{who}{Integer. Vector of the "who" numbers of the selected turtles.}

\item{breed}{Characters. Vector of "breed" names to select the \code{turtles}.
Must be of length 1 or of length \code{turtles}.
If missing, there is
no distinction based upon "breed".}
}
\value{
SpatialPointsDataFrame of the selected turtles.
}
\description{
Report the individuals among \code{turtles} based on their \code{who} numbers
and \code{breed}.
}
\details{
If no turtle matches the given \code{who} numbers, with potentially the given
         \code{breed}, inside \code{turtles}, then an empty SpatialPointsDataFrame is returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
t2 <- turtle(t1, who = 2)


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#turtle}
}

