% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\docType{methods}
\name{turtles2spdf}
\alias{turtles2spdf}
\alias{turtles2spdf,agentMatrix-method}
\title{Convert an agentMatrix object into a SPDF}
\usage{
turtles2spdf(turtles)

\S4method{turtles2spdf}{agentMatrix}(turtles)
}
\arguments{
\item{turtles}{description needed}
}
\value{
SpatialPointsDataFrame
}
\description{
Convert an agentMatrix object into a SPDF
}
\author{
Sarah Bauduin
}

