% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtlesAt}
\alias{turtlesAt}
\alias{turtlesAt,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,numeric,numeric,character-method}
\alias{turtlesAt,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,numeric,numeric,missing-method}
\alias{turtlesAt,NLworlds,SpatialPointsDataFrame,matrix,numeric,numeric,character-method}
\alias{turtlesAt,NLworlds,SpatialPointsDataFrame,matrix,numeric,numeric,missing-method}
\title{Turtles at}
\usage{
turtlesAt(world, turtles, agents, dx, dy, breed, torus = FALSE)


  \S4method{turtlesAt}{NLworlds,SpatialPointsDataFrame,matrix,numeric,numeric,missing}(world,
  turtles, agents, dx, dy, torus)


  \S4method{turtlesAt}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,numeric,numeric,missing}(world,
  turtles, agents, dx, dy, torus)


  \S4method{turtlesAt}{NLworlds,SpatialPointsDataFrame,matrix,numeric,numeric,character}(world,
  turtles, agents, dx, dy, breed, torus = FALSE)


  \S4method{turtlesAt}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,numeric,numeric,character}(world,
  turtles, agents, dx, dy, breed, torus = FALSE)
}
\arguments{
\item{world}{A \code{NLworlds} object where the turtles evolve onto.}

\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles among which the
one(s) at \code{(dx, dy)} distance of the \code{agents} is/are
going to be reported.}

\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates of the
patches from wich \code{(dx, dy)} are computed.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles from wich
\code{(dx, dy)} are computed.}

\item{dx}{Numeric. Distance to east from the agent. If \code{dx} is negative, the
distance to the west is computed. \code{dx} must be a single value or
of the length of \code{agents}.}

\item{dy}{Numeric. Distance to the north from the agent. If \code{dy} is negative,
the distance to the south is computed. \code{dy} must be a single value or
of the length of \code{agents}.}

\item{breed}{String of characters. To specify the breed of turtles to report.
If missing, there is no distinction based upon breed to report turtle(s).}

\item{torus}{Logical to determine if the \code{NLworlds} object is wrapped.
Default is \code{torus = FALSE}.}
}
\value{
A SpatialPointsDataFrame representing the turtle(s) among \code{turtles}
        which are located on the patch(es) at \code{(dx, dy)} distance of the
        \code{agents}.
}
\description{
Reports the turtle(s) on the patch(es) at \code{(dx, dy)} distance of the agent(s).
}
\details{
If \code{torus = FALSE} and the pacth at distance \code{(dx, dy)}
         of the agent is outside of the world's extent, no turtle is returned.
         If \code{torus = TRUE}, the turtle located on the patch whose coordinates
         are defined from the wrapped world is returned.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = cbind(xcor = 0:9, ycor = 0:9), breed = c(rep("sheep", 5), rep("wolf", 5)))
t2 <- turtlesAt(world = w1, turtles = t1, agents = turtle(turtles = t1, who = 0), dx = 1, dy = 1)
t3 <- turtlesAt(world = w1, turtles = t1, agents = patch(world = w1, xcor = c(3,4,5), ycor = c(3,4,5)),
                dx = 1, dy = 1, breed = "sheep")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

