% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtlesOn}
\alias{turtlesOn}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,character-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,missing-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,matrix,character-method}
\alias{turtlesOn,NLworlds,SpatialPointsDataFrame,matrix,missing-method}
\title{Turtles on}
\usage{
turtlesOn(world, turtles, agents, breed)

\S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,matrix,missing}(world,
  turtles, agents)

\S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,matrix,character}(world,
  turtles, agents, breed)


  \S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,missing}(world,
  turtles, agents)


  \S4method{turtlesOn}{NLworlds,SpatialPointsDataFrame,SpatialPointsDataFrame,character}(world,
  turtles, agents, breed)
}
\arguments{
\item{world}{\code{NLworlds} object, representing the world which the turtles
move onto.}

\item{turtles}{SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles among which the
ones on the same patches as \code{agents} are going to
be returned.}

\item{agents}{Matrix (ncol = 2) with the first column \code{pxcor} and the
               second column \code{pycor} representing the coordinates of the
               patches where to look for turtles.

               SpatialPointsDataFrame created by \code{createTurtles()} or by
               \code{createOTurtles()} representing turtles whose patches locations
               are the patches where to look for turtles.}

\item{breed}{Characters. To specify the breeds of turtles to report.
If missing, there is no distinction based upon breed to report turtles.}
}
\value{
SpatialPointsDataFrame representing the turtles among \code{turtles}
        which are located on the \code{agent} patches and of the given \code{breed}
        if specified.
}
\description{
Report the turtles at some patches locations or on the same patches as some turtles.
}
\details{
Patches or turtles given as \code{agents} must be located inside the
         world's extent.
}
\examples{
w1 <- createNLworld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
w1[] <- runif(length(w1))
t1 <- createTurtles(n = 500, coords = randomXYcor(w1, n = 500))

library(SpaDES)
clearPlot()
Plot(w1)
Plot(t1, addTo = "w1")

t2 <- turtlesOn(world = w1, turtles = t1, agents = patch(w1, 2, 2))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#turtles-on}
}

