% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle-functions.R
\docType{methods}
\name{turtlesOwn}
\alias{turtlesOwn}
\alias{turtlesOwn,SpatialPointsDataFrame,character,ANY-method}
\alias{turtlesOwn,SpatialPointsDataFrame,character,missing-method}
\title{Turtles own variable}
\usage{
turtlesOwn(turtles, tVarName, tVar)

\S4method{turtlesOwn}{SpatialPointsDataFrame,character,missing}(turtles,
  tVarName)

\S4method{turtlesOwn}{SpatialPointsDataFrame,character,ANY}(turtles, tVarName,
  tVar)
}
\arguments{
\item{turtles}{A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing turtles.}

\item{tVarName}{Characters. Name of the new variable to create.}

\item{tVar}{Any type of vector representing the value of \code{tVarName}.
\code{length(tVar)} must be equal to 1 or to \code{length(turtles)}.
If missing, \code{NA} is given.}
}
\value{
A SpatialPointsDataFrame representing the \code{turtles} with the new
        variable \code{tVarName} added.
}
\description{
Create a new variable for the turtle(s) provided.
}
\examples{
t1 <- createTurtles(n = 5, coords = cbind(xcor = 0, ycor = 0))
t1 <- turtlesOwn(turtles = t1, tVarName = "sex", tVar = c("F", "F", "F", "M", "M"))


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

