% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{withMax}
\alias{withMax}
\alias{withMax,SpatialPointsDataFrame,missing,character-method}
\alias{withMax,matrix,NLworld,missing-method}
\alias{withMax,matrix,NLworldStack,character-method}
\title{With maximum}
\usage{
withMax(agents, world, varName)

\S4method{withMax}{matrix,NLworld,missing}(agents, world)

\S4method{withMax}{matrix,NLworldStack,character}(agents, world, varName)

\S4method{withMax}{SpatialPointsDataFrame,missing,character}(agents, varName)
}
\arguments{
\item{agents}{A matrix (ncol = 2) with the first column \code{pxcor} and the
second column \code{pycor} representing the coordinates for the
patches to evaluate.
A SpatialPointsDataFrame created by \code{createTurtles()} or by
\code{createOTurtles()} representing the turtles to evaluate.}

\item{world}{A \code{NLworlds} object. Only needed if \code{agents} are patches.
Must not be provided if \code{agents} are turtles.}

\item{varName}{Characters. The name of the variable to evaluate for the maximum value.
If \code{agents} are patches and the \code{world} is a \code{NLworld}
object, \code{varName} must not be provided. If \code{agents} are
patches and the \code{world} is a \code{NLworldStack} object,
\code{varName} refers to the layer used for evaluating patches
values. If \code{agents} are turtles, \code{varName} is one of
the turtles' variable. \code{varName} can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
A matrix (ncol = 2) with the first column \code{pxcor} and the second column
        \code{pycor} representing the coordinates for the patches with the maximum
        value among the \code{agents}.
        A SpatialPointsDataFrame representing the turtles which have their
        \code{varName} equals to the maximum values among the \code{agents}.
}
\description{
Reports the coordinates \code{pxcor} and \code{pycor} of the patches which
have their variable equals to the maximum value.
Reports the turtles which have their variable equals to the maximum values.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1[] <- sample(1:5, size = 25, replace = TRUE)
plot(w1)
pMax <- withMax(agents = patches(world = w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(world = w1, n = 10), heading = sample(1:3, size = 10, replace= TRUE))
t2 <- withMax(agents = t1, varName = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}

