% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentset-functions.R
\docType{methods}
\name{withMin}
\alias{withMin}
\alias{withMin,SpatialPointsDataFrame,missing,character-method}
\alias{withMin,agentMatrix,missing,character-method}
\alias{withMin,matrix,NLworld,missing-method}
\alias{withMin,matrix,NLworldArray,character-method}
\alias{withMin,matrix,NLworldMatrix,missing-method}
\alias{withMin,matrix,NLworldStack,character-method}
\title{Agents with minimum}
\usage{
withMin(agents, world, var)

\S4method{withMin}{matrix,NLworld,missing}(agents, world)

\S4method{withMin}{matrix,NLworldStack,character}(agents, world, var)

\S4method{withMin}{SpatialPointsDataFrame,missing,character}(agents, var)

\S4method{withMin}{matrix,NLworldMatrix,missing}(agents, world)

\S4method{withMin}{matrix,NLworldArray,character}(agents, world, var)

\S4method{withMin}{agentMatrix,missing,character}(agents, var)
}
\arguments{
\item{agents}{Matrix (ncol = 2) with the first column "pxcor" and the second
              column "pycor" representing the patches coordinates, or

              SpatialPointsDataFrame created by \code{createTurtles()} or
              by \code{createOTurtles()} representing the moving agents.}

\item{world}{NLworlds object.}

\item{var}{Character. The name of the selected \code{agents} variable.
If \code{agents} are patches and the \code{world} is a
\code{NLworld} object, \code{var} must not be provided. If
\code{agents} are patches and the \code{world} is a \code{NLworldStack}
object, \code{var} is the name of the layer to use to define the patches
values. If \code{agents} are turtles, \code{var} is one of
the turtles' variable and can be equal to \code{"xcor"},
\code{"ycor"}, any of the variables created when turtles were created,
as well as any variable created using \code{turtlesOwn()}.}
}
\value{
Matrix (ncol = 2) with the first column "pxcor" and the second column
        "pycor" representing the coordinates of the patches among the \code{agents}
        which have their variable
        equal to the minimum value among the \code{agents}, or

        SpatialPointsDataFrame representing the turtles among the \code{agents}
        which have their variable
        \code{var} equal to the minimum value among the \code{agents}.
}
\description{
Report the patches or turtles among \code{agents} which have their variable
equals to the minimum value.
}
\details{
\code{world} must not be provided if \code{agents} are turtles.
}
\examples{
# Patches
w1 <- createNLworld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
w1 <- NLset(world = w1, agents = patches(w1),
          val = sample(1:5, size = NLcount(patches(w1)), replace = TRUE))
plot(w1)
p1 <- withMin(agents = patches(w1), world = w1)

# Turtles
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10),
                    heading = sample(1:3, size = 10, replace= TRUE))
t2 <- withMin(agents = t1, var = "heading")


}
\author{
Sarah Bauduin
}
\references{
Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/.
            Center for Connected Learning and Computer-Based Modeling,
            Northwestern University. Evanston, IL.
}
\seealso{
\url{https://ccl.northwestern.edu/netlogo/docs/dictionary.html#with-min}
}

