% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldNLR-classes-methods.R
\docType{class}
\name{worldArray-class}
\alias{worldArray-class}
\alias{worldArray}
\title{The worldArray class}
\description{
This is an s4 class extension of \code{array}. It is a collection of several
\code{worldMatrix} objects with the same extent (i.e., same values for all their
slots) stacked together. It is used to keep more than one value per patch.
}
\seealso{
\code{\link{worldMatrix}}
}
\author{
Sarah Bauduin, Eliot McIntire, and Alex Chubaty
}
