% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world-functions.R
\name{worldHeight}
\alias{worldHeight}
\alias{worldHeight,worldNLR-method}
\title{\code{World} height}
\usage{
worldHeight(world)

\S4method{worldHeight}{worldNLR}(world)
}
\arguments{
\item{world}{\code{WorldMatrix} or \code{worldArray} object.}
}
\value{
Integer.
}
\description{
Report the height of the \code{world} in \code{patch} number.
}
\examples{
w1 <- createWorld()
worldHeight(w1)

}
\references{
Wilensky, U. 1999. NetLogo. https://www.netlogo.org.
Center for Connected Learning and Computer-Based Modeling,
Northwestern University. Evanston, IL.
}
\seealso{
\url{https://docs.netlogo.org/dictionary.html#world-dim}
}
\author{
Sarah Bauduin
}
