% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldNLR-classes-methods.R
\docType{class}
\name{worldNLR-class}
\alias{worldNLR-class}
\alias{worldNLR}
\title{The worldNLR class}
\description{
The \code{worldNLR} class is the union of the \code{worldMatrix} and \code{worldArray}
classes. Mostly used for building function purposes.
}
\author{
Sarah Bauduin, and Eliot McIntire
}
